% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbAgStatRec.R, R/AbAgStatRec.camp.R
\name{AbAgStatRec.camp}
\alias{AbAgStatRec.camp}
\title{Calcula las abundancias estratificadas por edad y hora en cada rectángulo ICES de la zona considerada}
\usage{
AbAgStatRec.camp(
  gr,
  esp,
  camp,
  dns = "Cant",
  plus = 8,
  mediahora = 2,
  cor.time = TRUE,
  AltAlk = NA,
  incl2 = FALSE,
  DatGraf = FALSE,
  plotrix = T,
  es = T
)

AbAgStatRec.camp(
  gr,
  esp,
  camp,
  dns = "Cant",
  plus = 8,
  mediahora = 2,
  cor.time = TRUE,
  AltAlk = NA,
  incl2 = FALSE,
  DatGraf = FALSE,
  plotrix = T,
  es = T
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camp}{Campaña de la que se extraen los datos un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{mediahora}{Pondera a una hora si se solicita por defecto en 2 suponiendo lances de media hora}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{AltAlk}{ALK alternativa tomada de un fichero de edad del Camp edadXYY.dbf sin ruta ni extensión}

\item{incl2}{Si T incluye los datos de los lances extra, tener en cuenta que no son lances oficiales para calcular los índices de abundancia}

\item{DatGraf}{Si F saca tabla con abundancias en cada rectángulo ICES muestreado en la campaña, si T saca los datos para sacar gráfica con MapIStatRec}

\item{es}{si T saca para etiquetas en español (Ed para edad), si F en inglés (Ag para edad)}
}
\description{
Función de resultados: abundancias estratificadas por edad en cada rectángulo ICES dentro del área de la cqampaña a partir de los datos del camp.

Función de resultados: abundancias estratificadas por edad en cada rectángulo ICES dentro del área de la cqampaña a partir de los datos del camp.
}
\examples{
AbAgStatRec.camp(1,74,"N21","Cant",plus=3,DatGraf = F)
AbAgStatRec.camp(1,43,"P10","Porc",DatGraf = T)
AbAgStatRec.camp(1,74,"N21","Cant",plus=3,DatGraf = F)
AbAgStatRec.camp(1,43,"P10","Porc",DatGraf = T)
}
\seealso{
Other edades: 
\code{\link{ALKs.dns.camp}()},
\code{\link{GetAlk.camp}()},
\code{\link{bubbage.camp}()},
\code{\link{datagegr.camp}()},
\code{\link{edadsect.camp}()},
\code{\link{edadstr.camp}()},
\code{\link{grafedtal.camps}()},
\code{\link{grafedtal.camp}()},
\code{\link{logabage.camp}()},
\code{\link{logabage2.camp}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maphistage}()}

Other edades: 
\code{\link{ALKs.dns.camp}()},
\code{\link{GetAlk.camp}()},
\code{\link{bubbage.camp}()},
\code{\link{datagegr.camp}()},
\code{\link{edadsect.camp}()},
\code{\link{edadstr.camp}()},
\code{\link{grafedtal.camps}()},
\code{\link{grafedtal.camp}()},
\code{\link{logabage.camp}()},
\code{\link{logabage2.camp}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maphistage}()}
}
\concept{edades}
