% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbrvEsp.R
\name{AbrvEsp}
\alias{AbrvEsp}
\title{Crea abreviaturas de especies a partir del nombre científico, género+especie}
\usage{
AbrvEsp(x, nGen = 3, nEsp = 4, sep = "_", unique = TRUE)
}
\arguments{
\item{x}{Vector con la serie de nombres científicos a abreviar}

\item{nGen}{Número de caracteres del género a utilizar}

\item{nEsp}{número de catacteres del nombre de especie a utilizar}

\item{sep}{Caracter para separar Género de especie}
}
\value{
Devuelve un data.frame con dos columnas, nombres originales $names y los codigos $codes
}
\description{
Transforma series de nombres científicos de especies en abreviaturas de 9 caracteres, se puede elegir el número de
caracteres y el caracter de separación. Ignora borrando el sp. los nombres genéricos de especies con sp., no evita
los spp. cuidado.
Avisa si hay códigos duplicados en el resultado de los códigos solicitados, para evitar crear códigos repetidos para distinas especies
}
\examples{
AbrvEsp(c("Merluccius merluccius","Merlangius merlangus"))
AbrvEsp(buscaesp(1,50),1,8,".")
}
\seealso{
Other datos_especies: 
\code{\link{BuscaAphia}()},
\code{\link{buscacod}()},
\code{\link{buscaesp}()},
\code{\link{talpes.camp}()}
}
\concept{datos_especies}
