% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agradec.R
\name{agradec}
\alias{agradec}
\title{Transforma grados decimales a grados y minutos decimales}
\usage{
agradec(x)
}
\arguments{
\item{x}{vector con grados y minutos en formato decimal}
}
\description{
Transforma grados decimales a grados y minutos decimales en formato XXºxxxx' donde el valor tras el decimal son los minutos decimales pero sin decimal
}
\examples{
agradec(43.968)
}
\concept{Conversión unidades}
