% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALKS.dns.camp.R
\name{ALKs.dns.camp}
\alias{ALKs.dns.camp}
\title{Datos de claves talla edad en campañas en dns}
\usage{
ALKs.dns.camp(dns, camp = "")
}
\arguments{
\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa", Mediterraneo "Medit"}

\item{camp}{Permite elegir las campañas para filtrar sólo las que interesen, N para demersales, P para Porcupine, 1 o 2 para Arsa, M para Medits... Si es "" como es por defecto saca todos los datos de los ficheros edadXXX}
}
\value{
Devuelve un data.frame con las especies con datos de edad en las campañas presentes en el "directorio" dns
}
\description{
Función de acceso a datos:
A partir de ficheros de fauna.dbf presentes en directorio comprueba presencia de especie
}
\examples{
ALKs.dns.camp("Porc","P")
ALKs.dns.camp("Arsa","1")
}
\seealso{
Other edades: \code{\link{GetAlk.camp}},
  \code{\link{bubbage.camp}}, \code{\link{datagegr.camp}},
  \code{\link{edadsect.camp}}, \code{\link{edadstr.camp}},
  \code{\link{grafedtal.camps}},
  \code{\link{grafedtal.camp}},
  \code{\link{logabage.camp}},
  \code{\link{logabage2.camp}}, \code{\link{maphistage}}
}
\concept{edades}
