% Generated by roxygen2 (4.0.2): do not edit by hand
\name{armap.camp}
\alias{armap.camp}
\title{Mapa con las estaciones realizadas}
\usage{
armap.camp(camp, dns = "Pnew", lwdl = 1, col = 2, argr = 2, cuadr = F,
  cuadrMSFD = F, arrow = F, leg = T, es = F, cols = T, noval = T,
  CTDs = F, strat = F, Nlans = F, NCTDs = F, places = F, bw = F,
  lans = T, xlims = c(-10.25, -1.4), ylims = c(41.82, 44.48))
}
\arguments{
\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{lwdl}{Anchura de las líneas}

\item{col}{Color de los puntos que representan estaciones}

\item{argr}{Grosor de las flechas de recorrido}

\item{cuadr}{Si T dibuja cuadrícula}

\item{cuadrMSFD}{Si T dibuja caudrícula de 10 millas naúticas utilizada para la evaluación de la estrategia marina (MSFD)}

\item{arrow}{Si T añade flechas de reccorrido}

\item{leg}{Si T añade leyenda}

\item{es}{Si T nombres en castellano, si F nombres en inglés}

\item{cols}{Si T colorea los estratos y sectores, si F mapa en blanco y negro}

\item{noval}{Si T añade las estaciones nulas}

\item{CTDs}{Si T añade las estaciones de CTD realizadas}

\item{strat}{Si T delimita los sectores y los identifica}

\item{Nlans}{Si T identifica las estaciones numéricamente}

\item{NCTDs}{Si T identifica las estaciones de CTD numéricamente}

\item{places}{Si T añade los nombres de las principales ciudades}

\item{bw}{Gráfico en blanco y negro si T o en color si F}

\item{lans}{Si T marca las estaciones de muestreo con un punto (si Nlans=F)}

\item{xlims}{Delimita la longitud del mapa mediante un vector (ejem.c(-10.25,-1.4))}

\item{ylims}{Delimita la latitud del mapa mediante un vector (ejem.c(41.82,44.48))}
}
\description{
Crea un mapa de un único año con las estaciones realizadas
}

