% Generated by roxygen2 (4.0.2): do not edit by hand
\name{armap.tot}
\alias{armap.tot}
\title{Mapa ampliado de Porcupine con las estaciones realizadas}
\usage{
armap.tot(camp, dns = "Pnew", lwdl = 1, col = 2, argr = 2, arrow = F,
  leg = F, es = F, bw = T, noval = F, CTDs = F, strat = F)
}
\arguments{
\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant" (ARSA para sacar datos al IBTS, no gráficos)}

\item{lwdl}{Anchura de las líneas}

\item{col}{Color de los puntos que representan estaciones}

\item{argr}{Grosor de las flechas de recorrido}

\item{arrow}{Si T añade flechas de reccorrido}

\item{leg}{Si T añade leyenda}

\item{es}{Si T nombres en castellano, si F nombres en inglés}

\item{bw}{Si T mapa en blanco y negro, si F colorea los estratos y sectores}

\item{noval}{Si T añade las estaciones nulas}

\item{CTDs}{Si T añade las estaciones de CTD realizadas}

\item{strat}{strat Si T represent los estratos}
}
\description{
Crea un mapa de Porcupine con las estaciones realizadas(al igua que "armap.camp") pero con la referencia de Irlanda para su mejor ubicación.
}

