% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arMapLansGPS.R
\name{arMapLansGPS}
\alias{arMapLansGPS}
\title{Combinación en dos mapas de armap.camp con etiquetas de los lance y mapa con los lances en segmentos}
\usage{
arMapLansGPS(
  camp,
  dns = "Porc",
  incl0 = FALSE,
  xlims = NA,
  ylims = NA,
  col = 2,
  lwd = 2,
  cexlab = 0.6,
  places = TRUE,
  Nlans = FALSE,
  es = T,
  bw = FALSE,
  layout = NA
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa"}

\item{incl0}{Si T se incluyen los lances nulos}

\item{xlims}{Define los limites longitudinales del mapa, si se deja en NA toma los límites de long del área definida en la campaña}

\item{ylims}{Define los limites latitudinales del mapa, si se deja en NA toma los límites de lat del área definida en la campaña}

\item{col}{Define el color de los segmentos}

\item{lwd}{Define el ancho de los segmentos}

\item{cexlab}{Tamaño de los números de lance utilizados en el gráfico de armap.lan parte del gráfico}

\item{places}{si T por defecto, incluye las etiquetas de países y ciudad en tierra, no funciona en Porcupine}

\item{Nlans}{se T incluye los números de lance por encima de los segmentos}

\item{bw}{Si T gráfico en blanco y negro por default, si F gráfico en color}

\item{rumbo}{si T incluye un punto al final del lance, en el punto de la virada, sólo en el segundo mapa}
}
\value{
Devuelve un mapa, pero no objetos de datos.
}
\description{
Muestra un mapa con el número de lance en el punto del lance y debajo o al lado mapa de los lances en segmentos
}
\examples{
arMapLansGPS("N20","Cant",xlims=c(-9.8,-7.8),ylims=c(43.5,44.2))
}
\seealso{
{\link{datlan.camp}}, {\link{qcdistlan.camp}}

Other mapas: 
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{NepFUs.camp}()},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{map.check}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other PescaWin: 
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{pasa.arte}()},
\code{\link{pasa.lan}()}
}
\concept{PescaWin}
\concept{mapas}
