% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/As1.R
\docType{data}
\name{As1}
\alias{As1}
\title{Serie histórica de la campaña ARSA de primavera}
\format{A vector with codes for the surveys from 1992 to 2018}
\usage{
As1
}
\description{
Serie histórica Campañas ARSA Primavera
}
\details{
Serie completa con las campañas de otoño realizadas en el Golfo de Cádiz (Arrastre Sur Atlántica)
}
\seealso{
Other Series datos: 
\code{\link{As2}},
\code{\link{Msh}},
\code{\link{Nsh}},
\code{\link{Psh}}

Other ARSA: 
\code{\link{As2}},
\code{\link{MapArsa}()}
}
\concept{ARSA}
\concept{Series datos}
\keyword{datasets}
