% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Ash}
\alias{Ash}
\title{Serie histórica de la campaña Arsa de primavera}
\format{A vector with codes for the surveys from 1997 to 2012}
\usage{
Ash
}
\description{
Serie histórica Campañas ARSA
}
\details{
Serie completa con las campañas de otoño realizadas en el Golfo de Cádiz (Arrastre Sur Atlántica)
}
\seealso{
Nsh {\link{Nsh}} Psh {\link{Psh}}
}
\keyword{datasets}

