% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignLS.R
\name{assignLS}
\alias{assignLS}
\title{assignLS}
\usage{
assignLS(runDat, G = NULL, USEWeight = NULL)
}
\arguments{
\item{DATA}{describe argument}

\item{groupN=NULL}{describe argument}

\item{USEWeight=NULL}{describe argument}
}
\value{
describe return value
}
\description{
Jared Studyvin
 10 Feb 2016
 Assign life stage


 This function will be called within F.get.catch.data
 The purpose is to replace the life stage column with an updated assignment based on a clustering routine
}
\details{
Need some details
}
\examples{
# insert examples

}
\author{
WEST Inc.
}
\seealso{
\code{\link{related routine}}, \code{\link{related routine}}
}

