% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignLSCompare.R
\name{assignLSCompare}
\alias{assignLSCompare}
\title{assignLSCompare}
\usage{
assignLSCompare(Data, muLIST, sigmaLIST, SAVE = TRUE)
}
\arguments{
\item{Data}{describe argument}

\item{SAVE=TRUE}{describe argument}
}
\value{
describe return value
}
\description{
Assigns life stage
}
\details{
other comments found in file
 for debugging
data <- subset(Data,FinalRun=='Fall')
 this run
 confusion matrix
head(data)
 mixture distribution life stage level colors
 biologist life stage symbols
 data to be plotted now
nrow(plotData)
nrow(data)
head(plotData)
 add ellipse to the figure
 add ellipse to the figure
 save.image(file="C:/Users/jmitchell/Desktop/FirstLineBigLooper.RData")
monthLabel
par(mfrow=c(1,2))
 title for figure
 plot forklength and date
with(data,table(col,pch))
 legend info for mixture life stage
 legend info for biologist life stage
legMix
legBio
 prepares the legend so the color and pch match up correctly
 for debugging
ddply(Data,~FinalRun,compare,save=FALSE)
}
\examples{
\dontrun{
#insert examples

}
}
\author{
WEST Inc.
}
\seealso{
\code{\link{related routine}}, \code{\link{related routine}}
}

