% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbage.camp.R
\name{bubbage.camp}
\alias{bubbage.camp}
\title{Crea gráficos de burbujas de la abundancia sqrt(num)}
\usage{
bubbage.camp(gr, esp, camps, dns = "Pnew", plus = 8, recr = 0,
  cor.time = TRUE)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camps}{Serie historica de campañas de la que se extraen los datos. Todas las campañas han de tener ALKs para las especie en cuestión}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{recr}{edad de reclutamiento para especies con reclutamiento en edad 1 o mayor}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}
}
\value{
Saca un lattice de descenso de abundancia logarítmica con la edad con dato de la pendiente para cada año. También saca una matriz de abundancias por años x edad columnas x filas
}
\description{
Tipico bubleplot de análisis de VPA y evaluación de poblaciones con gráfico por edad y año
}
\examples{
bubbage.camp(1,43,Nsh[1:29],"Cant",8,0)
}
\seealso{
Other edades: \code{\link{GetAlk.camp}},
  \code{\link{datagegr.camp}}, \code{\link{edadsect.camp}},
  \code{\link{edadstr.camp}},
  \code{\link{grafedtal.camps}},
  \code{\link{grafedtal.camp}},
  \code{\link{logabage.camp}},
  \code{\link{logabage2.camp}}, \code{\link{maphistage}}
}
