% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuscaAphia.R
\name{BuscaAphia}
\alias{BuscaAphia}
\title{Recupera los AphiaId de las especies sin Aphiaid en el fichero especies.dbf}
\usage{
BuscaAphia(gr = 1, dns = "Camp", export = TRUE)
}
\arguments{
\item{gr}{Grupo del que se desea buscar el AphiaID: 1 peces, 2 crustáceos, 3 moluscos, 4 equinodermos, 5 otros invertebrados}

\item{dns}{Directorio en el que buscar el especies.dbf según se haya definido con los ODBC}

\item{export}{Si T crea un fichero especies.csv con todos los datos corregidos (APHIAs) en el directorio CAMP donde está el especies.dbf este es importable al especies.dbf con un append from deli with, quitando todos los peces grupo="1"}
}
\value{
Devuelve un data.table con datos de cada especie en el formato HL de DATRAS. DATRAS requiere que los datos no tengan cabecera y el trimestre sea el que corresponde a la campaña, además de no tener "". Por ello se debe pasar a fichero con la orden: write.table(CAMPtoHH(Xyy,dns),"nombrearchivo.csv",sep=",",quote=F,col.names=F,row.names=F))
}
\description{
Función de comprobación y recuperación de AphiaIDs, funciona sólo con el fichero especies.dbf del raiz del camp, no utiliza lances ni fauna de una campaña
Sirve para crear un fichero del que se pueden importar los códigos AphiaID al especies.dbf a través de los programas de dBase recaphia.prg o similares.
}
\examples{
# BuscaAphia(1,"P14","Porc")
}
\seealso{
Other datos_especies: 
\code{\link{AbrvEsp}()},
\code{\link{buscacod}()},
\code{\link{buscaesp.datras}()},
\code{\link{buscaesp}()},
\code{\link{camptoyear}()},
\code{\link{hidrotodec}()},
\code{\link{talpes.camp}()}
}
\concept{datos_especies}
