% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buscacod}
\alias{buscacod}
\title{Función de búsqueda del código del grupo y la familia}
\usage{
buscacod(nomb)
}
\arguments{
\item{nomb}{Nombre científico de la especie o algún fragmento del nombre usando "".}
}
\description{
Busca el código del grupo, especie y la familia a partir del nombre o fragmento de la especie. Es independiente de mayúsculas y minúsculas y puede sacar más de una especie si una parte de sus nombres coincide.
}
\seealso{
buscaesp {/link[buscaesp]}
}

