% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscaesp.datras.R
\name{buscaesp.datras}
\alias{buscaesp.datras}
\title{Función de búsqueda del datos de una especie en una campaña de DATRAS}
\usage{
buscaesp.datras(gr, esp, camp, year, quarter, dns = "Camp")
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{este es el código de campaña DATRAS, no los códigos del IEO}

\item{year}{el año de la campaña}

\item{quarter}{el trimestre de la campaña}

\item{id}{Salida como: "l" Nombre científico en latín, "e" nombre en español, o "i" nombre en inglés}
}
\description{
La búsqueda se hace a través de los códigos CAMP, el nombre de la campaña en DATRAS, el año y el trimestre de la campaña.
}
\examples{
buscaesp(1,50)
}
\seealso{
Other datos_especies: 
\code{\link{AbrvEsp}()},
\code{\link{BuscaAphia}()},
\code{\link{buscacod}()},
\code{\link{buscaesp}()},
\code{\link{camptoyear}()},
\code{\link{hidrotodec}()},
\code{\link{talpes.camp}()}
}
\concept{datos_especies}
