% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buscaesp}
\alias{buscaesp}
\title{Función de búsqueda del nombre de una especie}
\usage{
buscaesp(gr, esp, id = "l")
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{id}{Nombre científico de la especie ("l"), nombre común ("e"), o nombre en inglés ("i")}
}
\description{
Busca el nombre de una especie a partir del código del grupo y especie.Puede dar el nombre científico, común o en inglés
}
\seealso{
buscacod {/link[buscacod]}
}

