% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_imports.R
\docType{package}
\name{CampR-package}
\alias{CampR}
\alias{CampR-package}
\title{CampR: Libreria de funciones en R para el programa CAMP}
\description{
Funciones para sacar resultados de los datos de campañas de arrastre demersal guardados en el CAMP version 12 Windows7 - Windows10 (Harbour) Solo para 32 bits y Rstudio < 1.2 no funciona con R de 64 bits
}
\keyword{internal}
