% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CampR-package.r
\docType{package}
\name{CampR}
\alias{CampR}
\alias{CampR-package}
\title{Libreria de funciones de R para el programa Camp}
\description{
Libreria en R para trabajar con datos de campañas cientificas de arrastre demersal almacenados en las bases de datos del programa CAMP
}
\details{
CampR.

Contiene funciones para:
\itemize{
\item {Control de calidad de los datos y deteccion de errores}
\item {Analisis y procesado de los datos}
\item {Salidas graficas y de tablas para informes y publicaciones}
\item {Mapeado de los resultados en las zonas implementadas en la libreria}
}
}
