% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CampsDNS.camp.R
\name{CampsDNS.camp}
\alias{CampsDNS.camp}
\title{Listado de campañas presentes en el directorio e información sobre ellas}
\usage{
CampsDNS.camp(dns)
}
\arguments{
\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}
}
\value{
Un data.frame con columnas con la lista de archivos campXXX,lanceXXX,faunaXXX,ntallXXX e hidroXXX en el directorio, rellena los que falten con "."
}
\description{
Función de acceso a datos:
A partir de ficheros de fauna.dbf presentes en directorio comprueba presencia de especie
}
\examples{
CampsDNS.camp("Porc")
}
\seealso{
Other Series datos: 
\code{\link{As1}},
\code{\link{As2}},
\code{\link{Msh}},
\code{\link{Nsh}},
\code{\link{PrSh}},
\code{\link{PresenciaEsp.camp}()},
\code{\link{Psh}}
}
\concept{Series datos}
