% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camptoyear.r
\name{camptoyear}
\alias{camptoyear}
\title{Transforma series de nombres de campaña en años}
\usage{
camptoyear(x)
}
\arguments{
\item{x}{Vector con la serie de nombres de campaña a transformar a añoa}
}
\description{
Transforma series de nombres de campañas en formato Camp XYY a años
}
\examples{
camptoyear(Nsh)
}
