% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capwords.r
\name{capwords}
\alias{capwords}
\title{capwords}
\usage{
capwords(s, strict = FALSE)
}
\arguments{
\item{s}{describe argument}

\item{strict}{= FALSE describe argument}
}
\value{
describe return value
}
\description{
add description
}
\details{
other comments found in file
add details
}
\examples{
# insert examples

}
\author{
WEST Inc.
}
\seealso{
\code{\link{related routine}}, \code{\link{related routine}}
}

