% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVlist.camp.r
\name{CVlist.camp}
\alias{CVlist.camp}
\title{Abundancias en número y peso totales para todas las especies de un grupo en una campaña}
\usage{
CVlist.camp(gr = "1", camp, dns, kg = TRUE, dec = 3,
  percents = FALSE, cor.time = TRUE, excl.sect = NA, minim = F)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros. 9 incluye todos los grupos a excepción del 6}

\item{camp}{Campaña de la que se solicitan los resultados un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew" o "Porc", Cantábrico "Cant" o "Cnew", Golfo de Cádiz "Arsa" y Mediterráneo "Medi" P}

\item{kg}{Si T saca los resultados en Kg, si F en gramos}

\item{dec}{Número de decimales de los resultados}

\item{percents}{Si T devuelve los resultados en porcentaje del total}

\item{cor.time}{si T corrige abundancias a duración de lance de 30 minutos}

\item{excl.sect}{Excluye los sectores o subsectores dados como caracteres}

\item{minim}{Excluye por defecto los valores con peso menor de el valor que se elija (atención si kg=T), también se puede hacer a posteriori para cualquiera de los campos con tabla[tabla$NombColumn>XX,]}
}
\value{
Devuelve un data.fram con valores en orden decreciente de abundancia con columnas gr,esp,especie,familia,camp,weight,number,nlan (número de lances en que aparece cada especie)
}
\description{
Función de resultados
}
\examples{
CVlist.camp(gr=1,camp="P06",dns="Porc")
}
