% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVtal.camp.R
\name{CVtal.camp}
\alias{CVtal.camp}
\title{Abundancias, SE y coeficientes de variación paramétricos por subestrato, estrato y total para rango tallas}
\usage{
CVtal.camp(
  gr,
  esp,
  camp,
  dns,
  tmin,
  tmax,
  cor.time = TRUE,
  Nas = FALSE,
  excl.sect = NA
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros. 9 incluye todos los grupos a excepción del 6}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que sacar los dato un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{tmin}{Talla mínima}

\item{tmax}{Talla máxima}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{Nas}{Si F el error estándar de los subestratos y estratos con un solo lance son NA. Si T utiliza el valor del estrato y evita los NAs}

\item{excl.sect}{Excluye los sectores o subsectores dados como caracteres}
}
\value{
Devuelve una lista con valorees de media, SE y CV para los subestratos, estratos y total de la campaña
}
\description{
Función de resultados
}
\seealso{
{\link{CVtal.camps}}, {\link{CVtal.bt.camp}}
}
