% Generated by roxygen2 (4.0.2): do not edit by hand
\name{databEstr}
\alias{databEstr}
\title{Salida con abundancias estratificadas por estrato de profundidad}
\usage{
databEstr(gr, esp, camp, dns = "Cant", cor.time = T, Nas = F)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{Nas}{permite calcular los errores est?ndar aunque sólo haya un lance en algún estrato (haciendo varianza =0 en ese estrato, incorrecto pero da una idea válido cuando sólo un estrato entre varios tiene sólo un lance)}
}
\value{
Devuelve un número con nombres organizado en dos líneas (biomasa y número) en columnas por estratos de profundidad segun definidos en el Camp e información por columnas abundancia estratificada media por estrato (avgestr) y error estándar (SEestr) y totales (avgsect, SEsect). Preparado para pegarlo de año en año en los ficheros excel de abundancia en grupo de trabajo
}
\description{
Salida de datos a csv para rellenar los informes de grupo de trabajo por estrato, filas con datos ab estratificada (Biomasa y N) y error estándar por sector. Da la abundancia y error estándar en cada sector (geográfico definido en el Camp)
}
\examples{
databEstr(1,51,"N05","Cant")
}

