% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagegr.camp.r
\name{datagegr.camp}
\alias{datagegr.camp}
\title{Extrae los datos del FAUNA de una especie en concreto en una edad.}
\usage{
datagegr.camp(
  gr,
  esp,
  camp,
  dns = "Porc",
  plus = 8,
  excl.sect = NA,
  cor.time = TRUE,
  n.ots = FALSE,
  AltAlk = NA,
  incl2 = TRUE,
  mediahora = 1
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustaceos: Sólo existen datos de edad para algunos peces y la cigala.}

\item{esp}{Codigo de la especie numerico o caracter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña o campañas a representar en el mapa}

\item{dns}{Elige el origen de las bases de datos Porcupine "Pnew????" Cantábrico "Cant????" Golfo de Cadiz "Arsa" proporciona los datos para Medits pero no saca mapas}

\item{plus}{Edad plus: incluir la edad considerada como plus, solo afecta si se pide la edad solicitada que une todas las edades mayores}

\item{excl.sect}{excluye sectores si se quiere.}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{n.ots}{Valor interno para decir que no se saque el número de otolitos en la clave sino las proporciones}

\item{AltAlk}{Clave talla edad alternativa sin ruta ni extensión, NA por defecto usa la clave de la campaña edadXYY.dbf}

\item{incl2}{si T coge los datos de lances especiales, si F los excluye}

\item{mediahora}{Valor para obtener abundancias por hora si media hora es mayor}
}
\value{
Devuelve un data.frame con campos lance, lat, long y abundancia por edad 0,1,2...Plus de edad, lance, peso número subestrato...
}
\description{
Función de acceso a datos: saca las abundancias por edad y lance de la especie
}
\seealso{
Other edades: 
\code{\link{ALKs.dns.camp}()},
\code{\link{AbAgStatRec.camp}()},
\code{\link{GetAlk.camp}()},
\code{\link{bubbage.camp}()},
\code{\link{edadsect.camp}()},
\code{\link{edadstr.camp}()},
\code{\link{grafedtal.camps}()},
\code{\link{grafedtal.camp}()},
\code{\link{logabage.camp}()},
\code{\link{logabage2.camp}()},
\code{\link{mapICESStatRec}()},
\code{\link{maphistage}()}
}
\concept{edades}
