% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dathidro.camp.r
\name{dathidro.camp}
\alias{dathidro.camp}
\title{Características del hidrografía de la campaña desde el fichero hidro}
\usage{
dathidro.camp(camp, dns, year = TRUE, quarter = TRUE)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa", combinados con "dnsred" busca los datos en el servidor de Santander si se han creado las odbcs}

\item{year}{si T incluye una columna con el año al final de los datos}

\item{quarter}{si T incluye una columna con el trimestre de los datos teniendo en cuenta la fecha del lance, puede cambiar a mitad de la campaña, cuidado con campañas IBTS adscritas a un trimestre particular.}
}
\value{
Devuelve un data.frame con datos de cada estación hidrografica y la correspondencia a los lances del fichero lance
}
\description{
Función de acceso a datos:
Extrae las características de las hidrología para una campaña determinada
}
\details{
Un problema que ocurre al utilizar el CampR con ficheros dbf de las primeras campañas
puede ser que al fichero lanceXXX.dbf le falte algún campo, habitualmente
el campo \strong{ESTN} utilizado en las últimas versiones del \strong{CAMP} para relacionar los lances con las estaciones de CTD.
El error usual es \strong{$ operator is invalid for atomic vectors}
Si se detecta este error revisar la estructura de lanceXXX.dbf con la de
otros ficheros de lances de los últimos años
}
\examples{
  print(dathidro.camp(Nsh[24],"Cant"))
  print(datlan.camp("P16","Porc",bio=T,outhidro=T))
}
\seealso{
{\link{MapLansGPS}}
}
