% Generated by roxygen2 (4.0.2): do not edit by hand
\name{datlan.camp}
\alias{datlan.camp}
\title{Características del lance}
\usage{
datlan.camp(camp, dns, incl2 = T, incl0 = F, hidro = F, excl.sect = NA,
  redux = F)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)gr Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{incl2}{Si T se incluyen los lances extra no incluidos para las abundancias o biomasas estratificadas}

\item{incl0}{Si T se incluyen los lances nulos}

\item{hidro}{Si T muestra datos de hidrografía. Necesita que exista fichero base de datos de hidro: HIDROXYY.dbf}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{redux}{Si T elimina datos de longitud y latitud de virada y muestra la media de las profundidades de largada y virada}
}
\value{
Devuelve un data.frame con datos de cada lance, las variables dependen de la selección de hidro y redux. En cualquier caso incluye variables weight.time con el factor de calibración para lances con menos tiempo del estándar y arsect: el área del sector al que corresponde el lance dentro del muestreo
}
\description{
Función de acceso a datos:
Extrae las características de los lances para una campaña determinada
}
\examples{
# datlan.camp(Nsh[24:28],"Cant",hidro=F,excl.sect=c("A"))
}
\seealso{
{\link{MapLansGPS}}
}

