% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datmuest.camp.r
\name{datmuest.camp}
\alias{datmuest.camp}
\title{Datos del muestreo de la especie esp en la campaña}
\usage{
datmuest.camp(gr, esp, camps, dns = "Cant", excl.sect = NA)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Utilizar buscacod(gr,esp) para ver codigos}

\item{camps}{Campañas de las que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}
}
\value{
Devuelve un data.frame con datos del muestreo en los años/campañas elegidas y contendio: nombre especie,campaña, número de lances en la campaña,
número de lances con la especie,  pesos y número totales muestreados, numero total capturado, peso medio de los bichos medidos y el rango de tallas,
talla mínima, talla máxima y número total capturado
}
\description{
Saca un resumen de los ejemplares y peso muestreados de la especie esp en la campaña camp
}
\examples{
 datmuest.camp(2,19,c("P16","P17"),"Porc")
datmuest.camp(1,50,Psh,"Porc")
}
