% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dattal.camp}
\alias{dattal.camp}
\title{Abundancia estratificada por talla y sexo}
\usage{
dattal.camp(gr, esp, camp, dns, cor.time = T, excl.sect = NA, sx = T)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{sx}{Si T muestra los datos por sexo}
}
\value{
Devuelve un data.frame con variables: talla, machos, hembras e indet(erminados) si existen todos y si sx=T
}
\description{
Función de acceso a datos:
Extrae los datos de abundancia por talla y sexo de una especie o conjunto de especies a partir de las distribuciones de talla
}
\seealso{
datos.camp {\link{datos.camp}}
}

