% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dattalb.campb.r
\name{dattal.campb}
\alias{dattal.campb}
\title{Abundancia estratificada por talla y sexo forzando a que sean en cm. Para datos medidos en mm o .5}
\usage{
dattal.campb(
  gr,
  esp,
  camp,
  dns,
  cor.time = TRUE,
  cor.5 = FALSE,
  excl.sect = NA,
  sex = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{cor.5}{si T cambia las distribuciones de tallas .5 a 1 cm, sólo sirve para las campañas/peces medidos a 1 cm .5 o todos a .5, distingue si una campaña es a .5 y otra a 1}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{sex}{Si T muestra los datos por sexo}

\item{verbose}{Si T muestra avisos problemas de tallas entre distintas especies}
}
\value{
Devuelve un data.frame con variables: talla e indeterminados, no
}
\description{
Función de acceso a datos:
Extrae los datos de abundancia por talla y sexo de una especie o conjunto de especies a partir de las distribuciones de talla
}
\examples{
dattal.campb("1"," 18","N24","Cant")
}
\seealso{
{\link{datos.camp}}
}
