% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datTalaGeo.camp.r
\name{datTalaGeo.camp}
\alias{datTalaGeo.camp}
\title{Datos de tallas y sexo con geografía (para uso en especies a evaluar con datos de tallas: L. circularis, Phycis...)}
\usage{
datTalaGeo.camp(
  gr,
  esp,
  camps,
  dns,
  cor.time = TRUE,
  excl.sect = NA,
  sex = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{sex}{Si T muestra los datos por sexo}

\item{verbose}{Si T muestra avisos problemas de tallas entre distintas especies}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}
}
\value{
Devuelve un data.frame con variables: Species,	year,	survey,	station,	length,	sex,	number,	latitude,	longitude,	depth,
}
\description{
Función de transformación de datos:
Extrae los datos de abundancia por talla y sexo de una especie y les añade los datos de su ubicación geográfica y profundidad
}
\examples{
datTalaGeo.camp(1,87,Nsh[10:38],"Cant")
}
\seealso{
{\link{datos.camp}}
}
