% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dattalgr.camp}
\alias{dattalgr.camp}
\title{Biomasa y abundancia para un rango de talla}
\usage{
dattalgr.camp(gr, esp, camp, dns = "Pnew", tmin = 1, tmax = 999,
  cor.time = T, sx = NA, ind = "n")
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)gr Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{tmin}{Talla mínima}

\item{tmax}{Talla máxima}

\item{cor.time}{Si T corrige abundancias con la duración del lance para llevarlo a 30 minutos}

\item{sx}{Permite elegir entre machos(1), hembras(2) o indeterminados(3), NA escoge sin tener en cuenta el sexo}

\item{ind}{Parámetro a representar saca los datos en "p"eso o "n"úmero}
}
\description{
Muestra los datos de abundancia de una especie o conjunto de especies de un rango de tallas determinado a partir de las distribuciones de tallas. También puede mostrar los datos de biomasa a partir de la relación talla-peso
}
\examples{
dattalgr.camp("1",c(44:45),"N94","Cant",0,45,ind="p")
}
\seealso{
datgr.camp {\link{datgr.camp}}
}

