% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dattalmean.camp.R
\name{dattalmean.camp}
\alias{dattalmean.camp}
\title{\strong{Talla media} y \strong{percentil} de la distribución de tallas por lance en la campaña \strong{Xxx}}
\usage{
dattalmean.camp(
  gr,
  esp,
  camp,
  dns = "Porc",
  quant = 0.5,
  cor.time = TRUE,
  incl2 = TRUE,
  sex = NA
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña a representar en el mapa de un año comcreto (xx): Demersales "Nxx", Porcupine "Pxx", Arsa primavera "1XX", Arsa otoño "2xx" Medits "Mxx"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa" y Mediterráneo "Medi", gr Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{quant}{elije el percentil deseado, 0.5 para la mediana}

\item{cor.time}{Si T corrige abundancias con la duración del lance para llevarlo a 30 minutos}

\item{incl2}{Si F no tiene en cuenta los lances especiales, si T si los tiene en cuenta, pero da problemas por que no puede calcular las abundancias estratificadas}

\item{sex}{Permite elegir entre machos(1), hembras(2) o indeterminados(3), NA escoge sin tener en cuenta el sexo}
}
\description{
Muestra la talla media, el percentil \strong{quant} de la distribuicón de tallas de las espcie \emph{esp} en cada lance en la campaña seleccionada
}
\examples{
dattalmean.camp(gr=1,esp=44,camp="N94",dns="Cant",quant=0.85)
}
\seealso{
{\link{dattalgr.camp}}
}
