% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decgrad.R
\name{decgrad}
\alias{decgrad}
\title{Transforma grados y minutos decimales a grados decimales}
\usage{
decgrad(x)
}
\arguments{
\item{x}{vector con grados y minutos en formato decimal}
}
\description{
Transforma grados y minutos decimales a grados decimales, avisa si el dato es incorrecto, es decir minutos más de 60
}
\examples{
decgrad(43.7923)
}
\seealso{
Other Conversion unidades: 
\code{\link{HdectoH}()},
\code{\link{gradec}()},
\code{\link{gradms}()}
}
\concept{Conversion unidades}
