% Generated by roxygen2 (4.0.2): do not edit by hand
\name{denstall.camp}
\alias{denstall.camp}
\title{Función de densidad distribución de tallas}
\usage{
denstall.camp(gr, esp, camp, dns, ti = F, bw = T, es = T, plot = T,
  idi = "l", layout = NA, years = F, cexleg = 1)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos. 4 equinodermos y 5 invertebrados no se miden}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{ti}{Si T añade título al gráfico, el nombre de la especie en latín}

\item{bw}{Gráfico en blanco y negro si T o en color si F}

\item{es}{Si T gráfico en castellano, si F gráfico en inglés}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}

\item{years}{Si T saca los años como nombre de campaña en los paneles lattice de campañas}

\item{cexleg}{Varía el tamaño de letra de los ejes y del número de la leyenda}
}
\description{
Crea gráfico con la función de densidad de la distribución de tallas estratificada para la especie y las campañas seleccionadas
}

