% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtall.campa.r
\name{dtall.campa}
\alias{dtall.campa}
\title{Crea un histograma distribución de tallas estratificada especie y campañas corrigiendo el que haya unas campañas a .5 y otras a 1 cm}
\usage{
dtall.campa(
  gr,
  esp,
  camp,
  dns,
  cor.time = TRUE,
  cor.5 = FALSE,
  ti = FALSE,
  sub = NA,
  leg = TRUE,
  cexleg = 1,
  bw = TRUE,
  es = TRUE,
  sex = FALSE,
  plot = T,
  idi = "l",
  clms = 2,
  layout = NA,
  excl.sect = NA,
  ymax = NA,
  out.dat = FALSE,
  years = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo. Si se saca más de una especie avisa de que distintas especies pueden estar medidas en diferentes unidades, pero saca títulos y datos como si todos tuvieran las propiedades de la primera.}

\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{cor.5}{si T cambia las distribuciones de tallas .5 a 1 cm, sólo sirve para las campañas/peces medidos a 1 cm .5 o todos a .5, distingue si una campaña es a .5 y otra a 1}

\item{ti}{Si T añade título al gráfico, el nombre de la especie en latín}

\item{sub}{Si caracter lo añade como subtítulo en el gráfico}

\item{leg}{Si T añade leyenda}

\item{cexleg}{Varía el tamaño de letra de los ejes y del número de la leyenda}

\item{bw}{Gráfico en blanco en negro si T o en color si F}

\item{es}{Si T gráfico en castellano, si F gráfico en inglés}

\item{sex}{Sacar la distribución de tallas por sexos si existen (T), o conjunta (F)}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{clms}{Número de columnas para ordenar la serie histórica}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{ymax}{Valor máximo del eje y}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{years}{Si T saca los años como nombre de campaña en los paneles lattice de campañas}

\item{verbose}{Si T saca avisos de consistencia en tallas, sino los omite}
}
\value{
Si plot=T saca el gráfico, pero si out.dat=T puede exportar una matriz talla(filas)xCampañas(columnas)
}
\description{
Histograma de la distribución de tallas media estratificada por sexos a partir de ficheros del camp
}
\examples{
dtall.campa(5,15,"N24","Cant",es=F,sex=F,cor.5=T,out.dat=T)
dtall.campa(1,111,c("P22","P23"),"Porc",cor.5=T,sex=F)
}
\seealso{
Other Distribuciones de tallas: 
\code{\link{denstall.camp}()},
\code{\link{dtall.campb}()},
\code{\link{dtall.camp}()},
\code{\link{dtall.lan}()},
\code{\link{dtallan.camp}()},
\code{\link{dtallan.peso}()},
\code{\link{dtallbarplot}()}
}
\concept{Distribuciones de tallas}
