% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dtallan.camp}
\alias{dtallan.camp}
\title{Valores absolutos en número por talla}
\usage{
dtallan.camp(gr, esp, camp, dns, lances = NA, sex = F)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{lances}{Da la opción de escribir un número de lance y saca los valores solo para ese lance o grupo de lances.}

\item{sex}{por defecto (F) suma todos los individuos como indet. T saca los datos por sexo si los hay, no afecta si sólo hay indeterminados (3)}
}
\description{
Da Valores absolutos (totales por lance o grupos de lances) en número por talla y sexo (si lo hay)
}
\examples{
dtallan.camp(gr=1,esp=9,camp="N02",dns="Cant",lances=50)
}

