% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtallan.peso.R
\name{dtallan.peso}
\alias{dtallan.peso}
\title{Peso de la distribución de tallas de esp en los lances solicitados}
\usage{
dtallan.peso(gr, esp, camp, dns = "Cant", lances = NA, sex = TRUE,
  areg = NA, breg = NA)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña para representar el histograma de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{lances}{Número de lance o lances de los que se extrae el dato de peso de la especie solicitada. Si lances se deja en NA, da todos los lances de la campaña.}

\item{sex}{por defecto (T) da valores por sexos si los hay, si F suma todos, si todos son indeterminados no afecta.}

\item{areg}{Para cambiar el coeficiente "a" de la regresión si es distinto al archivado en el CAMP.}

\item{breg}{Para cambiar el coeficiente "b" de la regresión si es distinto al archivado en el CAMP.}
}
\description{
Da el peso de la distribución de tallas de las capturas de la especie en los lances solicitados
}
\examples{
dtallan.peso(gr=1,esp=9,camp="N02",dns="Cant",lance=50)
}
\seealso{
Other Distribuciones de tallas: \code{\link{denstall.camp}},
  \code{\link{dtall.lan}}, \code{\link{dtallan.camp}},
  \code{\link{dtallbarplot}}
}
