% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtallboxplot.camps.R
\name{dtallboxplot.camps}
\alias{dtallboxplot.camps}
\title{\strong{Talla media} de la distribución de tallas media estandarizada del conjunto de la campaña \strong{Xxx}}
\usage{
dtallboxplot.camps(
  gr,
  esp,
  camps,
  dns = "Porc",
  cor.time = TRUE,
  cor.5 = T,
  excl.sect = NA,
  plot = TRUE,
  es = FALSE,
  idi = "l",
  cexleg = 1,
  col = "steelblue",
  ymax = NA,
  notch = TRUE,
  outline = FALSE,
  ti = TRUE,
  sub = TRUE,
  graf = FALSE,
  xpng = 1200,
  ypng = 800,
  ppng = 15
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa" y Mediterráneo "Medi", gr Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{cor.time}{Si T corrige abundancias con la duración del lance para llevarlo a 30 minutos}

\item{ymax}{si F el límite superior del gráfico es el percentil X de la distribución de tallas, si es un valor pone el valor dado en la opción}

\item{notch}{si T incluye el notch en el boxplot}

\item{outline}{si T incluye el los outliers en el boxplot}

\item{ti}{Si T añade título al gráfico, el nombre de la especie en latín}

\item{graf}{si F el gráfico sale en la pantalla, si nombre fichero va a fichero en el directorio de trabajo del Rstudio ver getwd()}

\item{xpng}{width archivo png si graf es el nombre del fichero}

\item{ypng}{height archivo png si graf es el nombre del fichero}

\item{ppng}{points png archivo si graf es el nombre del fichero}

\item{camp}{Campaña a representar en el mapa de un año comcreto (xx): Demersales "Nxx", Porcupine "Pxx", Arsa primavera "1XX", Arsa otoño "2xx" Medits "Mxx"}
}
\value{
Devuelve un data.frame con campos: camp,mean.size
}
\description{
Muestra la talla media de la distribución de tallas de la especie \emph{esp} el conjunto de la campaña seleccionada
}
\examples{
dtallboxplot.camps(gr=1,esp=50,camps=Nsh,dns="Cant")
dtallmean.camp(gr=2,esp=19,camps=Psh,dns="Porc",graf="miratu")
}
\seealso{
{\link{dattal.camp}}
}
