% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ecolgr.camp}
\alias{ecolgr.camp}
\title{Índices ecológicos para un grupo en una campaña}
\usage{
ecolgr.camp(gr, esp = 999, camp, dns = "Pnew", ind = "n")
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{ha de ser 999 cuando se quiere incluir todas las especies del grupo, o elegir todas las especies deseadas con los codigos de las especies}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{ind}{Elige el valor sobre el que se calculan los índices de diversidad, dominancia....}
}
\value{
Devuelve un data.frame con campos lan,lat,long,prof,div(Diversidad de Shanon),numbesp (riqueza: número de especies),simp (diversidad de simpson),domsimp (indice de dominancia de simpson).
}
\description{
Utiliza los datos del Camp para calcular la riqueza, diversidad y dominancia
}
\examples{
ecolgr.camp(1,999,"P08","Pnew",ind="n")
}
\seealso{
MapEcol.camp {\link{MapEcol.camp.r}}
}

