% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandUnmarked.R
\name{expandUnmarked}
\alias{expandUnmarked}
\title{expandUnmarked}
\usage{
expandUnmarked(dat, colKeep, colRep)
}
\arguments{
\item{dat}{describe argument}

\item{colKeep}{describe argument}

\item{colRep}{describe argument}
}
\value{
describe return value
}
\description{
Jared Studyvin
 15 Jan 2016
 replicate the rows of a dataframe based on one of the columns
}
\details{
other comments found in file
 colRep = single string of the frequency to repeat the data
}
\examples{
\dontrun{
#insert examples

}
}
\author{
WEST Inc.
}
\seealso{
\code{\link{related routine}}, \code{\link{related routine}}
}

