% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_passage.r
\name{F.annual.passage}
\alias{F.annual.passage}
\title{F.annual.passage}
\usage{
F.annual.passage(site, taxon, run, min.date, max.date, output.file, ci = TRUE)
}
\arguments{
\item{site}{describe argument}

\item{taxon}{describe argument}

\item{run}{describe argument}

\item{min.date}{describe argument}

\item{max.date}{describe argument}

\item{output.file}{describe argument}

\item{ci}{TRUE  describe argument}
}
\value{

}
\description{
Estimate passage for all years in the data base.

   site = site number
   taxon = taxon number
   run = run number
   output.file = name of .CSV file to put passage estimates into.  Full path is allowed.

   Annual production is estimated for every year in the data base, from minimum date to maximum 
   date.  "years" are (min.date to min.date+365days), (min.date+365days, min.date+2(365)days), etc...


   ---- Pull the min and max dates from the data base
}
\details{

}
\examples{


}
\author{
WEST Inc.
}
\seealso{
\code{\link{}}, \code{\link{}}
}

