% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_1dim.r
\name{F.assign.1dim}
\alias{F.assign.1dim}
\title{F.assign.1dim}
\usage{
F.assign.1dim(catch, present.var, absent.var)
}
\arguments{
\item{catch}{<describe argument>}

\item{present.var}{<describe argument>}

\item{absent.var}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Values in one dimension of the basic run X lifstage table is present, the other is missing.

   Assign the missing one based on frequency dist of the other.


   We make sure to TOSS RANDOMSELECTION == "NO"
}
\details{

}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

