% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_2dim.r
\name{F.assign.2dim}
\alias{F.assign.2dim}
\title{F.assign.2dim}
\usage{
F.assign.2dim(catch, var1, var2)
}
\arguments{
\item{catch}{<describe argument>}

\item{var1}{<describe argument>}

\item{var2}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Values are missing in two dimensions.  I.e., var1 and var2 are missing, but n>0.

   Assign a var1 and var2 value based on joint distribution
}
\details{
<other comments found in file>
   Idea is to make a new variable with joint levels, then assign1dim
   Call the routine that assigns plus counts based on a 1 dimensional distribution    
   Now, put the new assignments back in FinalRun and lifeStage
 Finally, drop the columns we created.
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

