% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_batch_date.r
\name{F.assign.batch.date}
\alias{F.assign.batch.date}
\title{F.assign.batch.date}
\usage{
F.assign.batch.date(df)
}
\arguments{
\item{df}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Assign batch data to records.

 df <- catch                      # df <- nvCatch2
}
\details{
<other comments found in file>
min.day <- as.POSIXlt( min(df$EndTime) - 24*60*60, format="%Y-%m-%d %H:%M:%S", tz=time.zone)
max.day <- as.POSIXlt( max(df$EndTime) + 2*24*60*60, format="%Y-%m-%d %H:%M:%S", tz=time.zone)
   Bin the sampleEnd's to cut.seq
   Establish when the cut time "wraps" to the next day.  I.e., at some point, as cut time increases from 0, you
   stop calling the sample from the night before, and attribute it to the current night.  This time is set in 
   wrap.POSIX.
   Compute batchDate    
   If we allow biologists to over ride batchDate, uncomment the following code
   Figure out which batchDates are missing and replace just those.
miss <- is.na(df$batchDate)
df$batchDate[miss] <- bDate[miss]
   Add batch date to data frame
   overwrite with batchDate
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

