% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_gaplen.r
\name{F.assign.gaplen}
\alias{F.assign.gaplen}
\title{F.assign.gaplen}
\usage{
F.assign.gaplen(df)
}
\arguments{
\item{df}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Function to look for and assign gaps in sampling


   Make sure data base is sorted
}
\details{
<other comments found in file>
NA
   Look for gaps by trapPositionID
}
\examples{
\dontrun{
#insert examples

}
}
\author{
WEST Inc.
}
\seealso{
\code{\link{related routine}}, \code{\link{related routine}}
}

