% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biweekly_report.r
\name{F.biweekly.report}
\alias{F.biweekly.report}
\title{F.biweekly.report}
\usage{
F.biweekly.report(site, min.date, max.date)
}
\arguments{
\item{site}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Generate a bi-weekly report
}
\details{
<other comments found in file>
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

