% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_passage.r
\name{F.bootstrap.passage}
\alias{F.bootstrap.passage}
\title{F.bootstrap.passage}
\usage{
F.bootstrap.passage(grand.df, catch.fits, catch.Xmiss, catch.gapLens,
  catch.bDates.miss, eff.fits, eff.X, eff.ind.inside, eff.X.dates, sum.by, R,
  ci = T)
}
\arguments{
\item{grand.df}{<describe argument>}

\item{catch.fits}{<describe argument>}

\item{catch.Xmiss}{<describe argument>}

\item{catch.gapLens}{<describe argument>}

\item{catch.bDates.miss}{<describe argument>}

\item{eff.fits}{<describe argument>}

\item{eff.X}{<describe argument>}

\item{eff.ind.inside}{<describe argument>}

\item{eff.X.dates}{<describe argument>}

\item{sum.by}{<describe argument>}

\item{R}{<describe argument>}

\item{ci=T}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Bootstrap or Monte Carlo simulate data sufficient to compute confidence intervals
   for passage.

   catch.fits = list of Poisson fitted objects used to imput missing catches.  One per trap.  i.e., list
       has length equal to number of traps at this site.
   catch.Xmiss = X matrix for days where catch is missing.  This is constructed in F.catch.model, and passed into here.
   catch.bDates.miss = batchDates for missing catches.  Needed because it is possible for 2 gaps to be on same batch date.
   eff.fits = list of Binomial logistic regression fitted objects used to compute efficiency.  One per trap.
   R = number of Monte Carlo iterations to do.
}
\details{
<other comments found in file>
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

