% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Report_Criteria.r
\name{F.buildReportCriteria}
\alias{F.buildReportCriteria}
\title{F.buildReportCriteria}
\usage{
F.buildReportCriteria(site, min.date, max.date)
}
\arguments{
\item{site}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}
}
\value{
<describe return value>
}
\description{
This updates a table in the Access file to contain the trapVisitID's we want. 
   i.e., the trapVisitID's for a particular site between min.date and max.date
}
\details{
<other comments found in file>
NA
 testing
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

