% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Report_Criteria_Release.r
\name{F.buildReportCriteriaRelease}
\alias{F.buildReportCriteriaRelease}
\title{F.buildReportCriteriaRelease}
\usage{
F.buildReportCriteriaRelease(site, min.date, max.date)
}
\arguments{
\item{site}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}
}
\value{
<describe return value>
}
\description{
This updates a table in the Access file to contain the trapVisitID's for releases.
   i.e., the trapVisitID's for a particular site between min.date and max.date.
   Difference between this and the other buildReportCriteria is this for releases and subsequent recaptures.
}
\details{
<other comments found in file>
NA
NA
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

