% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinook_by_date.r
\name{F.chinookByDate.table}
\alias{F.chinookByDate.table}
\title{F.chinookByDate.table}
\usage{
F.chinookByDate.table(site, min.date, max.date, output.file)
}
\arguments{
\item{site}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}

\item{output.file}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Summarize chinook by date between two dates



   ---- Retrieve all chinook records between two dates. 
        run does not work at this point
}
\details{
<other comments found in file>
NA
   ---- Send messages back to the interface
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

