% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_model.enh.r
\name{F.efficiency.model.enh}
\alias{F.efficiency.model.enh}
\title{F.efficiency.model.enh}
\usage{
F.efficiency.model.enh(obs.eff.df, plot = T, max.df.spline = 4,
  plot.file = NA)
}
\arguments{
\item{obs.eff.df}{A data frame with at least variables \code{batchDate} and 
\code{efficiency}, where \code{efficiency} is \code{NA} for all days 
requiring an estimate.}

\item{plot}{A logical indicating if raw efficiencies and the model(s)
should be plotted.}

\item{max.df.spline}{The maximum degrees of freedom allowed for splines.}

\item{plot.file}{The name of the file prefix under which output is to be 
saved.  Set to \code{NA} to plot to the plot window.}
}
\value{
Results from running the enhanced efficiency model fitting process 
  include \code{csv}s of efficiency trials missing a covariate, for each
  \code{TrapPositionID} at a \code{subSiteID}.  This prevents these
  efficiency trials from inclusion in the model fitting process.
  
  Each \code{TrapPositionID} also results in a series of plots depicting the
  fitted temporal spline, at each point of the backwards-fitting process. 
  These could number many, depending on the number of covariates available
  for possible exlcusion.
  
  Each \code{trapPositionID} also outputs a \code{png} containing model
  fitting information, including plots of efficiency versus each considered
  covariate, along with plotted temporal trends of each covariate against
  time.  Additional plots include the final fitted temporal spline (along
  with an prediction "curve" derived from the available data), as well as a
  final "plot" depicting model summary statistics, obtained via the
  \code{summary} function against the logistic efficiency-trial model fits.
  
  Note that no passage estimates result from the fitting of enhanced
  efficiency models.  This is because bootstrapping does not occur, but also
  because estimation of passage is not the goal of the model fitting process.
  Use the regular function sequence; i.e., functions without the
  \code{".enh"} to estimation passage for one-year intervals of interest.
}
\description{
Estimate the efficiency model from a data frame containing 
efficiency trials.
}
\examples{
\dontrun{
#   ---- Fit an efficiency model for each unique trapPositionID 
#   ---- in data frame obs.eff.df.  
F.enh.efficiency.model( obs.eff.df, plot=T, max.df.spline=4, plot.file=NA)
}
}
\seealso{
\code{run_passage.enh}, \code{get_release_data},
  \code{est_passage.enh}, \code{est_efficiency.enh}, \code{eff_model.enh}
}
\author{
WEST Inc.
}
