% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_catch.r
\name{F.est.catch}
\alias{F.est.catch}
\title{F.est.catch}
\usage{
F.est.catch(catch.df, plot = TRUE, plot.file = "raw_catch.pdf")
}
\arguments{
\item{catch.df}{<describe argument>}

\item{plot=TRUE}{<describe argument>}

\item{plot.file="raw_catch.pdf"}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Estimate capture for every day of the season.
   This assumes that only species and life stages we want are
   in catch.df.  I.e., that appropriate filtering has already been done.

   Input:
   catch.df = data frame resulting from call to F.get.catch.data.  Run season is
       an attribute of this data frame

   Output:
   A data frame containing $batchDay and $catch

  plot.file=file.root
}
\details{
<other comments found in file>
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

