% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_catch_trapN.r
\name{F.est.catch.trapN}
\alias{F.est.catch.trapN}
\title{F.est.catch.trapN}
\usage{
F.est.catch.trapN(catch.df, plot = TRUE, plot.file = "raw_catch.pdf")
}
\arguments{
\item{catch.df}{A data frame, possibly restricted to one \code{lifeStage},
resulting from a call to \code{F.get.catch.data}.  Run season is an
attribute of this data frame.}

\item{plot}{A logical indicating if catch is to be plotted over time, per
trap.}

\item{plot.file}{The name to which a graph of catch is to be output, if 
\code{plot=TRUE}.}
}
\value{
A list \code{ans} containing catch model information for each trap 
  listed in \code{catch.df}.  See Details.
}
\description{
Estimate catch for every day of the season, per trap.
}
\details{
Function \code{F.est.catch.trapN} performs exactly the same 
  functionality as function \code{F.est.catch}.  However,
  \code{F.est.catch.trapN} utilizes the full observed catch dataset in its
  implementation, i.e., including all preceding and antecedent strings of
  zeros, for each individual trap contained in the data.  Via the use of
  these modified data, accurate data regarding the number of deployed traps,
  per day, are obtained.
}
\examples{
\dontrun{
#   ---- Estimate catch for each unique trap in data 
#   ---- frame catch.df.  Also output a plot.  
F.est.catch.trapN(catch.df, plot=TRUE, plot.file="raw_catch.pdf")
}
}
\author{
WEST Inc.
}
\seealso{
\code{F.est.catch}
}

