% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_passage.enh.r
\name{F.est.passage.enh}
\alias{F.est.passage.enh}
\title{F.est.passage.enh}
\usage{
F.est.passage.enh(catch.df, release.df, summarize.by, file.root, ci)
}
\arguments{
\item{catch.df}{A data frame with one row per \code{trapvisitID} for a 
particular \code{FinalRun} and \code{lifeStage}.}

\item{release.df}{A data frame resulting from a call to function 
\code{F.get.release.data}.  Contains efficiency data.}

\item{summarize.by}{A text string indicating the temporal unit over which 
daily estimated catch is to be summarized.  Can be one of \code{"day"}, 
\code{"week"}, \code{"month"}, or \code{"year"}.}

\item{file.root}{A text string indicating a prefix to append to all output.}

\item{ci}{A logical indicating if bootstrapped confidence intervals should be
estimated along with passage estimates.  The default is 95\%, although
levels other than 95\% can be set in function \code{F.bootstrap.passage}.}
}
\value{
Results from running the enhanced efficiency model fitting process 
  include \code{csv}s of efficiency trials missing a covariate, for each
  \code{TrapPositionID} at a \code{subSiteID}.  This prevents these
  efficiency trials from inclusion in the model fitting process.
  
  Each \code{TrapPositionID} also results in a series of plots depicting the
  fitted temporal spline, at each point of the backwards-fitting process. 
  These could number many, depending on the number of covariates available
  for possible exlcusion.
  
  Each \code{trapPositionID} also outputs a \code{png} containing model
  fitting information, including plots of efficiency versus each considered
  covariate, along with plotted temporal trends of each covariate against
  time.  Additional plots include the final fitted temporal spline (along
  with an prediction "curve" derived from the available data), as well as a
  final "plot" depicting model summary statistics, obtained via the
  \code{summary} function against the logistic efficiency-trial model fits.
  
  Note that no passage estimates result from the fitting of enhanced
  efficiency models.  This is because bootstrapping does not occur, but also
  because estimation of passage is not the goal of the model fitting process.
  Use the regular function sequence; i.e., functions without the
  \code{".enh"} to estimation passage for one-year intervals of interest.
}
\description{
Compute passage estimates, given catch and efficiency trial
  data.
}
\section{Functions Used to Fit Enhanced Efficiency Models}{
 
  Five programs make up the specialized procedure for fitting enhanced
  efficiency models.  This means actually compiling the data of efficiency 
  trials obtained over several years, and then fitting a generalized additive
  model (GAM) to those data.  All five programs have suffixes of 
  \code{".enh"}, and originated with the program versions without the suffix.
  As such, they are very similar to the originals.
  
  The first, \code{run_passage.enh} corrals the fitting.  It is different 
  from \code{run_passage} in that all of the passage summary that is usually 
  created has been suppressed.  This is because there is no need to 
  bootstrap, once enhanced efficiency models have been obtained.
  
  The second, \code{get_release_data.enh} modifies the obtaining of release 
  data, so as to obtain astrological data and mean fork-length.  It is very 
  similar to its originator.
  
  The third, \code{est_passage.enh}, corrals the data from 
  \code{get_release_data.enh} for use in \code{est_efficiency.enh}.  It too 
  should be very similar to its originator.
  
  The fourth, \code{est_efficiency.enh}, ensures the calculation of weighted 
  averages for the three efficiency-trial covariates have to do with mean 
  fork-lengths, and percent of fishing performed at night, or while the moon 
  is up.  It also emulates closely its originator.
  
  Finally, the fifth, \code{eff_model.enh}, fits the enhanced efficiency 
  models.  It follows a backwards selection procedure, allowing for both 
  variable covariate selection, as well as variable temporal spline 
  complexity.  It creates graphical output, for each trap, so as to provide 
  further hypothesis generation.
}

\examples{
\dontrun{
#   ---- Estimate passage based on a given set of 
#   ---- catch and release dataframes, over weeks.
thePassage <- F.est.passage(catch.df, release.df, "week", "myFileRoot", ci=TRUE )
}
}
\seealso{
\code{run_passage.enh}, \code{get_release_data},
  \code{est_passage.enh}, \code{est_efficiency.enh}, \code{eff_model.enh}
}
\author{
WEST Inc.
}
