% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_plus_counts.r
\name{F.expand.plus.counts}
\alias{F.expand.plus.counts}
\title{F.expand.plus.counts}
\usage{
F.expand.plus.counts(catch)
}
\arguments{
\item{catch}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Expand the data set in 'catch' to account for plus counts.  Plus counts are lines with missing FinalRun or lifeStage.


print( catch[is.na(catch$FinalRun),] )

   On entry there cannot be any rows with Unmarked = 0 or with missing Unmarked. 
   The Unmarked == 0 correspond to unsuccessful visits.  Missing Unmarked correspond to non-fishing. 
   Eliminate these lines if they exist.
}
\details{
<other comments found in file>
readline()
   Assign plus counts one dimension at a time, then for the double missing


   ---- Deal with missing run, but present lifestage (i.e., Assign a run)
   ---- Deal with missing lifeStage, but present finalRun (i.e., Assign a life stage)
   ---- Deal with missing lifeStageID AND missing finalRunID (i.e., assign a life stage and a run)
source("assign_1dim.r")
source("assign_2dim.r")
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

