% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_fish_data.r
\name{F.get.all.fish.data}
\alias{F.get.all.fish.data}
\title{F.get.all.fish.data}
\usage{
F.get.all.fish.data(site, min.date, max.date)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}
}
\value{
A data frame containing all catch, regardless of taxon, between the 
  specified dates.
}
\description{
Fetch all fish data, regardless of taxon, from an Access data 
  base between two dates.
}
\details{
This is a generalization of \code{F.get.indiv.fish.data} which
  returns records for a single taxon.  Function \code{F.get.all.fish.data}
  utilizes the all-catch query series in order to obtain the proper data.
}
\examples{
\dontrun{
# Obtain all catch on the American between the specified dates.  
F.get.all.fish.data(57000,"2014-01-01","2014-06-30")
}
}
\author{
WEST Inc.
}
\seealso{
\code{F.get.indiv.fish.data}
}

