% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indiv_fish_data.r
\name{F.get.indiv.fish.data}
\alias{F.get.indiv.fish.data}
\title{F.get.indiv.fish.data}
\usage{
F.get.indiv.fish.data(site, taxon, run, min.date, max.date, keep = "unmarked")
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{run}{The text seasonal identifier.  This is a one of \code{"Spring"}, \code{"Fall"},
\code{"Late Fall"}, or \code{"Winter"}.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{by}{A text string indicating the temporal unit over which daily
estimated catch is to be summarized.  Can be one of \code{day},
\code{week}, \code{month}, \code{year}.}

\item{keep="unmarked"}{A text string specifying the type of fish to retain.
\code{keep="unmarked"} keeps all fish without efficiency trail marks, i.e.,
all fish not in efficiency trial.  \code{keep="marked"} keeps only fish
that were involved in an efficiency trial. \code{keep="all"}, i.e.,
anything else, will keep all fish records --- both marked and unmarked.}
}
\value{

}
\description{
Fetch the fish data for a single taxon from an Access data base. The resulting data 
   set has one line per group of fish with the same forklength.
}
\details{
To be included in the catch data, a record has to be from the site, 
   of the correct taxon, of the correct run, and between min and max date. F.size.by.date
}
\examples{
\dontrun{
<insert examples>

}
}
\author{
WEST Inc.
}
\seealso{
\code{F.get.catch.data}, \code{F.sql.error.check}
}

