% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indiv_visit_data.r
\name{F.get.indiv.visit.data}
\alias{F.get.indiv.visit.data}
\title{F.get.indiv.visit.data}
\usage{
F.get.indiv.visit.data(site, run, min.date, max.date)
}
\arguments{
\item{site}{<describe argument>}

\item{run}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Fetch the visit data for a particular site and run from an Access data base. The resulting data
   set should have one line for each visit to a trap at the site. .

   input:
   db = full path and name of the Access data base to retrieve data from
   tables = vector with named components containing names
           of the table in db to pull values from
   site = site ID of place we want to do estimates for.  Remember, site encompases multiple traps.   Individual screw 
           traps are subsites.  Project could be multiple sites.
   run = the single run ID of the fish we want.  If run = NA, all records  will be pulled.

       YOU DO NOT ACTUALLY NEED RUN HERE.  IT MAKES NO DIFFERENCE, TAKE THIS OUT OF PARAMETERS TO THIS CALL.

   min.date and max.date = minimum and maximum dates for data to be included. Format = "YYYY-MM-DD"



   *******
   First, put the start and stop dates of the run into a list
   Need these to filter visits.
}
\details{
This routine is very similar to get_catch_data, except that it does not average over subsites.
   Raw catches are returned.

   Here is the way to interpret visitTime and visitTime2
    "visit start" field (named "visitTime")
    for visitTypeID = 1 Date/time of arrival at trap
    for visitTypeID = 2 Date/time the trap was emptied
    for visitTypeID = 3 Date/time the trap was emptied
    for visitTypeID = 4 Date/time the trap was emptied
    for visitTypeID = 5 Date/time of arrival at trap
    for visitTypeID = 6 Date/time of arrival at trap


    "visit end" field (named "visitTime2")
    for visitTypeID = 1 Date/time the trap began fishing
    for visitTypeID = 2 Date/time the trap began fishing
    for visitTypeID = 3 Date/time the trap began fishing
    for visitTypeID = 4 Date/time of the end of the visit
    for visitTypeID = 5 Date/time of the end of the visit
    for visitTypeID = 6 Date/time of the end of the visit
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

