% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_release_data.enh.r
\name{F.get.release.data.enh}
\alias{F.get.release.data.enh}
\title{F.get.release.data.enh}
\usage{
F.get.release.data.enh(site, taxon, min.date, max.date, visit.df)
}
\arguments{
\item{site}{The identification number of the site for which estimates are
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{min.date}{The start date for data to include. This is a text string in
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as
\code{min.date}.}
}
\value{
Results from running the enhanced efficiency model fitting process 
  include \code{csv}s of efficiency trials missing a covariate, for each
  \code{TrapPositionID} at a \code{subSiteID}.  This prevents these
  efficiency trials from inclusion in the model fitting process.
  
  Each \code{TrapPositionID} also results in a series of plots depicting the
  fitted temporal spline, at each point of the backwards-fitting process. 
  These could number many, depending on the number of covariates available
  for possible exlcusion.
  
  Each \code{trapPositionID} also outputs a \code{png} containing model
  fitting information, including plots of efficiency versus each considered
  covariate, along with plotted temporal trends of each covariate against
  time.  Additional plots include the final fitted temporal spline (along
  with an prediction "curve" derived from the available data), as well as a
  final "plot" depicting model summary statistics, obtained via the
  \code{summary} function against the logistic efficiency-trial model fits.
  
  Note that no passage estimates result from the fitting of enhanced
  efficiency models.  This is because bootstrapping does not occur, but also
  because estimation of passage is not the goal of the model fitting process.
  Use the regular function sequence; i.e., functions without the
  \code{".enh"} to estimation passage for one-year intervals of interest.
}
\description{
Fetch data on efficiency trials from an Access database.
}
\examples{
\dontrun{
#   ---- Fetch all Chinook salmon efficiency data on the American River 
#   ---- between Jan. 16, 2013 and June 8th, 2013.  
site <- 57000
taxon <- 161980
min.date <- "2013-01-01"
max.date <- "2013-06-01"
df <- F.get.release.data(site,taxon,min.date,max.date)
}
}
\seealso{
\code{run_passage.enh}, \code{get_release_data},
  \code{est_passage.enh}, \code{est_efficiency.enh}, \code{eff_model.enh}
}
\author{
WEST Inc.
}
