% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.lifestage.passage.assignLS3group.R
\name{F.lifestage.passage.assignLS3group}
\alias{F.lifestage.passage.assignLS3group}
\title{F.lifestage.passage.assignLS3group}
\usage{
F.lifestage.passage.assignLS3group(site, taxon, min.date, max.date, output.file,
  ci = TRUE, autoLS = TRUE)
}
\arguments{
\item{site}{<describe argument>}

\item{taxon}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}

\item{output.file}{<describe argument>}

\item{ci=TRUE}{<describe argument>}
}
\value{
<describe return value>
}
\description{
ANNUAL PRODUCTION ESTIMATES BY LIFE STAGE AND RUN ? TABULAR SUMMARY
   A table of passage estimates, with lifestages down the rows, and runs across the columns.

   Input:
   site = site ID of the place we want, trap locaton
   taxon = taxon number (from luTaxon) to retrieve



 3 life stage groups are fit and the program will decide if weight will be used or not
}
\details{
<other comments found in file>
NA
NA
NA
 3 life stage groups are fit and the program will decide if weight will be used or not
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

