% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.lifestage.passage.assignLSNoWeight.R
\name{F.lifestage.passage.assignLSNoWeight}
\alias{F.lifestage.passage.assignLSNoWeight}
\title{F.lifestage.passage.assignLSNoWeight}
\usage{
F.lifestage.passage.assignLSNoWeight(site, taxon, min.date, max.date,
  output.file, ci = TRUE, autoLS = TRUE)
}
\arguments{
\item{site}{<describe argument>}

\item{taxon}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}

\item{output.file}{<describe argument>}

\item{ci=TRUE}{<describe argument>}
}
\value{
<describe return value>
}
\description{
ANNUAL PRODUCTION ESTIMATES BY LIFE STAGE AND RUN ? TABULAR SUMMARY
   A table of passage estimates, with lifestages down the rows, and runs across the columns.

   Input:
   site = site ID of the place we want, trap locaton
   taxon = taxon number (from luTaxon) to retrieve



 allow the program to decide the number of life stage groups but weight will not be used to assign lifestage
}
\details{
<other comments found in file>
NA
NA
NA
 allow the program to decide the number of life stage groups but weight will not be used to assign lifestage
}
\examples{
\dontrun{
<insert examples>

}
}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

