% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.lifestage.passage.forkLength.R
\name{F.lifestage.passage.forkLength}
\alias{F.lifestage.passage.forkLength}
\title{F.lifestage.passage.forkLength}
\usage{
F.lifestage.passage.forkLength(site, taxon, min.date, max.date, by, output.file,
  ci = TRUE, autoLS = FALSE, reclassifyFL = TRUE)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{by}{A text string indicating the temporal unit over which daily
estimated catch is to be summarized.  Can be one of \code{day},
\code{week}, \code{month}, \code{year}.}

\item{output.file}{A text string indicating a prefix to append to all output.}

\item{ci}{A logical indicating if 95% bootstrapped confidence intervals
should be estimated along with passage estimates.}

\item{autoLS}{Default is FALSE, no analytical life stage assignment is done. If TRUE, the analytical life stage assignment is done, see details.}

\item{reclassifyFL}{A logical indicating if passage should be estimated via forklength-based class groups.}

\item{nLS}{Number of life stage groups to be estimated. Ignored if autoLS=FALSE, see details.}

\item{weightUse}{Boolean variable indicating if weight should be used for the analytical assignment, default is NULL. Ignored if autoLS=FALSE, see details.}
}
\value{

}
\description{
ANNUAL PRODUCTION ESTIMATES BY LIFE STAGE AND RUN ? TABULAR SUMMARY
   A table of passage estimates, with lifestages down the rows, and runs across the columns.

   Input:
   site = site ID of the place we want, trap locaton
   taxon = taxon number (from luTaxon) to retrieve



 2 life stage groups are fit and the program will decide if weight will be used or not
}
\details{

}
\examples{
\dontrun{
F.lifestage.passage.forkLength

}
}
\author{
WEST Inc.
}
\seealso{
\code{\link{related routine}}, \code{\link{related routine}}
}

