% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_model.r
\name{F.plot.catch.model}
\alias{F.plot.catch.model}
\title{F.plot.catch.model}
\usage{
F.plot.catch.model(df, file = NA)
}
\arguments{
\item{df}{<describe argument>}

\item{file=NA}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Plot the catch GAM model, showing observed and imputed
   values.

   input:
   df = data frame input to F.catch.model.  GAM fit an attribute of the df. 
   file = If file is not NA, save the graph to this file name.

   Output:
   A time series plot of catch by date.



 df <- masterCatch
  file <- plot.file
}
\details{
<other comments found in file>
   If file=NA, a pdf graphing device is assumed to be open already.
   Shut down all graphics devices
   ---- Open PNG device
 impute the imputed number of fish into unassd catch days.  awkward, b/c imputed catch
 based on plus-counted fish, but unassd based on measured fish (non-plus-counted).
df$unassdcatch <- ifelse(df$imputed.catch > 0,df$catch,df$unassdcatch)
 jason add 4/20/2015 -- plotting only measured values means that many days in df have NA.
 we need to pull those out before we plot--sometimes, a trap could have caught many fish,
 none of which were measured.  so, these traps end up not having anything to plot, even
 though the code thinks data for that plot is there.  this causes errors.
df <- df[ df$imputed.catch != 0 | !is.na(df$unassdcatch) ,]
 jason turn on -- have to be creative -- imputed and unassd catch generally on
 different scales, since imputed based on plus-counts, but unassigned on raw catches.
   This adds the smoothed interpolation model to the plot.  
     ind <- df$trapPositionID == traps[i]          # jason turns off
     lines( supsmu(df$batchDate[ind], df$catch[ind]), lwd=2, lty=1, col=my.colors[i] )
 jason 4/20/3015 -- use new var indOld?
catch.df.sites <- unique(df[,c('trapPositionID','TrapPosition')])       # jason add   # delete later?  added by github merge process.
 subsite.name <- attr(df, "subsites")
   Add title
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

