% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eff_model.r
\name{F.plot.eff.model}
\alias{F.plot.eff.model}
\title{F.plot.eff.model}
\usage{
F.plot.eff.model(fit, df, file)
}
\arguments{
\item{fit}{<describe argument>}

\item{df}{<describe argument>}

\item{file}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Plot trap efficiencies
   values.

   input:
   fit = for now, a data frame containing date and efficiency values.
   df = data frame input to F.eff.model
   file = file to write plot to

   Output:
   A time series plot of catch by date.

   df <- ans
   file <- plot.file
}
\details{
<other comments found in file>
if(length(dev.list()) > 1) windows()
 get subsite Names
 bring in subsite info on the data
 write out estimates.
   Draw points and lines
   indicators for season and such
 The following plots period between trial
 The following plots all time
   Draw lines
   Draw points
   Draw legend
   Add title
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

