% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lifestages.r
\name{F.plot.lifestages}
\alias{F.plot.lifestages}
\title{F.plot.lifestages}
\usage{
F.plot.lifestages(df, file.root, plot.pies = FALSE)
}
\arguments{
\item{df}{A data frame containing lifestages in the rows and runs across the
columns.}

\item{file.root}{A text string indicating a prefix to append to all output.}

\item{plot.pies=FALSE}{If true, a pie chart is produced instead of a bar
chart.}
}
\value{
A graphical png plotting a bar chart showing relative numbers of 
  estimated passage, with lifestages down the rows, and runs across the 
  columns.  Generally, all passage routines utilize the bar-chart option via
  paramter \code{plot.pies};  thus, pie charts are never returned.
}
\description{
Produce a plot of the passage associated with all distinct 
  combinations of lifestage and run.
}
\details{

}
\examples{
\dontrun{
# Plot passage estimates housed in a data frame
# entitled allCatch.  
F.plot.lifestages(df=allCatch,file.root="PassageResults",plot.pies=FALSE)
}
}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

