% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassifyLS.R
\name{F.reclassifyLS}
\alias{F.reclassifyLS}
\title{F.reclassifyLS}
\usage{
F.reclassifyLS(catch)
}
\arguments{
\item{catch}{A catch dataframe originating from a catch-query sequence.  See
function \code{F.get.catch.data}.  At the least, \code{catch} must contain
variables entitled \code{forkLength} and \code{lifeStage}.}
}
\value{
A dataframe with original variable \code{lifeStage} values containing
  "\code{Fry}," "\code{Parr}," and "\code{Smolt}" into the label values
  specified via the Global environment dataframe \code{forkLengthCutPoints}.
  See \code{GlobalVars}.
}
\description{
Reclassify lifeStage via forklength into groups specified
via Global Enviornment dataframe \code{forkLengthCutPoints}.
}
\details{

}
\examples{
 
\dontrun{
catchWithNewLifeStageCats <- reclassifyLS(catch)
}
}
\seealso{
\code{GlobalVars.R}, \code{F.get.catch.data}
}

