% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_summary.r
\name{F.release.summary}
\alias{F.release.summary}
\title{F.release.summary - Summarize releases}
\usage{
F.release.summary(site, taxon, run, min.date, max.date, output.file)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{run}{The biologist-assigned seasonal run.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{output.file}{The name of the file prefix under which output is to be 
saved.  Set to NA to plot to the Plot window.}
}
\value{
A \code{csv} file summarizing efficiency trials for the specified
  criteria.  A data frame containing the same information is also returned,
  but invisibly.
}
\description{
Summarize efficiency (catchability) trials, or releases, for a site, run, and taxon between
specified dates.
}
\details{

}
\examples{
\dontrun{

# requires an mdb.
    
}
}
\seealso{

}

