% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_passage.enh.r
\name{F.run.passage.enh}
\alias{F.run.passage.enh}
\title{F.run.passage.enh}
\usage{
F.run.passage.enh(site, taxon, min.date, max.date, by, output.file, ci = TRUE)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{by}{A text string indicating the temporal unit over which daily 
estimated catch is to be summarized.  Can be one of \code{day}, 
\code{week}, \code{month}, \code{year}.}

\item{output.file}{A text string indicating a prefix to append to all output.}

\item{ci}{A logical indicating if 95\% bootstrapped confidence intervals 
should be estimated along with passage estimates.}

\item{run}{The text seasonal identifier.  This is a one of \code{"Spring"},
\code{"Fall"}, \code{"Late Fall"}, or \code{"Winter"}.}
}
\value{
Results from running the enhanced efficiency model fitting process 
  include \code{csv}s of efficiency trials missing a covariate, for each
  \code{TrapPositionID} at a \code{subSiteID}.  This prevents these
  efficiency trials from inclusion in the model fitting process.
  
  Each \code{TrapPositionID} also results in a series of plots depicting the
  fitted temporal spline, at each point of the backwards-fitting process. 
  These could number many, depending on the number of covariates available
  for possible exlcusion.
  
  Each \code{trapPositionID} also outputs a \code{png} containing model
  fitting information, including plots of efficiency versus each considered
  covariate, along with plotted temporal trends of each covariate against
  time.  Additional plots include the final fitted temporal spline (along
  with an prediction "curve" derived from the available data), as well as a
  final "plot" depicting model summary statistics, obtained via the
  \code{summary} function against the logistic efficiency-trial model fits.
  
  Note that no passage estimates result from the fitting of enhanced
  efficiency models.  This is because bootstrapping does not occur, but also
  because estimation of passage is not the goal of the model fitting process.
  Use the regular function sequence; i.e., functions without the
  \code{".enh"} to estimation passage for one-year intervals of interest.
}
\description{
Estimate production by life stage and run for all days within a
  date range.
}
\section{Functions Used to Fit Enhanced Efficiency Models}{
 
  Five programs make up the specialized procedure for fitting enhanced
  efficiency models.  This means actually compiling the data of efficiency 
  trials obtained over several years, and then fitting a generalized additive
  model (GAM) to those data.  All five programs have suffixes of 
  \code{".enh"}, and originated with the program versions without the suffix.
  As such, they are very similar to the originals.
  
  The first, \code{run_passage.enh} corrals the fitting.  It is different 
  from \code{run_passage} in that all of the passage summary that is usually 
  created has been suppressed.  This is because there is no need to 
  bootstrap, once enhanced efficiency models have been obtained.
  
  The second, \code{get_release_data.enh} modifies the obtaining of release 
  data, so as to obtain astrological data and mean fork-length.  It is very 
  similar to its originator.
  
  The third, \code{est_passage.enh}, corrals the data from 
  \code{get_release_data.enh} for use in \code{est_efficiency.enh}.  It too 
  should be very similar to its originator.
  
  The fourth, \code{est_efficiency.enh}, ensures the calculation of weighted 
  averages for the three efficiency-trial covariates have to do with mean 
  fork-lengths, and percent of fishing performed at night, or while the moon 
  is up.  It also emulates closely its originator.
  
  Finally, the fifth, \code{eff_model.enh}, fits the enhanced efficiency 
  models.  It follows a backwards selection procedure, allowing for both 
  variable covariate selection, as well as variable temporal spline 
  complexity.  It creates graphical output, for each trap, so as to provide 
  further hypothesis generation.
}

\examples{
 
\dontrun{
#   ---- Estimate passage on the American for the Fall run.
site <- 6000
taxon <- 161980
min.date <- "2010-12-07"
max.date <- "2011-06-02"
by <- "day"
output.file <- "Feather"
ci <- TRUE
}
}
\seealso{
\code{run_passage.enh}, \code{get_release_data},
  \code{est_passage.enh}, \code{est_efficiency.enh}, \code{eff_model.enh}
}
