% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_passage.r
\name{F.run.passage}
\alias{F.run.passage}
\title{F.run.passage}
\usage{
F.run.passage(site, taxon, min.date, max.date, by, output.file, ci = TRUE)
}
\arguments{
\item{site}{<describe argument>}

\item{taxon}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}

\item{by}{<describe argument>}

\item{output.file}{<describe argument>}

\item{ci=TRUE}{<describe argument>}
}
\value{
<describe return value>
}
\description{
ANNUAL PRODUCTION ESTIMATES BY LIFE STAGE AND RUN ? TABULAR SUMMARY
   A table of passage estimates, with lifestages down the rows, and runs across the columns.

   Input:
   site = site ID of the place we want, trap locaton
   taxon = taxon number (from luTaxon) to retrieve


   ********
   Check that times are less than 1 year apart
}
\details{
<other comments found in file>
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

