% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_by_date.r
\name{F.size.by.date}
\alias{F.size.by.date}
\title{F.size.by.date}
\usage{
F.size.by.date(site, taxon, run, min.date, max.date, output.file)
}
\arguments{
\item{site}{<describe argument>}

\item{taxon}{<describe argument>}

\item{run}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}

\item{output.file}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Plot fork length of fish by date of catch ro a particular site and year. 

   Input:
   db = full path and name of the Access data base to retrieve data from
   site = site ID of the place we want, trap locaton 
   taxon = taxon number (from luTaxon) to retrieve
   run = run ID of fish we want to do estimates for. 

   Output:
   A graph, in "file" of the length by date data.
}
\details{
<other comments found in file>
   ---- Open PNG device
   *******
   Open ODBC channel and retrieve lifestage labels
   ********
   Retrieve basic data set, one line per fish or group of fish of same length.
 don't drop includeCatchID here.  this is a contrast with length_freq.r
 jason 3/25/2016 -- drop oldTrapPositionID here.  code assumes that var not here
 if catch.df has no data, it doesn't get batchDates added.  
catch.df <-  data.frame(catch.df,batchDate=integer(0))
  grab non-valid Catch
   Fetch run name
 no attributes to bring in -- do it now
   ********
   Now plot
   Define plotting variables
 x <- catch.df$visitTime
 n <- catch.df$n
 lstage <- catch.df$lifeStageID
   Drop record if any critical data is missing
   If we are talking salmon here, limit the lifestages to fry, parr, and smolt
     drop <- drop | (lstage > 8)
   Convert from lifestages the traps used to life stages that CAMP uses.  The conversion is in table rst.life.stages
 JASON OBSOLETE -- QUERY THAT GETS CATCH WORKS ON DESCRIPTORS INSTEAD OF IDS -- 1/26/2015
 u.l.s <- sort(unique(lstage))
 for( l.s in u.l.s ){
     camp.l.s <- rst.life.stage$lifeStageCAMPID[ rst.life.stage$lifeStageID == l.s ]
     lstage[ lstage == l.s ] <- camp.l.s
 }
   Rep the values for number of fish of that particular length
   Main plot
jity <- rnorm(sum(ind), sd=diff(range(y))/100)
95, 108, 122, 151, 152, 155
   Add quantile lines
 jason add.  this can sometimes be singular, apparently.  leads to problems in finding an inverse. 
     ans.qr <- data.frame( date=xpred, q.05=ypred[,1], q.95=ypred[,2] )
   Main title
   Legend
print(myleg)
print(myleg)
   ---- Close the graphics file
   ---- Write out the csv's
   ---- Send messages back to the interface
}
\examples{
\dontrun{
<insert examples>

}
}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

