% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fish_visit.r
\name{F.summarize.fish.visit}
\alias{F.summarize.fish.visit}
\title{F.summarize.fish.visit}
\usage{
F.summarize.fish.visit(catch, variable)
}
\arguments{
\item{catch}{A data frame containing one line per fish or fish group with 
identical lengths.  This data frame results from \code{F.get.catch.data}.}

\item{The}{type of fish for which summary is required.}
}
\value{
A data frame with one line per visit, with catch summarized. Summary
  statistics include N, mean, and standard deviation.
}
\description{
Summarize fish count and fork lengths over visits.
}
\details{
Results from the catch-query sequence summarize the number of fish 
  per \code{trapVisitID}, \code{lifeStage}, \code{FinalRun}, and 
  \code{forkLength};  these need to be summarized, depending on the type of 
  fish.
  
  Assigned fish are summarized via an index comprised of variables 
  \code{trapVisitID}, \code{FinalRun}, and \code{Unassd}.  Note that here, 
  assigned fish are assumed to have only their run assigned, and not 
  necessarily their life stage.  Assigned fish are separate from unassigned
  fish, meaning they were randomly selected, and thus form the basis of
  sampling proportions of run and life stage utilized via the plus-count
  algorithm.
  
  Unassigned fish are summarized via an index comprised of variables 
  \code{trapVisitID}, \code{FinalRun}, \code{lifeStage}, and \code{Unassd}.
  
  Inflated fish (all captured fish, including plus-counts) are summarized via
  an index comprised of \code{trapVisitID}, \code{FinalRun}, and 
  \code{lifeStage}.  Because inflated fish include plus-counted fish that 
  have been re-assigned to sampled-based counts of run and life stage, 
  indexing by variable \code{Unassd} is unnecessary.
  
  All other fish types for which summaries are required, namely those
  involved with the adjustment for half-cone operations, have an index of 
  \code{trapVisitID}, \code{FinalRun}, and \code{lifeStage}.
  
  Weighted fork lengths are recorded to the nearest millimeter.  Each is
  calculated over unique indices and weighted on the number of fish.
}
\examples{

}

