% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_passage.r
\name{F.summarize.passage}
\alias{F.summarize.passage}
\title{F.summarize.passage}
\usage{
F.summarize.passage(df, summarize.by)
}
\arguments{
\item{df}{<describe argument>}

\item{summarize.by}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Average passage over all traps at a site, then sum by summarize.by

   df must have the following columns: $batchDate, $imputed.catch
   df <- grand.df
   summarize.by <- sum.by
}
\details{
<other comments found in file>
   ---- First, average over traps
   ---- Now, summarize passage by time period.  
   Put catch and aux information into data frame
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

