% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_releases.r
\name{F.summarize.releases}
\alias{F.summarize.releases}
\title{F.summarize.releases}
\usage{
F.summarize.releases(release.df)
}
\arguments{
\item{release.df}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Summarize trap efficiency for every release.  This function
   produces a table of information that is used by other routines.

   Input:
   release.df = data frame read by F.get.release.data().  Contains
       information on releases, marks, etc.  Run season is a vector
       of dates for beginning and ending of run, stored as an attribute of the data frame.
}
\details{
<other comments found in file>
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

