% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSpline.r
\name{fitSpline}
\alias{fitSpline}
\title{fitSpline}
\usage{
fitSpline(covarString, df, eff.ind.inside, tmp.df, dist, max.df,
  eff.inside.dates)
}
\arguments{
\item{covarString}{A character string of length one housing all covariates to
be considered, with all covariates collapsed via \code{" + "}.}

\item{df}{The data frame for a specific \code{TrapPositionID} containing 
efficiency-trial information and covariates, if available, at the time of 
fitting enhanced efficiency trials in \code{eff_model.r} (or 
\code{F.efficiency.model }).}

\item{eff.ind.inside}{A vector of length equal to the number of rows in data 
frame \code{df} set equal to \code{TRUE} when the \code{df} 
\code{batchDate} is inside the data range specified by 
\code{eff.inside.dates}.}

\item{tmp.df}{The reduced data frame originating from \code{df} containing 
only efficiency-trial dates; i.e., those with non-\code{NA} 
\code{nReleased} values.}

\item{dist}{The distributional family used in calls to function \code{glm}. 
Almost always set to \code{"binomial"}.}

\item{max.df}{The highest number of degrees of freedom to be used in the 
fitting of temporal-dimension splines.  Set by \code{max.df.spline}, which 
is \code{4} in function \code{eff.models}.}

\item{eff.inside.dates}{A temporal \code{POSIX} vector of length two 
indicating the first and last \code{batchDate} efficiency trials.}
}
\value{
A list containing several different objects.  

\describe{
  \item{fit}{The \code{glm}-type object resulting from the fit.}
  \item{fit.AIC}{The AIC of \code{fit}.}
  \item{bspl}{The matrix of ALL \code{batchDate}s within \code{eff.inside.dates} of the final evaluated model. }
  \item{tmp.bs}{The matrix of REDUCED \code{batchDates}s from \code{bspl} tied to efficiency-trial dates.}
  \item{cur.df}{The final number of degrees-of-freedom used in the final temporal spline fit.}
  \item{disp}{}
  \item{s.beg}{The first date housed within the 1960-spline paradigm variable \code{batchDate2}.}
  \item{s.end}{The last date housed within the 1960-spline paradigm variable \code{batchDate2}.}
  \item{cur.bspl}{The matrix of ALL \code{batchDate}s within \code{eff.inside.dates} of \code{fit} model. }
}
}
\description{
Fit a temporal spline with a given set of covariates; i.e., fit
  a generalized additive model (GAM).
}
\details{
Note that returned object \code{bspl} is not the same as 
  \code{cur.bspl}.  In order to stop, the loop that evaluates temporal 
  splines must first evaluate the next model.  In the case that the next 
  model fails to be better than the current model, the current model is the 
  winner. Object \code{cur.bspl} is the basis matrix associated with the 
  winning current model, while \code{bspl} is the basis maxtri associated 
  with that (non-winning) next model.
  
  The overdispersion parameter is fit in the traditional way, i.e., via
  Pearson residuals.
}
\examples{
 
\dontrun{
fit <- fitSpline(covarString,
                 df,
                 eff.ind.inside,
                 tmp.df,
                 dist,
                 max.df,
                 eff.inside.dates)
}
}
\seealso{

}
