% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetAlk.camp}
\alias{GetAlk.camp}
\title{Extrae ALK para una especie de los ficheros del camp}
\usage{
GetAlk.camp(gr, esp, camp, dns = "Pnew", plus = 8, AltAlk = NA)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camp}{Campaña de la que se extraen los datos un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{AltAlk}{ALK alternativa tomada de un fichero edadXYY.dbf}
}
\description{
Función de acceso a datos: recupera la clave talla edad de una especie y campaña determinadas
}
\examples{
GetAlk.camp("1"," 43","N93","Cant",8)
}
\seealso{
grafedtal.camp {\link{grafedtal.camp}}
}

