% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatchDataWeight.R
\name{getCatchDataWeight}
\alias{getCatchDataWeight}
\title{getCatchDataWeight}
\usage{
getCatchDataWeight(taxon, site, min.date, max.date)
}
\arguments{
\item{taxon}{<describe argument>}

\item{site}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Jared Studyvin
 14 Jan 2016
 get the data for the life stage assignment
}
\details{
<other comments found in file>
 ---- end building ----
 ---- finally, the new query ----
 ---- end new query -----
 return the data
}
\examples{
\dontrun{
<insert examples>

}
}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

