% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICESarea.R
\name{getICESarea}
\alias{getICESarea}
\title{Includes ICESrect in an HH file with lat and long data}
\usage{
getICESarea(camp, dns, incl2 = T, incl0 = F)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)gr Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{incl2}{Si T se incluyen los lances extra no incluidos para las abundancias o biomasas estratificadas}

\item{incl0}{Si T se incluyen los lances nulos}
}
\value{
A data.table with haul variables, ICESrect and ICESarea
}
\description{
Function to complete HH with ICESrect and area
}
\examples{
# getICESrect(icesDatras::getHHdata("SP-NORTH",2015,4),latname="haullat",longname="haullong")
}
