% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeProp.r
\name{getTimeProp}
\alias{getTimeProp}
\title{getTimeProp}
\usage{
getTimeProp(df, rise, set, traps, tmp, metric)
}
\arguments{
\item{df}{A data frame of skinny astrological data describing one of either 
sun or moon rise or set datetimes.}

\item{rise}{A text string describing the rise time from data frame 
\code{dates}.  Either sun or moon rise times.}

\item{set}{A text string describing the set time from data frame 
\code{dates}.  Either sun or moon set times.}

\item{traps}{A vector of \code{trapPositionID}s over which rise and set data 
are required.}

\item{tmp}{A data frame consisting of \code{StartTime}s and \code{EndTime}s
for fishing for the current data run.}

\item{metric}{A text vector of size one.  Typically intended to be one of 
either \code{"sun"} or \code{"moon"}.}
}
\value{
A data frame with the proportion of time spent fishing during the day
  (when the sun was up), or the same when the moon was up.
}
\description{
Shuffle fishing \code{StartTime}s and \code{EndTime}s with rise 
  and set times associated with one of either sun or moon astrological data.
}
\details{
Denominators are defined for each unique trapping instance via
  variable \code{trapVisitID}, and are defined as the length of time in
  fishing, in minutes.  Numerators are the length of time the sun was up, in
  minutes, given the fishing interval, or the corresponding metric for moon
  times.
  
  Note that this variable, usually labeled as \code{"SampleMinutes"} in
  results deriving from Connie's SQL queries, varies slightly from Connie's
  calculated value.  Generally, however, the number of deviations is small,
  given a large sample of unique \code{trapVisitIDs}; i.e., fishing
  instances.  In the overwheling majority of cases, the results between 
  Connie's value and here exactly agree.  
  
  The temporal \code{dates} table is defined via the earliest and 
  latest efficiency trial date for the run in question.  To this, an 
  additional month is appended at each end.
}
\examples{
\dontrun{
dfNew <- getTimeProp(df,rise,set,traps,tmp,metric)
}
}
\seealso{
\code{makeSkinnyTimes}
}
\author{
WEST Inc.
}
