% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTogetherCovarData.r
\name{getTogetherCovarData}
\alias{getTogetherCovarData}
\title{getTogetherCovarData}
\usage{
getTogetherCovarData(obs.eff.df, min.date, max.date, traps, useEnhEff)
}
\arguments{
\item{obs.eff.df}{A data frame with at least variables \code{batchDate} and
\code{efficiency}, where \code{efficiency} is \code{NA} for all days 
requiring an estimate.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{traps}{A set of traps for which efficiency data are available.}

\item{useEnhEff}{A logical indicating if function \code{getTogetherCovarData}
should use query dates based on provided \code{min.date} and
\code{max.date} (\code{TRUE}), or efficency time frames (\code{FALSE}).}
}
\value{
Several items, all housed within a list.  The main piece of output is 
data frame \code{obs.eff.df}, which contains the same number of rows as the 
\code{obs.eff.df} submitted to the function, but with several additional 
columns, due to appended covariates.  

\describe{
  \item{obs.eff.df}{Same as submitted to the function, but with covariate columns added.}
  \item{dbDisc}{Available CAMP discharge data.}
  \item{dbDpcm}{Available CAMP water-depth data.}
  \item{dbATpF}{Available CAMP air-temperature data.}
  \item{dbTurb}{Available CAMP turbidity data.}
  \item{dbWVel}{Available CAMP water-velocity data.}
  \item{dbWTpC}{Available CAMP water-temperature data.}
  \item{dbLite}{Available CAMP light-penetration discharge data.}
  \item{dbFlPG}{Available Environmental Covariate Database discharge data.}
  \item{dbTpPG}{Available Environemtnal Covariate Database water temperature data.}
}
}
\description{
Put together available covariate data from both the external
  Environmental Covariate Database, as well as CAMP-collected environmental
  variables.
}
\details{
Function \code{getTogetherCovarData} appends covariate information 
  for each unique \code{batchDate} and \code{TrapPositionID} combination 
  avaiable in data frame \code{obs.eff.df}.
  
  Environmental covariate data are queried from the WEST-supported database
  created for housing USGS and CDEC data at sites important to the CAMP
  program.
  
  Other covariates derive directly from data stored within a particular
  river's CAMP mdb database.  Prior to use in efficiency models, data are
  first converted to standardized units via the \code{QryCleanEnvCov} SQL
  query sequence in function \code{getCAMPEnvCov}, which is called once 
  for each covariate.    
  
  Following querying, all covariates are then fit with a smoothing spline via
  function \code{estCovar}, with penalization parameter \eqn{\lambda}
  selected via cross-validation. Smoothing occurs to ensure easy fitting of
  efficiency models, whose temporal resolution is that of a \code{batchDate}.
}
\examples{
 
\dontrun{
ans <- getTogetherCovarData(obs.eff.df,
                            min.date,
                            max.date,
                            traps,
                            useEnhEff)
}
}
\references{
Hastie, T., Tibshirani, R., and Friedman, J. 2009.  The Elements
  of Statistical Learning.  2nd Edition.  Springer, New York, New York.
}
\seealso{
\code{getCAMPEnvCov}, \code{estCovar}
}
