% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradec.R
\name{gradec}
\alias{gradec}
\title{Transforma grados y minutos decimales a grados decimales}
\usage{
gradec(x)
}
\arguments{
\item{x}{grados en formato decimal}
}
\description{
Transforma grados y minutos decimales a grados decimales, avisa si el dato es incorrecto, es decir minutos más de 60
}
\examples{
camptoyear(Nsh)
}
