% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradms.R
\name{gradms}
\alias{gradms}
\title{Transforma grados, minutos y segundos a grados decimales}
\usage{
gradms(x)
}
\arguments{
\item{x}{grados, minuto y segundos en formato "GGG:MM:SS", vector de caracter}
}
\description{
Transforma grados y minutos decimales a grados decimales, avisa si el dato es incorrecto, es decir minutos o segundos más de 60
}
\examples{
gradms("43:14:25")
}
\seealso{
Other Conversion unidades: 
\code{\link{HdectoH}()},
\code{\link{decgrad}()},
\code{\link{gradec}()}
}
\concept{Conversion unidades}
