% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grafhistbox.comp}
\alias{grafhistbox.comp}
\title{Gráficas grafhistbox combinadas biomasa y peso}
\usage{
grafhistbox.comp(gr, esp, camps, dns = "Pnew", cor.time = T, kg = T,
  ci.lev = 0.8, idi = "l", SE = T, es = T, sector = NA, Nas = F,
  excl.sect = NA, ymax = c(NA, NA), tline = F, years = F, ti = T,
  mar = NA)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Codigo de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{campañas de la serie de datos a representar en el gráfico de abundancias Demersales Nsh, Porcupine Psh, Arsa primavera Ash y Arsa otoño 2sh}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{kg}{Si T el gráfico está en kgs, si F en gramos}

\item{ci.lev}{El intervalo de confianza a representar}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{SE}{Si T dibuja las cajas representando el error estándar, si F no aparecen la cajas}

\item{es}{Si T ejes y unidades en español, si F en inglés}

\item{sector}{Alternativa a excl.sect para un sólo sector. Si especificado como carácter solo toma el sector elegido}

\item{Nas}{Si F no representa las cajas en los sectores/estratos en los que algún sector o estrato tiene un sólo lance. Si T utiliza el valor del estrato y evita los NAs}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{ymax}{Valor máximo del eje de las y, tiene doble valor para regular biomasa y número}

\item{tline}{Si T dibuja una línea de tendencia a traves de un glm con los datos de abundancia. Gráficos evaluación MSFD.}

\item{years}{Si T saca los años como nombre de campaña en el eje de las equis en vez del nombre de campaña}

\item{ti}{Permíte sólo el nombre de la especie que aparece en el gráfico superior y sale siempre en cursiva.}

\item{mar}{Si se quiere dejar un margen ya establecido hacerlo igual a F}
}
\value{
Crea una gráfica doble de evolución de las abundancias en biomasa y número.
}
\description{
Ver documentación grafhistbox
}
\examples{
grafhistbox(1,45,Nsh[7:27],"Cant",es=F,years=T,tline=T,ti=T,subtit=T)
}
\seealso{
grafhistbox {\link{grafhistbox}}

grafhistbox.comp {\link{grafhistbox.comp}}
}

