% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grafhistbox}
\alias{grafhistbox}
\title{Evolución en biomasa o abundancia de la especie en la serie histórica}
\usage{
grafhistbox(gr, esp, camps, dns = "Pnew", ind = "p", cor.time = T,
  kg = T, ci.lev = 0.8, idi = "l", SE = T, es = T, excl.sect = NA,
  sector = NA, ti = T, Nas = F, ymax = NA, mar = NA, tline = F,
  years = F, subtit = F, cex.main = 1.1)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Codigo de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{campañas de la serie de datos a representar en el gráfico de abundancias Demersales Nsh, Porcupine Psh, Arsa primavera Ash y Arsa otoño 2sh}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{ind}{Parámetro a representar saca los datos en "p"eso o "n"úmero}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{kg}{Si T el gráfico está en kgs, si F en gramos}

\item{ci.lev}{El intervalo de confianza a representar}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{SE}{Si T dibuja las cajas representando el error estándar, si F no aparecen la cajas}

\item{es}{Si T ejes y unidades en español, si F en inglés}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{sector}{Alternativa a excl.sect para un sólo sector. Si especificado como carácter solo toma el sector elegido}

\item{ti}{Si T muestra el título con el nombre de la especie (y el sector si se elige sect o excl.sect si falso no muestra nada. También se puede dar un valor y aparece el texto incluido o dar un list(label=(título), font, cex)}

\item{Nas}{Si F no representa las cajas en los sectores/estratos en los que algún sector o estrato tiene un sólo lance. Si T utiliza el valor del estrato y evita los NAs}

\item{ymax}{Valor máximo del eje y}

\item{mar}{Si se quiere dejar un margen ya establecido hacerlo igual a F}

\item{tline}{Si T dibuja una línea de tendencia a traves de un glm con los datos de abundancia. gráficos evaluación MSFD.}

\item{years}{Si T saca los años como nombre de campaña en el eje de las equis en vez de de campañas}

\item{subtit}{Añade un subtítulo con el valor que se le ponga si no F}

\item{cex.main}{Sirve para modificar los tamaños de letra del gráfico, funciona con ejes y títulos. Por defecto en 1.1}

\item{mar}{Sirve para definir los margenes de la figura alrededor del gráfico mismo. Por defecto es NA y deja los definicos por defecto por R. Si se definen tiene que ser un vector con cuatro valores numéricos en líneas: c(bottom,left,upper,right)}
}
\value{
Crea una gráfica de evolución de las abundancias en biomasa o número y devuelve en consola un data.frame con columnas: avg,SE (error estándar),camp
}
\description{
Gráfico de cajas con la evolución en biomasa o abundancia en la serie histórica
}
\examples{
grafhistbox(1,45,Nsh[7:27],"Cant",es=F,years=T,tline=T,ti=T,subtit=T)
}
\seealso{
grafhistbox.comp {\link{grafhistbox.comp}}
}

