% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrafMarksGPS_2.R
\name{GrafMarksGPS_2}
\alias{GrafMarksGPS_2}
\title{Gráficos de dragas o CTDs a partir de las marcas del PescaWin con leyenda móvil}
\usage{
GrafMarksGPS_2(camp, dns = "Cant", event = "Draga", label = "Dredges",
  add = FALSE, ti = TRUE, years = TRUE, pch = NA, ptbg = NA,
  xlims = NA, ylims = NA, cuadr = TRUE, es = T, places = TRUE,
  bw = FALSE)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa"}

\item{event}{Elige el evento que se representa, dragas o CTDs}

\item{add}{Si T añade los puntos al gráfico previo, si F saca el gráfico nuevo}

\item{ti}{Si T incluye un título con el nombre de la campaña}
}
\description{
Gráfica todas las dragas o CTDs de una campaña siempre que se hayan marcado con el PescaWin durante la misma
Requiere que los ficheros de marcas del PescaWin estén en los directorios "C:/GPS/mrk/porcupine/" para Porcupine o "C:/GPS/mrk/Norte/" para Demersales. Muestra el gráfico y queda pendiente de colocar la leyenda
}
\examples{
GrafMarksGPS_2("N14","Cant")
}
\seealso{
Other mapas: \code{\link{GrafMarksGPS}},
  \code{\link{MapEcol.camp}}, \code{\link{MapGaltal}},
  \code{\link{MapLansGPS}}, \code{\link{NepClos}},
  \code{\link{armap.camp}}, \code{\link{armap.tot}},
  \code{\link{map.check}}, \code{\link{maparea}},
  \code{\link{maphistage}}, \code{\link{maphistal}},
  \code{\link{maphist}}

Other PescaWin: \code{\link{GrafMarksGPS}},
  \code{\link{MapLansGPS}}, \code{\link{pasa.lan}}
}
