% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidrotodec.R
\name{hidrotodec}
\alias{hidrotodec}
\title{Transforma fichero hidroXXX.dbf a datos decimales}
\usage{
hidrotodec(camp, dns, year = T)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa", combinados con "dnsred" busca los datos en el servidor de Santander si se han creado las odbcs}

\item{year}{si T incluye una columna con el año al final de los datos}
}
\value{
Devuelve un data.frame con datos de cada lance, las variables dependen de la selección de hidro y redux. En cualquier caso incluye variables weight.time con el factor de calibración para lances con menos tiempo del estándar y arsect: el área del sector al que corresponde el lance dentro del muestreo
}
\description{
Añade variables lat y long a los datos en el fichero hidroXXX.dbf (no toca el fichero dbf)
}
\examples{
camptoyear(Nsh)
}
\seealso{
Other datos_especies: 
\code{\link{AbrvEsp}()},
\code{\link{BuscaAphia}()},
\code{\link{buscacod}()},
\code{\link{buscaesp.datras}()},
\code{\link{buscaesp}()},
\code{\link{camptoyear}()},
\code{\link{talpes.camp}()}
}
\concept{datos_especies}
