% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histboxplot.comp.R
\name{histboxplot.comp}
\alias{histboxplot.comp}
\title{Gráficos de boxplot para la serie histórica incluyendo lances especiales o no y con rangosx de profundidad}
\usage{
histboxplot.comp(
  gr,
  esp,
  camps,
  dns = "Porc",
  cor.time = TRUE,
  incl2 = TRUE,
  es = T,
  bw = TRUE,
  ti = TRUE,
  sub = NULL,
  idi = "l",
  ceros = TRUE,
  cex.leg = 1,
  years = TRUE,
  profrange = NA
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 desechos y otros, 9 escoge todos los orgánicos pero excluye desechos}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{Campaña a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o  "Pnew", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{incl2}{Si F no tiene en cuenta los lances especiales, si T si los tiene en cuenta, pero da problemas por que no puede calcular las abundancias estratificadas}

\item{es}{Si T rotulos gráfico en español, si F en inglés}

\item{bw}{Gráfico en blanco en negro si T o en color si F}

\item{ti}{Añade el nombre de la especie en latín sin T, si F no añade titulo}

\item{sub}{Añade un subtítulo debajo del gráfico, sin texto por defecto.}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{ceros}{por defecto incluye los valores de 0 al calcular los rangos y medianas, si T los quita, reflejarlo en el pie del gráfico}

\item{years}{Si T saca los años como nombre de campaña en los paneles lattice de campañas}

\item{profrange}{Si c(profmin,profmax) filtra por ese rango de profundidad}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{ind}{Parámetro a representar saca los datos en "p"eso o "n"úmero}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}

\item{escmult}{Varía la relación de tamaño de los puntos con la leyenda y el máximo en los datos}

\item{cexleg}{Varía el tamaño de letra de los ejes y del número de la leyenda}
}
\value{
Si out.dat=TRUE devuelve un data.frame con columnas: lan,lat,long,prof,peso.gr,numero (de individuos entre tmin y tmax),camp, si out.dat=F saca el gráfico en pantalla o como objeto para combinar con otros gráficos con print.trellis
}
\description{
Crea mapas con la distribución en biomasa o numero para distintas zonas: Porcupine (dns="Pnew"), el Cantábrico (dns=Cant), Cádiz= (dns=Arsa), y el Mediterráneo (dns=Medi)
}
\examples{
histboxplot(1,50,Nsh[7:27],"Cant",years=TRUE)
histboxplot(1,50,Nsh[7:27],"Cant",years=TRUE,ind="n")
}
\seealso{
Other abunds: 
\code{\link{grafhistbox.comp}()},
\code{\link{grafhistbox}()},
\code{\link{grafhisttalbox}()},
\code{\link{histboxplot}()}
}
\concept{abunds}
