% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_campR.R
\name{install_campR}
\alias{install_campR}
\title{install_campR}
\usage{
install_campR(lib = "../Rlibrary", pkg = "campR")
}
\value{
The result of \code{devtools::install}.
}
\description{
A utility function to document, build, 
install, and attach the CAMP_R package.  This is for 
development work in R. Normal users will never call this.
}
\details{
All paths are hard coded. This version 
installs to the "win-library" library.
}
\examples{
install_campR()

}
