% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lan0.camp}
\alias{lan0.camp}
\title{Lances nulos en una campaña}
\usage{
lan0.camp(camp, dns)
}
\arguments{
\item{camp}{Campaña a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}
}
\description{
Busca y devuelve los lances nulos de una campaña
}

