% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListCamps_b.R
\name{ListCamps_b}
\alias{ListCamps_b}
\title{Listado de campañas presentes en dns con ficheros que contienen cada una.}
\usage{
ListCamps_b(dns, tablas = TRUE)
}
\arguments{
\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{tablas}{Si T muestra al final de los resultados una lista con todos los ficheros de definición (CAMP con su identificador de campaña), fauna y tallas presentes en el directorio.}
}
\value{
Devuelve el listado de ficheros dbf campañas que hay en el directorio; campXXX.dbf,faunaXXX.dbf, ntallXXX.dbf, edadXXX.dbf, hidroxxx.dbf
}
\description{
Función de acceso a datos:
A partir de ficheros de camp.dbf,fauna.dbf,ntall.dbf da un listado de campañas presentes en el directorio con sus ficheros
}
\examples{
ListCamps_b(dns="Porc",tablas=T)
}
\seealso{
Other Series datos: 
\code{\link{As1}},
\code{\link{As2}},
\code{\link{ListCamps}()},
\code{\link{Msh}},
\code{\link{Nsh}},
\code{\link{PresenciaEsp.camp}()},
\code{\link{Psh}}
}
\concept{Series datos}
