% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ListFauna.camp}
\alias{ListFauna.camp}
\title{Capturas medias por lance de cada especie de gr (grupo) en una campaña}
\usage{
ListFauna.camp(gr = "1", camp, dns, excl.sect = NA, incl2 = F)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{camp}{Campaña de la que se extrae el listado de especies capturadas: un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa", Medits "Medi"}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos, NA no excluye ninguno}

\item{incl2}{Si T incluye los lances especiales "2"}
}
\value{
Devuelve un data.frame con las capturas medias por lance de cada especie capturada del grupo gr. Columnas: gr,esp,especie,peso(kg),número,nlan (nº de lances en que ha aparecido esp)
}
\description{
Muestra un listado de la biomasa y número medio de todas las especies capturadas del grupo gr
}

