% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListFauna.camps.r
\name{ListFauna.camps}
\alias{ListFauna.camps}
\title{Capturas medias por lance de cada especie de gr (grupo) de varias campañas}
\usage{
ListFauna.camps(gr = "1", camps, dns, excl.sect = NA, incl2 = FALSE)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros.}

\item{camps}{campañas de las que se extrae el listado de especies capturadas: un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa", Medits "Medi"}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{incl2}{Si T incluye los lances especiales "2"}
}
\value{
Devuelve un data.frame con las capturas medias por lance de cada especie capturada del grupo gr en cada campaña. Columnas: camp,gr,esp,especie,peso(kg),número,nlan (nº de lances en que ha aparecido esp)
}
\description{
Ver documentación ListFauna.camp
}
\seealso{
{\link{ListFauna.camp}}
}
