% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListFauna.groups.R
\name{ListFauna.groups}
\alias{ListFauna.groups}
\title{Capturas medias por lance de cada especie de gr (grupo) de varias campañas}
\usage{
ListFauna.groups(
  camp,
  dns,
  excl.sect = NA,
  verbose = TRUE,
  cor.time = TRUE,
  kg = TRUE
)
}
\arguments{
\item{camp}{campaña de la que se extrae el listado de especies capturadas: un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa", Medits "Medi"}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{verbose}{si T imprime resultados en pantalla si F no.....}
}
\value{
Devuelve un data.frame con las capturas medias por lance de cada especie capturada del grupo gr en cada campaña. Columnas: camp,gr,esp,especie,peso(kg),número,nlan (nº de lances en que ha aparecido esp) LOs valores NaN corresponden a especies que sólo aparecen en lances especiales
}
\description{
Ver documentación ListFauna.camp
}
\details{
Por defecto extrae los datos de todos los grupos taxonómicos menos el 6 (deshechos y no orgánicos), sólo se puede utilizar para los lances estándares y los resultados responden a las abundancias estandarizadas, sólo incluyen los lances estándare, ni extras  ni fuera de estatos áreas de la campaña. Para incluir estos lances utilizar las funciones de ListFaunaTals
}
\examples{
ListFauna.groups(Nsh[length(Nsh)],"Cant",excl.sect=NA,incl2=TRUE)
}
\seealso{
Other ListadosFauna: 
\code{\link{Fauna.camp}()},
\code{\link{ListFauna.camp}()},
\code{\link{ListFauna.lans}()},
\code{\link{ListFauna.lan}()},
\code{\link{ListFaunaTals.camps}()},
\code{\link{ListFaunaTals.camp}()},
\code{\link{ListFaunaTals.groups}()}
}
\concept{ListadosFauna}
