% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListFauna.lans.R
\name{ListFauna.lans}
\alias{ListFauna.lans}
\title{Capturas de un todos los lances en una campaña}
\usage{
ListFauna.lans(camp, dns = "Porc", gr = NA, excl.sect = NA,
  incl2 = FALSE)
}
\arguments{
\item{camp}{Campaña de la que se extrae los datos del lance (lan) un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa"}

\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 basura. Si NA, muestra todos los grupos}

\item{excl.sect}{Sirve para excluir uno o más sectores o estratos, ponerlos como un vector por separado, i.e. c("A","B")}

\item{incl2}{Si TRUE se incuyen los lances especiales}

\item{lan}{Lance del que se extrae la faunística}
}
\value{
Devuelve un data.frame con columnas grupo,esp, lance,especie,peso,numero
}
\description{
Muestra un listado de las especies capturadas en peso y número en un lance concreto. Se pueden seleccionar grupos de especies
}
\examples{
ListFauna.lans("12C","Cant",gr=c(1,2))
}
\seealso{
{\link{ListFauna.camps}}, {\link{ListFaunaTals.camps}}
}
