% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListFaunaTals.camp.r
\name{ListFaunaTals.camp}
\alias{ListFaunaTals.camp}
\title{Capturas y tallas mínima y máxima de grupos de especies para todos los lances de una campaña \strong{permite incluir lances especiales}}
\usage{
ListFaunaTals.camp(
  gr = "1",
  camp,
  dns,
  cor.time = TRUE,
  excl.sect = NA,
  profrange = NA,
  incl2 = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados. 6 desechos y no orgánicos no se puede incluir.}

\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{profrange}{Si c(profmin,profmax) filtra por ese rango de profundidad, por defecto NA no filtra por profunidades, debe ser o NA o un rango con dos profundidades}

\item{incl2}{Si T incluye los lances especiales "2"}
}
\description{
Muestras las abundancias totales \strong{sin estratificar}, en número y peso y las tallas mínima y máxima de un grupo taxonómico de especies para todos los lances de una campaña. Hay que tener en cuenta que en numerosas especies de invertebrados no se toman datos de tallas, sobre todo históricamente, por lo que para estos taxones Lmin y Lmax resultan NA
}
\examples{
ListFaunaTals.camp(gr="1",camp="N90",dns="Cant",excl.sect=NA)
ListFaunaTals.camp(gr="1",camp="N90",dns="Cant",excl.sect=NA,profrange=c(400,500))
}
\seealso{
Other ListadosFauna: 
\code{\link{Fauna.camp}()},
\code{\link{ListFauna.camps}()},
\code{\link{ListFauna.camp}()},
\code{\link{ListFauna.groups}()},
\code{\link{ListFauna.lans}()},
\code{\link{ListFauna.lan}()},
\code{\link{ListFaunaTals.camps}()},
\code{\link{ListFaunaTals.groups}()}
}
\concept{ListadosFauna}
