% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ListFaunaTals.camps}
\alias{ListFaunaTals.camps}
\title{Capturas y tallas mínima y máxima de grupos de especies para todos los lances de varias campaña}
\usage{
ListFaunaTals.camps(gr = "1", camps, dns, excl.sect = NA, incl2 = T)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{camps}{campañas (años) a representar en el mapa: Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{incl2}{Si T incluye los lances especiales "2"}
}
\description{
Ver documentación ListFaunaTals.camp
}
\seealso{
ListFaunaTals.camp {\link{ListFaunaTals.camp}}
}

