% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logabage.camp}
\alias{logabage.camp}
\title{Crea gráficos de descenso de abundancias con la edad}
\usage{
logabage.camp(gr, esp, camps, dns = "Pnew", plus = 8, cor.time = T,
  clms = 2, layout = NA)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camps}{Serie historica de campañas de la que se extraen los datos. Todas las campañas han de tener ALKs para las especie en cuestión}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{clms}{Número de columnas del gráfico}

\item{layout}{parámetro con dos valores, filas y columnas del gráfico de lattice}
}
\value{
Saca gráficos de descenso de la abundancia a lo largo de la edad con la mortalidad total
}
\description{
gráficos en lattice (uno por cohorte en la serie) en escala logarítimica para el seguimiento de la mortalidad por cohortes
}
\examples{
logabage.camp("1"," 43",Nsh,"Cant",8,0)
}
\seealso{
logabage2.camp {\link{logabage2.camp.r}}  edadstr.camp {\link{edadstr.camp}} bubbage.camp {\link{bubbage.camp.r}}
}

