% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logabage2.camp}
\alias{logabage2.camp}
\title{Crea gráficos de descenso de abundancias con la edad}
\usage{
logabage2.camp(gr, esp, camps, dns = "Pnew", plus = 8, cor.time = T,
  ti = F, layout = NA)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camps}{Serie historica de campañas de la que se extraen los datos. Todas las campañas han de tener ALKs para las especie en cuestión}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{ti}{Si T muestra el título con el nombre de la especie (y el sector si se elige sect o excl.sect si falso no muestra nada. También se puede dar un valor y aparece el texto incluido o dar un list(label=(título), font, cex)}

\item{layout}{parámetro con dos valores, filas y columnas del gráfico de lattice, si NA lattice rellena la pantalla para optimizar la vista l*c}
}
\value{
Saca gráficos de descenso de la abundancia a lo largo de la edad con la mortalidad total
}
\description{
Tipico gráfico en escala logarítimica para el seguimiento de la mortalidad por cohortes Todas las cohortes en un mísmo gráfico
}
\examples{
logabage2.camp("1"," 43",Nsh,"Cant",8)
}
\seealso{
logabage.camp {\link{logabage.camp.r}} edadstr.camp {\link{edadstr.camp}} bubbage.camp {\link{bubbage.camp}} logabage.camp {\link{logabage.camp.r}}
}

