% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFake_release.df.R
\name{makeFake_release.df}
\alias{makeFake_release.df}
\title{makeFake_release.df}
\usage{
makeFake_release.df(site, min.date, max.date, visit.df)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as
\code{min.date}.}

\item{visit.df}{A data frame resulting from the catch SQL sequence.}
}
\value{
A data frame with the number of rows equal to the number of days
  between \code{min.date} and \code{max.date}, inclusive.
}
\description{
Create a fake \code{release.df} data frame for use when no
  trials exist within the time period specified via \code{min.date} and
  \code{max.date}.
}
\details{
Days produced depend on the \code{seq.POSIXt} function with
  \code{by="1 DSTday"}.

  The "fake efficiency trials" produced by this function insert one released
  fish and zero captured fish, per unique \code{trapPositionID} found within
  the \code{visit.df} data frame.  These numbers are used to minimize the
  impact, if any, the construction of these fake data cause.
}
\seealso{
\code{F.get.release.data}
}
