% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSkinnyTimes.r
\name{makeSkinnyTimes}
\alias{makeSkinnyTimes}
\title{makeSkinnyTimes}
\usage{
makeSkinnyTimes(rise, set, dates)
}
\arguments{
\item{rise}{A text string describing the rise time from data frame
\code{dates}.  Either sun or moon rise times.}

\item{set}{A text string describing the set time from data frame
\code{dates}.  Either sun or moon set times.}

\item{dates}{A data frame copy of the table \code{Dates} from an Access CAMP
database.}
}
\value{
A skinny data frame with one column housing datetimes, another to
  identify the "Event," i.e., a rise or set, and a third with all-\code{NA} 
  \code{trapVisitID}s.
}
\description{
Convert one of either the sun or moon wide rise and set times 
  from table \code{Dates} from an Access CAMP database into a long skinny
  with only one time variable.
}
\details{
Function \code{makeSkinnyTimes} simply converts datetime data in a
  wide format housed in two variables (a "rise" and "set") into one data
  frame with all datetime data in one column. This function prepares data
  frames for use with function \code{getTimeProp}, the true workhorse
  function.
  
  Note that the temporal \code{dates} table is defined via the earliest and
  latest efficiency trial date for the run in question.  To this, an
  additional month us appended at each end.
}
\examples{
\dontrun{
rise <- "sunRise"
set <- "sunSet"
dates <- tblDates

sun <- makeSkinnyTimes(rise,set,dates)
}
}
\seealso{
\code{getTimeProp}
}
\author{
WEST Inc.
}
