% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.check.R
\name{map.check}
\alias{map.check}
\title{Compara la distribución geográfica de una especie entre campañas}
\usage{
map.check(gr, esp, camps, newcamp, dns = "Pnew", cor.time = TRUE,
  ind = "p", bw = TRUE, ti = TRUE, plot = TRUE, idi = "l",
  pts = TRUE, cexleg = 1)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 desechos y otros, 9 escoge todos los orgánicos pero excluye desechos}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{Campaña (o grupo de campañas) a representar en el mapa de un año (o grupo de años) concreto (s) (XX) con los que se quiere comparar la última campaña (newcamp): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{newcamp}{última campaña (que se compara con las anteriores) Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{cor.time}{si T corrige la captura del lance a media hora}

\item{ind}{saca los datos en "p"eso o "n"úmero.}

\item{bw}{Gráfico en blanco en negro si T o en color si F.}

\item{ti}{Si T añade título al gráfico, el nombre de la especie en latín.}

\item{plot}{saca el gráfico (T) o lo guardar como objeto para componer con otros gráficos (F)}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e").}

\item{pts}{Saca los puntos de los datos}

\item{cexleg}{Varia el tamaño de letra de los ejes y del número de la leyenda.}
}
\description{
Compara la distribución geográfica de la especie gr,esp entre las campañas camps y la newcamp
}
\examples{
map.check(1,17,Nsh[1:27],"N11","Cant",ind="n")
}
\seealso{
Other mapas: \code{\link{GrafMarksGPS_2}},
  \code{\link{GrafMarksGPS}}, \code{\link{MapEcol.camp}},
  \code{\link{MapGaltal}}, \code{\link{MapLansGPS}},
  \code{\link{armap.camp}}, \code{\link{armap.tot}},
  \code{\link{maparea}}, \code{\link{maphistage}},
  \code{\link{maphistal}}, \code{\link{maphist}}
}
