% Generated by roxygen2 (4.0.2): do not edit by hand
\name{maparea}
\alias{maparea}
\title{Mapa general del banco de Porcupine}
\usage{
maparea(es = T, leg = T, bw = F, dens = 0)
}
\arguments{
\item{es}{Textos en español, si F en inglés}

\item{leg}{incluye la leyenda con los colores/patrones de los estratos}

\item{bw}{Si T mapa en blanco y negro, si F colorea los estratos y sectores}

\item{dens}{si mayor de 0 las superficies de los estratos tienen patrones de líneas}
}
\description{
Mapa con la estratificación y el área general del banco de Porcupine incluyendo parte de Irlanda
}

