% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MapArsa}
\alias{MapArsa}
\title{Mapa del Golfo de Cádiz}
\usage{
MapArsa(lwdl = 1, cuadr = F, cuadrMSFD = F, ICESrect = F, ax = T,
  wmf = F, es = T, places = T)
}
\arguments{
\item{lwdl}{Ancho de las líneas del mapa}

\item{cuadr}{Si T saca las cuadrículas de 5x5 millas naúticas}

\item{cuadrMSFD}{Si T dibuja caudrícula de 10 millas naúticas utilizada para la evaluación de la estrategia marina (MSFD)}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{ax}{Si T saca los ejes x e y}

\item{wmf}{Si T saca a fichero metafile Arsaconc.emf}

\item{es}{Si T saca titulos y ejes en español}

\item{places}{Si T saca ciudades y puntos geográficos de referencia}
}
\value{
Saca en pantalla el mapa y es utilizada por otras funciones
}
\description{
Función auxiliar para sacar mapas de la campaña ARSA
}
\examples{
MapArsa()
}
\seealso{
{\link{MapCant}} {\link{mapporco}}
}

