% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MapCant}
\alias{MapCant}
\title{Mapa de Demersales Norte para una sola campaña}
\usage{
MapCant(gr, esp, camp, dns, color = 1, add = F, escala = NA, ti = F,
  peso = T)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 Moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie númerico o caracter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{color}{Color de los puntos que representan las abundancias}

\item{add}{Si T añade los puntos al gráfico actual, si F dibuja uno nuevo}

\item{escala}{Varia el tamaño de los puntos}

\item{ti}{Si T añade titulo al mapa, el nombre de la especie en latín}

\item{peso}{Si T representa los datos de biomasa, si F representa los datos de abundancia}
}
\description{
Crea un mapa para el Cantábrico y Galicia con información para la especie solicitada en la campaña solicitada, sólo una campaña para mas campañas ver maphist
}
\seealso{
maphist {\link{maphist}}
}

