% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapComp.R
\name{MapComp}
\alias{MapComp}
\title{Mapas de presentación de datos de una campaña comparando dos sectores/lances, creada para la campaña N21 con dos barcos, el lance 66 es el último que se hizo con el Miguel Oliver}
\usage{
MapComp(
  gr,
  esp,
  camp = "N21",
  dns = "Cant",
  lance = 66,
  ti = T,
  plot = T,
  out.dat = F,
  ind = "p",
  idi = "l",
  es = F,
  layout = NA,
  leg = T,
  pts = F,
  ceros = TRUE,
  escmult = 0.25,
  cexleg = 1,
  years = F,
  bw = FALSE,
  notch = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{lance}{Lance a partir del cual se utiliza el otro barco en principio solo para usar con la campaña de 2021}

\item{ti}{Añade el nombre de la especie en latín sin T, si F no añade titulo}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{ind}{Parámetro a representar saca los datos en "p"eso o "n"úmero}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{es}{if T usa español, si no etiquetas y textos en inglés}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}

\item{leg}{Si T añade la leyenda}

\item{pts}{Saca los puntos de los datos}

\item{ceros}{Añade puntos para los datos igual a 0 si T, si F no incluye x en los valores 0}

\item{escmult}{Varía la relación de tamaño de los puntos con la leyenda y el máximo en los datos}

\item{cexleg}{Varía el tamaño de letra de los ejes y del número de la leyenda}

\item{years}{Si T saca los años como nombre de campaña}

\item{bw}{Gráfico en blanco en negro si T o en color si F}
}
\description{
Crea un mapa con la distribucion en biomasa o numero para distintas el Cantábrico (dns=Cant)
}
