% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapEcol.camp.R
\name{MapEcol.camp}
\alias{MapEcol.camp}
\title{Mapa de índices ecológicos para un grupo en una campaña}
\usage{
MapEcol.camp(
  gr,
  esp = "999",
  camp,
  dns = "Porc",
  ind = "n",
  indec = "div",
  plot = TRUE,
  subtit = TRUE,
  bw = FALSE,
  ti = " ",
  idi = "l",
  es = TRUE,
  out.dat = FALSE,
  layout = NA,
  cex.pt = 1,
  cexleg = 1,
  years = TRUE,
  graf = FALSE,
  xpng = 1200,
  ypng = 800,
  ppng = 15
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados. 6 Desechos y otros no se incluye en esta función}

\item{esp}{ha de ser 999 cuando se quiere incluir todas las especies del grupo, o elegir todas las especies deseadas con los codigos de las especies}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{ind}{Elige el valor (n)úmero o (p)eso sobre el que se calculan los índices de diversidad, dominancia....}

\item{indec}{Elige el índice ecológico a representar: opciones disponibles: Shannon-Wiener: 'div', Número de especies: 'nesp' y Diversidad Simpson: 'simp'.}

\item{plot}{Saca el gráfico (si T) o si se salva como objeto se puede componer para componer con otros gráficos de lattice (F)}

\item{subtit}{Si T, es el valor por defecto, incluye un subtítulo abajo del gráfico con el índice representado, si F no se pone el subtítulo pero \emph{convendría que añadirlo a posteriori}}

\item{ti}{por defecto en blanco (" "), pero pone el texto que se quiera poner como título}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{es}{Si T saca los titulos y rotulos en español, si F en inglés}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}

\item{cex.pt}{Varía el tamaño de los puntos en los gráficos}

\item{cexleg}{varía el tamaño del texto de la leyenda y los ejes}

\item{years}{Si T saca los años como nombre de campaña en los paneles lattice de campañas}

\item{graf}{si F no el gráfico va a pantalla, si nombre fichero va a fichero en el directorio en que está wdf}

\item{xpng}{width archivo png si graf es el nombre del fichero}

\item{ypng}{height archivo png si graf es el nombre del fichero}

\item{ppng}{points png archivo si graf es el nombre del fichero}
}
\value{
Saca el mapa de diversidad en la campaña seleccionada.
}
\description{
Utiliza los datos del Camp representar la variación geográfica de los índices ecológicos
}
\examples{
MapEcol.camp(c(2:5),999,"N24","Cant",ind = "n",indec = "div",ti="Invertebrados")
dumbecol<-MapEcol.camp(1,999,Nsh[25:30],"Cant",ind="n",bw=TRUE,indec="simp",out.dat=TRUE,layout=c(2,3),ti=" ")
dumbecol<-MapEcol.camp(2:5,999,Nsh[25:30],"Cant",ind="n",ti="Invertebrados, sin peces",bw=TRUE,indec="simp",out.dat=TRUE,layout=c(2,3))
dumbecol$estrato<-cut(dumbecol$prof,c(0,70,120,200,500,900),c("A1","A","B","C","D"))
lattice::bwplot(numbesp~estrato|camp,dumbecol,horizontal=FALSE,main="Número de especies",xlab="Estrato")
lattice::bwplot(div~estrato|camp,dumbecol,horizontal=FALSE,main="Shannon Wiener",xlab="Estrato")
}
\seealso{
Other mapas: 
\code{\link{BelgicaMound}},
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapGaltal}()},
\code{\link{MapHidro.camp}()},
\code{\link{MapHidro_b.camp}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{MatchHidro.ctd}()},
\code{\link{NepFUs.camp}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{VedasBesugo}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{hmmSAC}},
\code{\link{map.check}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphistal.one}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other ecologia: 
\code{\link{ecolgr.camp}()}
}
\concept{ecologia}
\concept{mapas}
