% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapHidro.camp.R
\name{MapHidro.camp}
\alias{MapHidro.camp}
\title{Mapa de índices ecológicos para un grupo en una campaña}
\usage{
MapHidro.camp(
  camp,
  dns = "Porc",
  ind = "t",
  plot = TRUE,
  subtit = TRUE,
  bw = FALSE,
  ti = TRUE,
  es = TRUE,
  out.dat = FALSE,
  layout = NA,
  cex.pt = 1,
  cexleg = 1,
  years = TRUE,
  graf = FALSE,
  xpng = 1200,
  ypng = 800,
  ppng = 15,
  cuts = 4
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{ind}{Elige el valor (t)emperatura o (s)alinidad}

\item{plot}{Saca el gráfico (si T) o si se salva como objeto se puede componer para componer con otros gráficos de lattice (F)}

\item{ti}{Añade la Campaña, si F no añade titulo}

\item{es}{Si T saca los titulos y rotulos en español, si F en inglés}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}

\item{cex.pt}{Varía el tamaño de los puntos en los gráficos}

\item{cexleg}{varía el tamaño del texto de la leyenda y los ejes}

\item{years}{Si T saca los años como nombre de campaña en los paneles lattice de campañas}

\item{graf}{si F no el gráfico va a pantalla, si nombre fichero va a fichero en el directorio en que está wdf}

\item{xpng}{width archivo png si graf es el nombre del fichero}

\item{ypng}{height archivo png si graf es el nombre del fichero}

\item{ppng}{points png archivo si graf es el nombre del fichero}

\item{sub}{Si T añade el parámetro (temperatura, salinidad) como subtítulo bajo el gráfico en inglés o español}
}
\value{
Saca el mapa de diversidad en la campaña seleccionada.
}
\description{
Utiliza los datos del Camp representar la variación geográfica de los datos hidrologicos, temperatura y salinidad
}
\examples{
MapHidro.camp("N24","Cant",ind="t",ti=T)
}
\seealso{
Other mapas: 
\code{\link{BelgicaMound}},
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapHidro_b.camp}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{NepFUs.camp}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{hmmSAC}},
\code{\link{map.check}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other hidrología: 
\code{\link{MapHidro_b.camp}()}
}
\concept{hidrología}
\concept{mapas}
