% Generated by roxygen2 (4.0.2): do not edit by hand
\name{maphistage}
\alias{maphistage}
\title{Mapa distribución por edad en la campaña}
\usage{
maphistage(gr, esp, camp, dns = "Pnew", age, plus = 8, cor.time = T,
  bw = T, ti = T, plot = T, out.dat = F, ind = "n", idi = "l",
  es = T, layout = NA, ceros = F, years = F, mediahora = 1)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña o campañas a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa". Medits "Medi" (únicamente para sacar datos al IBTS, no gráficos)}

\item{age}{Edad solicitada}

\item{plus}{Edad plus: incluir la edad considerada como plus, solo afecta si se pide como plus la edad solicitada que suma todas las edades mayores}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{bw}{Gráfico en blanco en negro si T o en color si F}

\item{ti}{Si T añade titulo al mapa, el nombre de la especie en latín}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{ind}{Parámetro a representar saca los datos en "p"eso o "n"úmero}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{es}{Si T saca los ejes y legendas en español, si F lo saca en inglés}

\item{layout}{Organización de gráficos en filas columnas c(r,c)}

\item{ceros}{Añade puntos para los datos igual a 0 si T, si F no incluye x en los valores 0}

\item{years}{Si T permite sacar los años como nombre de campaña en los paneles lattice de campañas}

\item{mediahora}{Valor para obtener abundancias por hora si media hora es mayor}
}
\value{
Si out.dat=T devuelve un data.frame con columnas: lan,lat,long,0,1,2,...,Age+ (número de individuos edad 0,1,2...),camp. Si out.dat=F saca el Gráfico en pantalla o como objeto para combinar con otros Gráficos con print.trellis
}
\description{
Mapa de la distribución geográfica por edad de la especie en las campañas solicitadas
}
\examples{
maphistage("1"," 42","N05","Cant",2,ceros=T,out.dat=T)
}

