% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maphistal.r
\name{maphistal}
\alias{maphistal}
\title{Mapa distribución entre tallas tmin y tmax}
\usage{
maphistal(
  gr,
  esp,
  camps,
  dns = "Porc",
  tmin = 0,
  tmax = 999,
  cor.time = TRUE,
  incl2 = TRUE,
  ind = "n",
  ICESrect = FALSE,
  sex = NA,
  bw = FALSE,
  ti = TRUE,
  subti = TRUE,
  plot = TRUE,
  out.dat = FALSE,
  idi = "l",
  layout = NA,
  leg = TRUE,
  ceros = TRUE,
  escmult = 0.25,
  cexleg = 1,
  years = TRUE,
  graf = FALSE,
  xpng = 1200,
  ypng = 800,
  ppng = 15
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustaceos 3 moluscos 4 equinodermos 5 invertebrados. 6 Desechos y otros inorgánicos no tiene sentido sacar tallas, sólo recogidas en peces, crustáceos decápodos y algunos moluscos}

\item{esp}{Código de la especie numerico o caracter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{tmin}{Talla mínima del intervalo de tallas a incluir, si 0 el subtítulo sale =< tmax}

\item{tmax}{Talla máxima del intervalo de tallas a incluir, si 999 el subtítulo sale >= tmin}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{incl2}{Si F no presenta los lances especiales, si T si los tiene en cuenta, pero puede dar problemas por que no puede calcular las abundancias estratificadas}

\item{ind}{Permite elegir entre número "n" o peso "p" peso sólo funciona cuando existen a y b en especies.dbf y se elige sólo una especie}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{sex}{Permite elegir entre machos(1), hembras(2) o indeterminados(3), NA escoge sin tener en cuenta el sexo}

\item{bw}{gráfico en blanco en negro si T o en color si F}

\item{ti}{Si T añade titulo al mapa, el nombre de la especie en latín. También se puede poner como list con parámetros \strong{main}: el título, \strong{font} y demás.}

\item{subti}{ti T añade un subtítulo bajo la gráfica con el rango de tallas seleccionado.}

\item{plot}{Saca el gráfico (T) o si se salva como objeto se puede componer para componer con otros gráficos de lattice (F)}

\item{out.dat}{Si T el resultado final de la funcion es la figura en pantalla, pero los datos en objeto}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}

\item{leg}{Si T añade la leyenda}

\item{ceros}{Añade puntos para los datos igual a 0 si T, si F no incluye x en los valores 0}

\item{escmult}{Varía la relación de tamaño de los puntos con la leyenda y el máximo en los datos}

\item{cexleg}{Varía el tamaño de letra de los ejes y del número de la leyenda}

\item{years}{Si T saca los años como nombre de campaña en los paneles lattice de campañas}

\item{graf}{Si F no saca nada, si pones el nombre de un gráfico lo saca saca como archivo png y al final del proceso dice dónde está el mapa con ese nombre:}

\item{xpng}{width archivo png si graf es el nombre del fichero}

\item{ypng}{height archivo png si graf es el nombre del fichero}

\item{ppng}{points png archivo si graf es el nombre del fichero}
}
\value{
Si out.dat=TRUE devuelve un data.frame con columnas: lan,lat,long,prof,numero (de individuos entre tmin y tmax),camp, si out.dat=F saca el gráfico en pantalla o como objeto para combinar con otros gráficos con print.trellis
}
\description{
Saca mapas de distribución de una especie entre tallas tmin y tmax para varias campañas
}
\examples{
maphistal(1,50,Psh[1:12],"Porc",1,23,layout=c(4,3),out.dat=TRUE)
}
\seealso{
Other mapas: 
\code{\link{BelgicaMound}},
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapHidro.camp}()},
\code{\link{MapHidro_b.camp}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{NepFUs.camp}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{hmmSAC}},
\code{\link{map.check}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphist}()}

Other tallas: 
\code{\link{MapGaltal}()}
}
\concept{mapas}
\concept{tallas}
