% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapIberia.R
\name{MapIberia}
\alias{MapIberia}
\title{Mapa de la Península Ibérica completo}
\usage{
MapIberia(
  xlims = c(-10.7, 5),
  ylims = c(35.9, 44.5),
  lwdl = 1,
  cuadr = FALSE,
  cuadrMSFD = FALSE,
  ICESrect = FALSE,
  ICESlab = FALSE,
  ICESlabcex = 0.6,
  FU = NULL,
  ColFU = "chartreuse",
  FUsLab = FALSE,
  bw = F,
  ax = TRUE,
  wmf = FALSE,
  es = TRUE,
  places = TRUE,
  pais = TRUE,
  escmult = 1,
  dens = 20,
  add = FALSE
)
}
\arguments{
\item{xlims}{Define los limites longitudinales del mapa, los valores por defecto son los del total del ?rea de la campaña}

\item{ylims}{Define los limites latitudinales del mapa, los valores por defecto son los del total del ?rea de la campaña}

\item{lwdl}{Ancho de las líneas del mapa}

\item{cuadr}{Si T saca las cuadrículas de 5x5 millas naúticas}

\item{cuadrMSFD}{Si T dibuja caudr?cula de 10 millas naúticas utilizada para la evaluaci?n de la estrategia marina (MSFD)}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{ICESlab}{Si T incluye las etiquetas de los rectángulos ICES}

\item{ICESlabcex}{tamaño del ICESlab en cex, .5 por defecto subirlo si se quiere más grande}

\item{ColFU}{Por defecto chartreuse, color para las zonas de las FU de cigalas}

\item{FUsLab}{Por defecto F, pero si T incluye una etiqueta con los nombres de las FUs seleccionadas en FUs}

\item{bw}{si T mapa con tierra en gris, si F tierra en color}

\item{ax}{Si T saca los ejes x e y}

\item{wmf}{Si T saca a fichero metafile Iberia.emf}

\item{es}{Si T saca titulos y ejes en español}

\item{places}{Si T saca ciudades y puntos geográficos de referencia}

\item{pais}{si T incluye los carteles de los países represntados}

\item{escmult}{=1 aumenta o disminuye el tamaño de las etiquetas y textos menos de 1 lo disminuye, más de uno lo aumenta}

\item{add}{si T se añade al mapa si no saca la pantalla nueva}

\item{FUs}{Por defecto NULL pero si se incluye un vector con la lista de las unidades funcionales las pinta (las disponibles en MapIberia son FU31, FU25, FU26, FU27, FU28, FU29 y FU30)}
}
\value{
Saca en pantalla el mapa y es utilizada por otras funciones
}
\description{
Función auxiliar para sacar mapas de la campaña MEDITS
}
\examples{
MapIberia()
MapIberia(FU=c("FU25","FU31"),ICESrect=TRUE,ICESlab=T)
}
\seealso{
Other mapas base: 
\code{\link{MapArsa}()},
\code{\link{MapCant}()},
\code{\link{MapMedit}()},
\code{\link{MapNort}()},
\code{\link{MapPorc}()},
\code{\link{mapporco}()}

Other Medits: 
\code{\link{MapMedit}()}
}
\concept{Medits}
\concept{mapas base}
