% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapICESStatRec_plotrix.R
\name{mapICESStatRec.plotrix}
\alias{mapICESStatRec.plotrix}
\title{Mapa distribución por edad en rectángulos estadísticos ICES en la campaña}
\usage{
mapICESStatRec.plotrix(
  gr,
  esp,
  camp,
  dns = "Cant",
  plus = 3,
  cor.time = TRUE,
  ICESrect = TRUE,
  ICESlab = FALSE,
  AltAlk = NA,
  incl2 = FALSE,
  bw = FALSE,
  ti = TRUE,
  plot = TRUE,
  out.dat = FALSE,
  ind = "n",
  idi = "l",
  es = TRUE,
  years = TRUE,
  escmult = 0.1,
  ceros = FALSE,
  mediahora = 2,
  graf = FALSE,
  xpng = NA,
  ypng = NA,
  ppng = 15
)
}
\arguments{
\item{gr}{Grupo de la especie: Solo hay dados de edad para algunos peces y cigala ?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña o campañas a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa". Medits "Medi" no tiene sentido en esta gráfica por no utilizar rectángulos ICES}

\item{plus}{Edad plus: incluir la edad considerada como plus, solo afecta si se pide como plus la edad solicitada que suma todas las edades mayores}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{ICESrect}{Si T representa las líneas de los rectángulos ICES. No para dns="Medi"}

\item{ICESlab}{Si T saca etiqueta del rectángulo ices en la parte superior. No para dns="Medi"}

\item{AltAlk}{Clave talla edad alternativa sin ruta ni extensión, NA por defecto usa la clave de la campaña edadXYY.dbf}

\item{incl2}{si T representa los datos de lances especiales, si F los excluye}

\item{bw}{Gráfico en blanco en negro si T o en color si F}

\item{ti}{Si T añade titulo al mapa, el nombre de la especie en latín y año de la campaña o campaña si years=F}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{es}{Si T saca los ejes y legendas en español, si F lo saca en inglés}

\item{years}{Si T permite sacar el dato de la campaña como nombre campaña o como año}

\item{ceros}{Añade puntos para los datos igual a 0 si T, si F no incluye x en los valores 0}

\item{mediahora}{Valor para obtener abundancias por hora si media hora es mayor}

\item{graf}{si F el gráfico sale en la pantalla, si nombre fichero va a fichero en el directorio de trabajo del Rstudio ver getwd()}

\item{xpng}{width archivo png si graf es el nombre del fichero, tienen valores default para cant,porc,arsa, no tocar si no}

\item{ypng}{height archivo png si graf es el nombre del fichero, tienen valores default para cant,porc,arsa, no tocar si no}

\item{ppng}{points png archivo si graf es el nombre del fichero}

\item{age}{Edad solicitada}
}
\value{
Si out.dat=TRUE devuelve un data.frame con columnas: lan,lat,long,0,1,2,...,Age+ (número de individuos edad 0,1,2...),camp. Si out.dat=F saca el Gráfico en pantalla o como objeto para combinar con otros Gráficos con print.trellis
}
\description{
Mapa de la distribución geográfica en rectángulos ICES por edad de la especie en la campaña solicitada
}
\examples{
mapICESStatRec.plotrix(1,74,"N05","Cant",plus=3,ICESlab = T,bw=F)
mapICESStatRec.plotrix(1,43,"P21","Porc",plus=7,ICESlab = T,bw=F)
mapICESStatRec.plotrix(1,73,"219","Arsa",plus=3,ICESlab = T,bw=F)
}
\seealso{
Other mapas: 
\code{\link{BelgicaMound}},
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapHidro.camp}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{NepFUs.camp}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{hmmSAC}},
\code{\link{map.check}()},
\code{\link{mapICESStatRec}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other edades: 
\code{\link{ALKs.dns.camp}()},
\code{\link{AbAgStatRec.camp}()},
\code{\link{GetAlk.camp}()},
\code{\link{bubbage.camp}()},
\code{\link{datagegr.camp}()},
\code{\link{edadsect.camp}()},
\code{\link{edadstr.camp}()},
\code{\link{grafedtal.camps}()},
\code{\link{grafedtal.camp}()},
\code{\link{logabage.camp}()},
\code{\link{logabage2.camp}()},
\code{\link{mapICESStatRec}()},
\code{\link{maphistage}()}
}
\concept{edades}
\concept{mapas}
