% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MapLansGPS}
\alias{MapLansGPS}
\title{Gráfico con los lances en segmentos}
\usage{
MapLansGPS(camp, dns = "Pnew", incl0 = F, xlims = c(-10.25, -1.4),
  ylims = c(41.82, 44.48))
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)gr Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{incl0}{Si T se incluyen los lances nulos}

\item{xlims}{Define los limites longitudinales del mapa, Solo para el Cantabrico por defecto -10.25 y -1.4 oeste}

\item{ylims}{Define los limites latitudinales del mapa, Solo para el Cantabrico por defecto 41.82 y 44.48 norte}
}
\value{
Devuelve un data.frame con datos de cada lance, las variables dependen de la selección de hidro y redux
}
\description{
gráfica todos los lances de una campaña con las distancias recorridas en cada uno de ellos
}
\examples{
MapLansGPS("12C","Cant")
}
\seealso{
{\link{datlan.camp}}
}

