% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mapporco}
\alias{mapporco}
\title{Mapa del Banco de Porcupine sin referencias en tierra}
\usage{
mapporco(lwdl = 1, cuadr = F, ICESrect = F, label = F, colo = 2,
  ax = T, wmf = F, corners = F)
}
\arguments{
\item{lwdl}{Ancho de las líneas del mapa}

\item{cuadr}{Si T saca las cuadrículas de 5x5 millas naúticas}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{label}{Si T saca las etiquetas de cada una de las cuadriculas numeradas consecutivamente por estratos 1A,1B,2B,3A,3B}

\item{colo}{Color de las etiquetas, por defecto rojas}

\item{ax}{Si T saca los ejes x e y}

\item{wmf}{Si T saca a fichero metafile porconc.emf}

\item{corners}{Si T coloca dos puntos rojos en los extremos nordeste y suroeste para ajustar mapas al PescaWin con ax=F}
}
\value{
Saca en pantalla el mapa y es utilizada por otras funciones, si wmf=T lo saca a metafile para fondo del pescawin
}
\description{
Función auxiliar para sacar el mapa de la campaña Porcupine
}
\seealso{
MapNort {\link{MapNort.r}} MapCant {\link{MapCant.r}}
}

