% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.gaps.R
\name{mark.gaps}
\alias{mark.gaps}
\title{Busca huecos en un fichero de lances creado con sorteo()}
\usage{
mark.gaps(x, buffer = 0.1177, srv.dsg = Porc.dsg)
}
\arguments{
\item{x}{es el fichero de lances a comprobar}

\item{buffer}{Distancia entre cuadrículas contiguas que es posible elegir. El valor por defecto asegura que no se seleccionan dos cuadrículas contíguas}

\item{srv.dsg}{Diseño del muestreo. data.frame de dos columnas, strat, stations (estrato y número de lances)}
}
\value{
Devuelve un mapa con los lances y los huecos en rojo además de los puntos sin lances alredador
}
\description{
En un fichero de lances creado por sorteo busca los huecos de cuadrículas agrupadas sin ningun lance alrededor
}
\seealso{
{\link{mapsorteo}}, {\link{sorteo}}
}
