% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_buff_days.r
\name{max.buff.days}
\alias{max.buff.days}
\title{max.buff.days}
\usage{
\method{max}{buff.days}(df, trap)
}
\arguments{
\item{df}{A data frame containing catch records for one
\code{trapPositionID}, with catch counts contained in \code{n.tot}.}

\item{trap}{A vector of length one containing the \code{trapPositionID} for
which buffering is necessary.}
}
\value{
A vector containing three values.  The first contains the row index 
  of the first non-zero / non-\code{NA} in the catch record provided via data
  frame \code{df}. The second contains the row index of the last non-zero / 
  non-\code{NA} in the catch record provided via data frame \code{df}.  The 
  third contains the length of the vector containing 
  \code{trapVisitID}-specific fish counts (vector \code{n.tot}).
}
\description{
df <- df2
}
\details{

}
\examples{
\dontrun{
#   ---- Find the first and last non-zero and non-NA catch record
#   ---- for trap "12345".
vec <- max.buff.days(df,"12345")
}
}
\author{
WEST Inc.
}
\seealso{
\code{F.chuck.zeros}
}

