% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanMaxL.camp.r
\name{MeanMaxL.camp}
\alias{MeanMaxL.camp}
\title{Talla máxima media del grupo de especies elegidas}
\usage{
MeanMaxL.camp(gr = 1, esps, camp, dns = "Cant", lmax = NA,
  cor.time = TRUE, incl2 = FALSE, excl.sect = NA)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos (4 equinodermos 5 invertebrados habitualmente no medidos)}

\item{esps}{Código de las especies seleccionadas. Si se mezclan grupos hay que introducir el mismo numero de elementos en gr que en esps}

\item{camp}{Campaña de la que se extraen los datos: Demersales NYY, Porcupine PYY, Arsa primavera 1YY y Arsa otoño 2YY, Medits MYY. Para sacar resultados de varias campañas {\link{MeanMaxL.camps}}}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa", Medits: "Medi"}

\item{lmax}{Debe incluir los datos de Linf de cada una de las especies en el campo esps. Si se usa NA en un caso utiliza la talla máxima de la campaña camp, se puede dar para cada esps dejando un NA en las especies que se quiera utilizar el maximo de la campaña, o en general}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{incl2}{Si T incluye los lances especiales "2"}

\item{excl.sect}{Excluye los sectores y estratos en cuestion, si NA usa toda el area.}
}
\value{
Devuelve una lista con un data.frame datos por especie: gr, esps, unid, n, lmax y el valor del indice la talla maxima media ponderada a la abundancia
}
\description{
Debería de usar especies de un ecotipo similar (mismo gr) y sobre todo medidas en la misma unidad, por defecto divide los datos que se en el fichero especies salen en mm por 10 y el valor es utilizado en la meida ponderada, revisar que efectivamente están en mm.
Para cada especie calcula la abundancia estratificada y la talla máxima, con ello hace una media ponderada, la salida
}
\examples{
MeanMaxL.camp(gr=1,c(50,42,43,44,45),"N05","Cant")
}
\seealso{
Other MSFD: \code{\link{MeanMaxL.camps}},
  \code{\link{P95talesps.camp}}
}
\concept{MSFD}
