% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Msh}
\alias{Msh}
\title{Serie histórica de la campaña Medits}
\format{A vector with codes for the surveys from 1994 to 2011}
\usage{
Msh
}
\description{
Serie histórica campaña Medits
}
\details{
Serie completa con las campañas realizadas en el Mediterraneo en la campaña MEDITS
}
\seealso{
Nsh {\link{Nsh}} Ash {\link{Ash}}
}
\keyword{datasets}

