% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Nsh}
\alias{Nsh}
\title{Serie histórica de la campaña Demersales plataforma Norte}
\format{A vector with codes for the surveys from 1983 to 2013}
\usage{
Nsh
}
\description{
Serie histórica Demersales Norte
}
\details{
Serie completa con las campañas de cuarto trimestre realizadas en el Cantábrico y Galicia, el año 1987 no hubo campaña
}
\seealso{
Psh {\link{Psh}} Ash {\link{Ash}}
}
\keyword{datasets}

