% Generated by roxygen2 (4.0.2): do not edit by hand
\name{P95talesps.camp}
\alias{P95talesps.camp}
\title{Da el percentil .95 de la distribucion de tallas estratificada de un grupo de especies}
\usage{
P95talesps.camp(gr = 1, esps, camps, dns = "Cant", excl.sect = NA,
  perc = 0.95, graf = T, pch = 20, lty = 1, layout = NA, bw = F,
  cexleg = 0.8, years = T, es = T, ylim = NA)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos (4 equinodermos 5 invertebrados habitualmente no medidos)}

\item{esps}{El grupo de códigos de la especies seleccionadas, no mezclar grupos distintos}

\item{camps}{Campañas de las que se extraen los datos: Demersales Nsh, Porcupine Psh, Arsa primavera Ash y Arsa otoño Ash}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa", Medits: "Medi"}

\item{excl.sect}{Excluye el los sectores y estratos en cuestion, si NA usa toda el area.}

\item{perc}{El percentil de la distribucion de tallas a calcular, .95 es el valor utilizado habitualmente como indicador del maximo de la distribucione de tallas}

\item{graf}{Si T saca el gráfico de evolucion del p95 en la serie historica buscada para cada especie en lattice, si F ignora el gráfico}

\item{pch}{El simbolo a utilizar {\link{pchShow.r}}}

\item{lty}{Ancho de la linea que une los puntos}

\item{layout}{Configuracion del lattice en filas y columnas c(r,c)}

\item{bw}{Si T saca un gráfico en blanco y negro, si F los rotulos de los lattice van en color}

\item{cexleg}{Tamaño de la fuente de los nombres de las especies}

\item{years}{Si T en vez de campañas la serie historica se marca con los años en lugar del codigo de la campaña}

\item{es}{Si T los rotulos de los ejes se muestran en español, si F en ingles}

\item{ylim}{Valor maximo del eje de ordenadas, las tallas}
}
\value{
Devuelve un data.frame con columnas especie, camp y p95tal y year. Ademas crea un gráfico de lattice con la evolucion del p95 para cada especie en un panel distinto
}
\description{
Funcion indicador MSFD Da el percentil perc (por defecto 95, en decimal) de la distribución de tallas de las especies en las serie historica campañas seleccionadas.
}
\examples{
P95talesps.camp(1,c(5,8,9,14,105),Nsh[7:27],"Cnew",layout=NA,graf=T,years=T,es=F)
}

