% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasa.arte.r
\name{pasa.arte}
\alias{pasa.arte}
\title{Funcion auxiliar para el PescaWin}
\usage{
pasa.arte(
  camp,
  lan = NA,
  FileArte = "Arte_",
  folder = "c:/gps/mrk/",
  gps = FALSE,
  graphs = TRUE
)
}
\arguments{
\item{camp}{Campaña de las que se cogen los resultados un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{lan}{Elique el lance del que se quieren comprobar los datos, si no existe dice que no  se encuentra. Si se deja en NA se recoge el lance o los lances seleccionados}

\item{gps}{Si T exporta el fichero directamente como camp.lan al directorio de lances del PescaWin}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{valid}{Si T incluye sólo los lances válidos (>0) y excluye los nulos (=0)}
}
\value{
Devuelve un data.frame con datos long_l,lat_l,prof_l,long_v,lat_v,prof_v en grados decimales con signo
}
\description{
Analiza los datos del fichero de arte de una campaña a partir de fichero arte_XXX.csv en el directorio del pescawin (c:/gps/mrk/XXXX/arte_aXX.csv)
}
\seealso{
Other PescaWin: 
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{arMapLansGPS}()},
\code{\link{pasa.lan}()}
}
\concept{PescaWin}
