% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pasa.lan}
\alias{pasa.lan}
\title{Funcion auxiliar para el PescaWin}
\usage{
pasa.lan(camp, dns = "Pnew", gps = F, valid = T)
}
\arguments{
\item{camp}{Campaña de las que se cogen los resultados un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{gps}{Si T exporta el fichero directamente como camp.lan al directorio de lances del PescaWin}

\item{valid}{Si T incluye sólo los lances válidos y excluye los nulos}
}
\value{
Devuelve un data.frame con datos long_l,lat_l,prof_l,long_v,lat_v,prof_v en grados decimales con signo
}
\description{
Saca un fichero de lances de la campaña con el formato de datos del PescaWin
}

