% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percQ.R
\name{percQ}
\alias{percQ}
\title{percQ}
\usage{
percQ(hrflow)
}
\arguments{
\item{hrflow}{A dataframe housing hourly flow data resulting from a query 
against the Environmental Covariate Database.}
}
\value{
A data frame that mimics the structure of the other covariates. 
  Columns include \code{subSiteID}, \code{measureDate}, \code{percQ}, and
  \code{percQUnitID}, which is set to \code{99}.
}
\description{
Estimate per-trap percent-Q, given hourly flow data and recorded
  water velocity from a CAMP Access database.
}
\details{
Calculations for \code{percQ} utilize denominators obtained via the 
  Big Bend gauge station housed within the Environmental Covariate Database. 
  These numbers include an adjustment for the two canals that often divert 
  flow away from the Sacramento each summer.  In order to obtain an estimate 
  for these diverting flows, the daily means, as measured annually over 11
  years from 2002 to 2013, was subtracted from the values stored in the
  Environmental Covariate Database.  Package \code{campR} data frame
  \code{canal} houses these averages;  submit \code{data(canal)} to see them
  after loading the \code{campR} package.
  
  Numerators derive from velocities stored within the RBDD CAMP Access
  database.
  
  See the \code{"campR"} \code{"percQ"} vignette for more information on the
  derivation of percent-Q.
}
\examples{
 
\dontrun{
ans <- percQ(hrflow))
}
}
\seealso{
\code{data(canal)}, vignette
}
