% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PorcNWSAC.R
\docType{data}
\name{PorcNWSAC}
\alias{PorcNWSAC}
\title{Polígono SAC de}
\format{
A data.frame with long and lat points defining the closed area
}
\usage{
PorcNWSAC
}
\value{
data.frame con long y lat
}
\description{
PorcNWSAC: Area de especial protección irlandesa
}
\details{
Polígono que delimita una zona zona de protección especial de VME del banco de Porcupine denominada Porcupine NorthWestern shelf SAC
}
\examples{
mapporco(SACs=T);
}
\references{
Medidas de protección zonas vulnerables Porcupine \url{http://www.nwwac.org/_fileupload/Opinions and Advice/Year 12/Dictamen_CCANOC_Gestion_Cigala_Porcupine_UF16_2017_ES.pdf}
}
\seealso{
Other mapas: 
\code{\link{BelgicaMound}},
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapHidro.camp}()},
\code{\link{MapHidro_b.camp}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{MatchHidro.ctd}()},
\code{\link{NepFUs.camp}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcSWSAC}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{hmmSAC}},
\code{\link{map.check}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other Porcupine: 
\code{\link{BelgicaMound}},
\code{\link{MapPorc}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcSWSAC}},
\code{\link{Psh}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{hmmSAC}},
\code{\link{maparea}()},
\code{\link{mapporco}()}
}
\concept{Porcupine}
\concept{mapas}
\keyword{datasets}
