% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PorcShelfSAC.R
\docType{data}
\name{PorShelfSAC}
\alias{PorShelfSAC}
\alias{PorcShelfSAC}
\title{Polígono SAC de}
\format{
A data.frame with long and lat points defining the closed area
}
\usage{
PorcShelfSAC
}
\value{
data.frame con long y lat
}
\description{
PorcShelfSAC: Area de especial protección irlandesa
}
\details{
Polígono que delimita una zona ampliada más grande que la zona de veda de la cigala en el Porcupine Bank, la zona de veda fue propuesta por los armadores españoles, y puesta en vigor en 2010 desde el 1 de mayo al 31 de julio. Desde entonces ha estado vigente en diversos periodos, pero al menos todo el mes de mayo desde 2013 a 2016.
}
\examples{
mapporco(SACs=T);
}
\references{
Medidas de protección zonas vulnerables Porcupine \url{http://www.nwwac.org/_fileupload/Opinions and Advice/Year 12/Dictamen_CCANOC_Gestion_Cigala_Porcupine_UF16_2017_ES.pdf}
}
\seealso{
Other mapas: 
\code{\link{BelgicaMound}},
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapHidro.camp}()},
\code{\link{MapHidro_b.camp}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{MatchHidro.ctd}()},
\code{\link{NepFUs.camp}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{VedasBesugo}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{hmmSAC}},
\code{\link{map.check}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphistal.one}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other Porcupine: 
\code{\link{BelgicaMound}},
\code{\link{MapPorc}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{Psh}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{hmmSAC}},
\code{\link{maparea}()},
\code{\link{mapporco}()}
}
\concept{Porcupine}
\concept{mapas}
\keyword{datasets}
