% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresenciaEsp.camp.R
\name{PresenciaEsp.camp}
\alias{PresenciaEsp.camp}
\title{Datos de presencia de especie en campañas en dns}
\usage{
PresenciaEsp.camp(gr, esp, dns, tablas = FALSE, redux = TRUE)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 desechos y otros}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Función para una sola especie}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{tablas}{Si T muestra al final de los resultados una lista con todos los ficheros de definición (CAMP con su identificador de campaña), fauna y tallas presentes en el directorio.}
}
\value{
Devuelve dos vectores con las campañas con presencia de la especie, y con las campañas con datos de talla de la especie. Tb incluye información sobre campañas que no tienen datos en el fichero FAUNAXXX.dbf y/o NTALLXXX.dbf
}
\description{
Función de acceso a datos:
A partir de ficheros de fauna.dbf presentes en directorio comprueba presencia de especie
}
\examples{
PresenciaEsp.camp(1,220,"Porc")
}
\seealso{
Other Series datos: 
\code{\link{As1}},
\code{\link{As2}},
\code{\link{CampsDNS.camp}()},
\code{\link{Msh}},
\code{\link{Nsh}},
\code{\link{Psh}}
}
\concept{Series datos}
