% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrSh.R
\docType{data}
\name{PrSh}
\alias{PrSh}
\title{Serie histórica de la campaña Demersales Norte}
\format{
A vector with codes for the surveys between 1984 to 1988 (not in 1985)
}
\usage{
PrSh
}
\description{
Serie histórica Campañas Demersales Primavera
}
\details{
Serie completa con las campañas de primavera realizadas en el Cantábrico-Galicia (Arrastre noroeste)
}
\examples{
windows();maphist(2,19,PrSh,"Cant",layout=c(2,2))
}
\seealso{
Other Series datos: 
\code{\link{As1}},
\code{\link{As2}},
\code{\link{CampsDNS.camp}()},
\code{\link{Msh}},
\code{\link{Nsh}},
\code{\link{PresenciaEsp.camp}()},
\code{\link{Psh}}
}
\concept{Cant}
\concept{Series datos}
\keyword{datasets}
