% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Psh}
\alias{Psh}
\title{Serie hist칩rica de la campa침a Porcupine Bank}
\format{A vector with codes for the surveys from 2001 to 2014}
\usage{
Psh
}
\description{
Serie hist칩rica Porcupine
}
\details{
Serie completa con las campa침as realizadas en Porcupine
}
\seealso{
Nsh {\link{Nsh}} Ash {\link{Ash}}
}
\keyword{datasets}

