% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qcdistlan.camp}
\alias{qcdistlan.camp}
\title{Comprueba la distancia recorrida en los lances y la consistencia con recorrido por camp}
\usage{
qcdistlan.camp(camp, dns = "Cant", todos = F, pc.error = 2)
}
\arguments{
\item{camp}{campaña a revisar los datos en formato Camp Xyy}

\item{dns}{Origen de bases de datos: Cant cantábrico, Pnew Porcupine}

\item{todos}{Por defecto F. Si T lista todos los lances con valores, si no sólo los que pc.error>error}

\item{pc.error}{porcentaje de error aceptable para no mostrar los lances erróneos}
}
\value{
Devuelve un data.frame con lance, recorrido según camp, recorrido haversine y % de error
}
\description{
Sirve para control de calidad y asegurarse que los datos de distancias y posiciones son correctos
}
\examples{
qcdistlan.camp("C14","Cant",pc.error=.01)
}
\references{
gcd.hf function gives the haversine calculation of distance between two geographic points (see: Pineda-Krch, M. 2010. http://www.r-bloggers.com/great-circle-distance-calculations-in-r/
}
\seealso{
Other Control.de.calidad: \code{\link{qcLW.camp}};
  \code{\link{qcLWbucl.camp}}; \code{\link{qcTalPez.lan}};
  \code{\link{qcdtall.camp}}; \code{\link{qcdtallrev.camp}}
}

