% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcdtall.camp.r
\name{qcdtall.camp}
\alias{qcdtall.camp}
\title{Revisión del histograma de distribución de tallas}
\usage{
qcdtall.camp(
  gr,
  esp,
  camp = "P12",
  dns = "Porc",
  lance,
  ti = FALSE,
  legend = TRUE,
  idi = "l",
  ymax = NA,
  mm = FALSE,
  areg = NA,
  breg = NA
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{lance}{Permite seleccionar el lance}

\item{ti}{Si T añade título al gráfico, el nombre de la especie en latín}

\item{legend}{Si T añade leyenda}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{ymax}{Valor máximo del eje y}

\item{mm}{Usar T para especies medidas en milímetros}

\item{areg}{Para cambiar el coeficiente "a" de la regresión si es distinto al archivado en el CAMP}

\item{breg}{Para cambiar el coeficiente "b" de la regresión si es distinto al archivado en el CAMP}
}
\value{
Saca en pantalla un gráfico con el histograma de individuos medidos divididos en sus categorías y ponderaciones, ademas del numero ideal de individuos medidos
}
\description{
Revisa el histograma de distribución de tallas total de la especie en un lance concreto de una campaña.
Muestra la distribución de talla de cada categoría,los factores de ponderación por categoría, el número total de individuos medidos,
y el número ideal de individuos a medir (10 por talla: Gerritsen & McGrath 2007)
}
\examples{
qcdtall.camp(1,43,"P08","Porc",lance=35)
}
\references{
Gerritsen & McGrath. 2007. Precision estimates and suggested sample sizes for length-frequency data.Fish. Bull. 106: 116-120)
}
\seealso{
{\link{dtall.camp}}

Other Control de calidad: 
\code{\link{qcLW.camp}()},
\code{\link{qcLWbucl.camp}()},
\code{\link{qcTalPez.camp}()},
\code{\link{qcTalPez.lan}()},
\code{\link{qcdistlan.camp}()},
\code{\link{qcdtallrev.camp}()},
\code{\link{qclandatr.camp}()}
}
\concept{Control de calidad}
