% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qclandata.camp.R
\name{qclandata.camp}
\alias{qclandata.camp}
\title{A partir de los HH de DATRAS con CAMPtoHH comprueba que los lances son de día y avisa si son de noche}
\usage{
qclandata.camp(camp, dns = "Cant", out.dat = FALSE)
}
\arguments{
\item{camp}{campaña a revisar los datos en formato Camp Xyy}

\item{dns}{Origen de bases de datos: "Cant" cantábrico, "Porc" o "Pnew" Porcupine, "Arsa" para el Golfo de Cádiz y "Medi" para MEDITS}
}
\value{
Devuelve un data.frame con campaña, lance, recorrido, recorrido según la fórmula de Haversine, recorrido según la velocidad x el tiempo, velocidad, tiempo, rumbo, rumbo estimado según posiciones,velocidad calculada a partir de la distancia y el tiempo, y los porcentajes de errores de distancia, velocidad y rumbo.
}
\description{
Sirve para control de calidad y asegurarse que los datos de horas y fechas cuando las fechas son muy distintas canta.
}
\examples{
qcdistlan.camp("C14","Cant",pc.error=.01)
qcdistlan.camp("216","Arsa",pc.error=.01)
}
\references{
\code{\link[suncalc]{getSunlightTimes}}
}
\seealso{
{\link{MapLansGPS}}

Other Control de calidad: 
\code{\link{qcLW.camp}()},
\code{\link{qcLWbucl.camp}()},
\code{\link{qcTalPez.camp}()},
\code{\link{qcTalPez.lan}()},
\code{\link{qcdistlan.camp}()},
\code{\link{qcdtall.camp}()},
\code{\link{qcdtallrev.camp}()}
}
\concept{Control de calidad}
