% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcLanDatr.camp.R
\name{qclandatr.camp}
\alias{qclandatr.camp}
\title{Comprueba la distancia recorrida en los lances y la consistencia con recorrido y la velocidad en los datos del CAMP}
\usage{
qclandatr.camp(camp, dns)
}
\arguments{
\item{camp}{campaña a revisar los datos en formato Camp Xyy}

\item{dns}{Origen de bases de datos: "Cant" cantábrico, "Porc" o "Pnew" Porcupine, "Arsa" para el Golfo de Cádiz y "Medi" para MEDITS}

\item{plots}{si T abre una pantalla con cuatro gráficos de comparacion entre camp y DATRAS}
}
\value{
Devuelve un gráfico con cuatro plots puertas~prof, calones~prof, vertical~prof y lances en área
}
\description{
Sirve para control de calidad y asegurarse que los datos de distancias, posiciones, y las horas de largada y virada son correctos y no salen del estándar de largar y virar de día
}
\examples{
qclansdatr.camp("N23","Cant")
qclansdatr.camp("216","Arsa")
}
\seealso{
{\link{MapLansGPS}}

Other Control de calidad: 
\code{\link{qcLW.camp}()},
\code{\link{qcLWbucl.camp}()},
\code{\link{qcTalPez.camp}()},
\code{\link{qcTalPez.lan}()},
\code{\link{qcdistlan.camp}()},
\code{\link{qcdtall.camp}()},
\code{\link{qcdtallrev.camp}()}
}
\concept{Control de calidad}
