% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcLW.camp.r
\name{qcLW.camp}
\alias{qcLW.camp}
\title{Outliers capturas según tallas medidas y relación talla-peso}
\usage{
qcLW.camp(gr, esp, camp = "P11", dns = "Porc", margerr = 20,
  out.dat = FALSE, mm = FALSE, areg = NA, breg = NA)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{margerr}{Margen de error que delimita los datos que estan dentro o fuera del intervalo.}

\item{out.dat}{si T saca los datos estimados, observados y el error correspondiente. Por defecto (si F) saca estos datos, únicamente, de aquellos lances con errores mayores que margerr}

\item{mm}{Usar T para especies medidas en milímetros}

\item{areg}{Para cambiar el coeficiente "a" de la regresión si es distinto al archivado en el CAMP}

\item{breg}{Para cambiar el coeficiente "b" de la regresión si es distinto al archivado en el CAMP}
}
\value{
Saca una tabla con datos estimados, observados y el error asociado a la relación talla peso y la distribución de tallas. Si out.dat=T crea un data.frame con los datos para todos los lances de la campaña, en caso contrario sólo saca en la consola los que están fuera de los margenes de error
}
\description{
Busca outliers en las capturas en el fichero de tallas a partir de la relación talla-peso y los pesos y tallas muestreadas. El tamaño de los círculos reflejan el numero de peces medidos, y los distintos colores marcan las distintas categorias si existe muestreo por categorías.
}
\examples{
qcLW.camp(1,50,"N08","Cant")
}
\seealso{
Other Control de calidad: \code{\link{qcLWbucl.camp}},
  \code{\link{qcTalPez.lan}}, \code{\link{qcdistlan.camp}},
  \code{\link{qcdtall.camp}}, \code{\link{qcdtallrev.camp}}
}
