% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qcTalPez.lan}
\alias{qcTalPez.lan}
\title{Comprueba que están medidas todas las especies en capturas y viceversa}
\usage{
qcTalPez.lan(camp, dns = "Cant", lan, gr = 1)
}
\arguments{
\item{camp}{campaña a revisar los datos en formato Camp Xyy}

\item{dns}{Origen de bases de datos: Cant cantábrico, Pnew Porcupine}

\item{lan}{lance para el que se quiere comprobar que estan todas las especies medidas y talladas (sólo comprueba peces, por defecto gr=1, si se quiere otros cambiar gr)}

\item{gr}{El grupo que se quiere comprobar 1 peces, 2 crustaceos, 3 cefalópodos. 4, 5 y 6 nunca se miden.}
}
\value{
Devuelve la lista de especies capturadas pero no medidas o viceversa
}
\description{
Sirve para control de calidad y asegurarse que no faltan datos o estan incompletos
}
\examples{
qcTalPez.lan("C14",dns="Cant",24,gr=1)
}
\seealso{
Other Control.de.calidad: \code{\link{qcLW.camp}};
  \code{\link{qcLWbucl.camp}};
  \code{\link{qcdistlan.camp}}; \code{\link{qcdtall.camp}};
  \code{\link{qcdtallrev.camp}}
}

