% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SacaAphiaID.r
\name{SacaAphiaID}
\alias{SacaAphiaID}
\title{Función de búsqueda del código del grupo y la familia}
\usage{
SacaAphiaID(grupo, esp, dns = "Camp")
}
\arguments{
\item{grupo}{1=Peces, 2= crustáceos, 3 moluscos, 4 equinodermos, 5 otros invertebrados, 6 basuras sin AphiaID}

\item{esp}{Número que identifica la especie en el fichero especies.dbf del camp}

\item{dns}{origen de base de datos a utilizar en la busqueda, camp debería ser el que esté dirigido al del camp por defecto donde está el especies.dbf del camp, se puede crear uno de red.}
}
\description{
Busca el código el código AphiaID (WormS) a partir de los códigos del especies.dbf de CAMP
}
\examples{
buscaesp(1,87);SacaAphiaID(1,87)
}
