% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sefos.camp.R
\name{sefos.camp}
\alias{sefos.camp}
\title{Crea datos en formato SEFOS a partir de Camp, dns}
\usage{
sefos.camp(
  gr,
  esp,
  camp,
  dns = "Cant",
  plus = 8,
  cor.time = TRUE,
  AltAlk = NA,
  incl2 = FALSE
)
}
\arguments{
\item{gr}{Grupo de organismos, 1 peces 2 crustaceos, 3 moluscos, 4 equinodermos, 5 inv..}

\item{esp}{Código de especie camp}

\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)gr Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{plus}{Cual es la edad plus}

\item{cor.time}{Corrección del tiempo de arrastre al calcular las abundancias}

\item{AltAlk}{ALK alternativa tomada de un fichero de edad del camp edadxyy.dbf sin ruta ni extensión}

\item{incl2}{Si T se incluyen los lances extra no incluidos para las abundancias o biomasas estratificadas}
}
\value{
A data.table con el formato de datos de SEFOS para la evaluación de caballa
}
\description{
Function to complete HH with ICESrect and area
}
\examples{
# sefos.camp(1,74,"N23","Cant",plus=3)
}
