% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexr.camp.R
\name{sexr.camp}
\alias{sexr.camp}
\title{Sex ratio por talla de una especie}
\usage{
sexr.camp(
  gr,
  esp,
  camp,
  dns,
  cor.time = TRUE,
  excl.sect = NA,
  verbose = TRUE,
  plot = TRUE,
  es = F,
  sex.ratio = TRUE,
  sex.prop = FALSE,
  out.dat = FALSE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{verbose}{Si T muestra avisos problemas de tallas entre distintas especies}

\item{sex.ratio}{Si plot=T el gráfico será del sex ratio}

\item{sex.prop}{si plot=T el gráfico será de las proporciones por sexos}

\item{out.dat}{si T saca los datos de tallas por sexo}
}
\value{
Devuelve un data.frame con variables: talla, machos, hembras e indet(erminados) si existen todos y si sex=TRUE
}
\description{
Función de acceso a datos:
Extrae los datos de sex ratio de una especie a lo largo de la distribución de tallas muestreada
}
\examples{
dattal.camp("1"," 50",paste0("P0",7),"Porc",excl.sect=c("B","C"))
}
\seealso{
{\link{datos.camp}}
}
