% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.r
\name{stepper}
\alias{stepper}
\title{stepper}
\usage{
stepper(df, varVec, min.date, max.date)
}
\arguments{
\item{varVec}{A string vector containing the names of variables for which
step-expansion is required.  Typically set to
\code{c("bdMeanNightProp","bdMeanMoonProp","bdMeanForkLength")} for most
CAMP applications.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}
}
\value{
A dataframe with one row for all unique \code{batchDates} included 
  within \code{min.date} and \code{max.date}, inclusive.  All variables, as 
  provided via \code{varVec}, have a non-\code{NA} value.
}
\description{
For a set of covariates for which estimates have been obtained
  on non-continguous dates, expand recorded data values to fill in all dates,
  so as to create a step-like function.
}
\details{
The \code{stepper} function is usually only called within
  \code{eff_model} when applying previously obtained enhanced efficiency
  model results.
}
\examples{
\dontrun{
df <- stepper(df=df,
              varVec=c("bdMeanNightProp","bdMeanMoonProp","bdMeanForkLength"),
              min.date=min.date,
              max.date=max.date)
}
}
\seealso{
\code{eff_model}
}
\author{
WEST Inc.
}
