% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strmean.dt.R
\name{strmean.dt}
\alias{strmean.dt}
\title{Medias intraestrato y estratificadas por estratos geográfico y batimétrico}
\usage{
strmean.dt(x, sector, area, Nas = FALSE)
}
\arguments{
\item{x}{Es el vector de abundancias en peso o número}

\item{sector}{Es la asignación al sector de cada muestra igual que en strmean {\link{strmean.r}} pero ha de tener dos caracteres, primero el sector geográfico y segundo el estrato batimétrico}

\item{area}{El área del sector correspondiente}

\item{Nas}{Permite obtener desviación estandar cuando sólo existe un dato, dando una desviación standard de 0. Es util para obtener una estimacion de la variabilidad aunque sólo haya un lance en un estrato}
}
\value{
Devuelve la media estratificadda ponderada al área dentro de cada estrato y subestrato y la total del área
}
\description{
Función interna para cálculos y aplicación de bootstrap
}
