% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strmean.dtt}
\alias{strmean.dtt}
\title{Medias intraestrato y estratificadas por estratos geográfico y batimétrico}
\usage{
strmean.dtt(x, sector, area, w = rep(1, length(x)))
}
\arguments{
\item{x}{Es el vector de abundancias en peso o número}

\item{sector}{Es la asignación al sector de cada muestra igual que en strmean {\link{strmean.r}} pero ha de tener dos caracteres, primero el sector geográfico y segundo el estrato batimétrico}

\item{area}{El área del sector}

\item{w}{Es un factor de ponderación para cada uno de los valores de x. Sirve para la función de la librería boot boot(x,strmean,R,stype="f",Porc.map=sector,sector=sector,area=area)}
}
\value{
Devuelve la media estratificadda ponderada al área dentro de cada estrato y subestrato y la total del área
}
\description{
Función interna para cálculos y aplicación de bootstrap
}

