% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressBindingNotes.R
\name{suppressBindingNotes}
\alias{suppressBindingNotes}
\title{suppressBindingNotes}
\usage{
suppressBindingNotes(variablesMentionedInNotes)
}
\arguments{
\item{variablesMentionedInNotes}{A variable returned from
\code{devtools::check()} for which a NOTE is returned.}
}
\description{
Ensure resolution of NOTEs from the calling of package
  \code{devtools} function \code{check()}.
}
\details{
A helper function to assist in meeting R package devtools::check() 
  requirements.  Found via Stack Overflow post "In R, is it possible to 
  suppress 'Note: no visible binding for global variable?'"  This function is
  utilized in function \code{assignLSCompare} to wrap certain variables 
  utilized via function \code{dplyr}.
}
\examples{
#   ---- See the Stack Overflow post described in Details. 

}
\author{
Tyler Rinker
}
\keyword{internal}

