% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabAbsEsp.camp.r
\name{TabAbsEsp.camp}
\alias{TabAbsEsp.camp}
\title{Tabla resumen abundancias estratificadas por estrato y sector}
\usage{
TabAbsEsp.camp(gr, esp, camp, dns = "Cant", cor.time = TRUE, Nas = FALSE)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que se extraen los datos: un año concreto (XX): Demersales "NXX"}

\item{dns}{Elige el origen de las bases de datos: sólo sirve para el Cantábrico/Galicia "Cant"}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{Nas}{Permite calcular los errores estándar aunque sólo haya un lance en algún estrato (haciendo varianza =0 en ese estrato, incorrecto pero da una idea cuando sólo un estrato entre varios tiene sólo un lance)}
}
\value{
Devuelve un objeto en formato list con las abundancias y errores estándar de la especie elegida en la campaña elegida.
}
\description{
Salida de datos para informe de campaña del Cantábrico y Galicia (con 5 estratos y 3 sectores).
}
\examples{
TabAbsEsp.camp(1,36,"N08","Cant")
}
