% Generated by roxygen2 (4.0.2): do not edit by hand
\name{talpes.camp}
\alias{talpes.camp}
\title{Parámetros a y b relación talla-peso}
\usage{
talpes.camp(gr, esp)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}
}
\description{
Extrae los parámetros a y b de la relación talla-peso de una especie determinada
}

