% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryEnvCovDB.R
\name{tryEnvCovDB}
\alias{tryEnvCovDB}
\title{tryEnvCovDB}
\usage{
tryEnvCovDB(nTries, secSleep)
}
\arguments{
\item{nTries}{The number of attempts to query the Environmental Covariate
Database.}

\item{secSleep}{The number of seconds to wait between attempts.}
}
\value{
The function is nearly silent.  It simply reports messages to the log
  file indicating connection attempts.  In the case a connection is
  successful, a message of success is recorded.
}
\description{
Query the Environmental Covariate Database for other users prior
  to creating internal database tables.
}
\details{
The function is a simple \code{repeat}-type loop, running at most
  \code{nTries} times.
}
\author{
WEST, Inc.
}
