% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unid.camp.r
\name{unid.camp}
\alias{unid.camp}
\title{Unidades en que se mide una especie}
\usage{
unid.camp(gr, esp, dns = "Camp")
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{dns}{Es el ODBC donde se busca el fichero especies, deberías ser Camp por defecto pero se puede utilizar "Arsa" si se tiene el especies del Golfo de Cádiz añadiéndole el campo MED numérico y longitud 1}
}
\description{
Da información de las unidades en que se miden las especies y permite añadir datos a los gráficos
}
