% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VedaCigala.out.R
\docType{data}
\name{VedaCigala.out}
\alias{VedaCigala.out}
\title{Polígono área de veda}
\format{
A data.frame with long and lat points defining the closed area
}
\usage{
VedaCigala.out
}
\value{
data.frame con long y lat
}
\description{
VedaCigala.out: Zona ampliada de la zona de veda de Cigala en Porcupine
}
\details{
Polígono que delimita una zona ampliada más grande que la zona de veda de la cigala en el Porcupine Bank, la zona de veda fue propuesta por los armadores españoles, y puesta en vigor en 2010 desde el 1 de mayo al 31 de julio. Desde entonces ha estado vigente en diversos periodos, pero al menos todo el mes de mayo desde 2013 a 2016.
}
\examples{
mapporco();lines(lat~long,VedaCigala,col=2,lwd=2); lines(lat~long,VedaCigala.out,col="blue",lwd=2)
}
\references{
Medidas de gestión del stock de Cigala del banco de Porcupine FU 16 \url{http://www.nwwac.org/_fileupload/Opinions and Advice/Year 12/Dictamen_CCANOC_Gestion_Cigala_Porcupine_UF16_2017_ES.pdf}
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{VedaCigala}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{map.check}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other Porcupine: 
\code{\link{MapPorc}()},
\code{\link{Psh}},
\code{\link{VedaCigala}},
\code{\link{maparea}()},
\code{\link{mapporco}()}
}
\concept{Porcupine}
\concept{mapas}
\keyword{datasets}
