% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VedasBesugo.R
\docType{data}
\name{VedasBesugo}
\alias{VedasBesugo}
\title{Polígono áreas de veda besugo}
\format{
A data.frame with long and lat points defining the closed areas
}
\usage{
VedasBesugo
}
\value{
data.frame con lat y long
}
\description{
VedaBesugo: Veda Besugo Cantábrico y Galicia
}
\details{
Polígonos que delimitan las zonas de veda de juveniles de besugo en el Cantábrico, la zona de veda fue propuesta por los armadores españoles en 2019, sin llegar a entrar en vigor
}
\examples{
MapNort();lines(lat~long,VedasBesugo,col=2,lwd=2)
}
\references{
Propuesta de vedas protección juveniles besugo \url{https://www.mapa.gob.es/pesca/participacion-publica/detalle/nuevas_vedas_cantabrico}
}
\seealso{
Other mapas: 
\code{\link{BelgicaMound}},
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapHidro.camp}()},
\code{\link{MapHidro_b.camp}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{MatchHidro.ctd}()},
\code{\link{NepFUs.camp}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{hmmSAC}},
\code{\link{map.check}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphistal.one}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}
}
\concept{Cantábrico Galicia}
\concept{mapas}
\keyword{datasets}
