% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armap.camp.r
\name{armap.camp}
\alias{armap.camp}
\title{Mapa con las estaciones realizadas}
\usage{
armap.camp(
  camp,
  dns = "Porc",
  ti = FALSE,
  lwdl = 1,
  col = 2,
  argr = 2,
  cuadr = FALSE,
  cuadrMSFD = FALSE,
  ICESrect = FALSE,
  ICESlab = FALSE,
  arrow = FALSE,
  leg = TRUE,
  es = FALSE,
  cols = TRUE,
  noval = TRUE,
  CTDs = FALSE,
  strat = FALSE,
  Nlans = FALSE,
  NCTDs = FALSE,
  Dates = F,
  places = FALSE,
  bw = FALSE,
  lans = TRUE,
  xlims = c(-10.25, -1.4),
  ylims = c(41.82, 44.48)
)
}
\arguments{
\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX", Arsa otoño "2XX" y Medits "MXX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa", Mediterráneo "Medi"}

\item{ti}{si TRUE incluye un título con el nombre de la campaña}

\item{lwdl}{Anchura de las líneas}

\item{col}{Color de los puntos que representan estaciones}

\item{argr}{Grosor de las flechas de recorrido}

\item{cuadr}{Si T dibuja cuadrícula}

\item{cuadrMSFD}{Si T dibuja caudrícula de 10 millas naúticas utilizada para la evaluación de la estrategia marina (MSFD)}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{arrow}{Si T añade flechas de recorrido}

\item{leg}{Si T añade leyenda}

\item{es}{Si T nombres en castellano, si F nombres en inglés}

\item{cols}{Si T colorea los estratos y sectores, si F mapa en blanco y negro}

\item{noval}{Si T añade las estaciones nulas}

\item{CTDs}{Si T añade las estaciones de CTD realizadas}

\item{strat}{Si T delimita los sectores y los identifica}

\item{Nlans}{Si T identifica las estaciones numéricamente}

\item{NCTDs}{Si T identifica las estaciones de CTD numéricamente}

\item{Dates}{Si T saca las fechas en la que se han realizado los lances (dd-mm), no incluye información año.}

\item{places}{Si T añade los nombres de las principales ciudades}

\item{bw}{Gráfico en blanco y negro si T o en color si F}

\item{lans}{Si T marca las estaciones de muestreo con un punto (si Nlans=FALSE)}

\item{xlims}{Delimita la longitud del mapa mediante un vector (ejem.c(-10.25,-1.4))}

\item{ylims}{Delimita la latitud del mapa mediante un vector (ejem.c(41.82,44.48))}
}
\value{
Saca mapa con el desarrollo de la campaña con la zona y estratificación
}
\description{
Crea un mapa de un único año con las estaciones realizadas
}
\examples{
op<-par(no.readonly = TRUE)
par(mfrow=c(1,1))
armap.camp("N14",dns="Cant",noval=TRUE,CTDs=FALSE,bw=FALSE)
armap.camp("C14","Cant",xlims=c(-10.2,-7.6),es=TRUE)
par(mfrow=c(1,2),oma=c(0,0,2.5,0),mar=c(3,3,3,3))
armap.camp("P14","Porc",xlims=c(-10.2,-7.6),es=FALSE)
title("Trawl hauls",cex=1,line=2.5)
armap.camp("P14","Porc",xlims=c(-10.2,-7.6),lans=TRUE,es=FALSE,CTDs=TRUE)
title("CTD casts",cex=1,line=2.5)
mtext("Porcupine 2014 survey",line=1,outer=TRUE,cex=1.4,font=2)
par(op)
}
\seealso{
Other mapas: 
\code{\link{BelgicaMound}},
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapHidro.camp}()},
\code{\link{MapHidro_b.camp}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{MatchHidro.ctd}()},
\code{\link{NepFUs.camp}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.tot}()},
\code{\link{hmmSAC}},
\code{\link{map.check}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other resumen general: 
\code{\link{armap.tot}()}
}
\concept{mapas}
\concept{resumen general}
