% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armap.tot.r
\name{armap.tot}
\alias{armap.tot}
\title{Mapa ampliado de Porcupine con las estaciones realizadas}
\usage{
armap.tot(
  camp,
  dns = "Porc",
  ICESrect = FALSE,
  ICESlab = FALSE,
  lwdl = 1,
  col = 2,
  argr = 2,
  arrow = FALSE,
  leg = FALSE,
  es = FALSE,
  bw = FALSE,
  noval = FALSE,
  Nlans = FALSE,
  CTDs = FALSE,
  NCTDs = FALSE,
  Dates = FALSE,
  lans = TRUE,
  strat = FALSE
)
}
\arguments{
\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant" (ARSA para sacar datos al IBTS, no gráficos)}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{ICESlab}{si T incluye el rectángulo ICES en los rectángulos (arriba porcupine y norte, abajo en ARSA)}

\item{lwdl}{Anchura de las líneas}

\item{col}{Color de los puntos que representan estaciones}

\item{argr}{Grosor de las flechas de recorrido}

\item{arrow}{Si T añade flechas de reccorrido}

\item{leg}{Si T añade leyenda}

\item{es}{Si T nombres en castellano, si F nombres en inglés}

\item{bw}{Si T mapa en blanco y negro, si F colorea los estratos y sectores}

\item{noval}{Si T añade las estaciones nulas}

\item{CTDs}{Si T añade las estaciones de CTD realizadas}

\item{NCTDs}{Si T identifica las estaciones de CTD numéricamente}

\item{Dates}{Si T saca las fechas en la que se han realizado los lances (dd-mm), no incluye información año.}

\item{lans}{Si T marca las estaciones de muestreo con un punto (si Nlans=FALSE)}

\item{strat}{strat Si T represent los estratos}
}
\value{
Saca mapa con el desarrollo de la campaña con la zona y estratificación incluyendo tierra (Porcupine)
}
\description{
Crea un mapa de Porcupine con las estaciones realizadas(al igua que "armap.camp") pero con la referencia de Irlanda para su mejor ubicación.
}
\examples{
armap.tot("P14",dns="Porc",noval=TRUE,CTDs=FALSE,bw=TRUE,strat=FALSE,leg=TRUE)
armap.tot("N14",dns="Cant",noval=TRUE,CTDs=FALSE,bw=FALSE,leg=TRUE)
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{NepFUs.camp}()},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{map.check}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other resumen general: 
\code{\link{armap.camp}()}
}
\concept{mapas}
\concept{resumen general}
