% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/As2.R
\docType{data}
\name{As2}
\alias{As2}
\title{Serie histórica de la campaña Arsa de Otoño}
\format{A vector with codes for the surveys from 1997 to 2017}
\usage{
As2
}
\description{
Serie histórica Campañas ARSA Otoño
}
\details{
Serie completa con las campañas de otoño realizadas en el Golfo de Cádiz (Arrastre Sur Atlántica)
}
\seealso{
Other Series datos: \code{\link{As1}}, \code{\link{Msh}},
  \code{\link{Nsh}}, \code{\link{Psh}}

Other ARSA: \code{\link{As1}}, \code{\link{MapArsa}}
}
\keyword{datasets}
