% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscacod.r
\name{buscacod}
\alias{buscacod}
\title{Función de búsqueda del código del grupo y la familia}
\usage{
buscacod(nomb, dns = "Camp", buscfam = FALSE)
}
\arguments{
\item{nomb}{Nombre científico de la especie o fragmento del nombre entre ""}

\item{dns}{origen de base de datos a utilizar en la busqueda, camp debería ser el que esté dirigido al del camp por defecto donde está el especies.dbf del camp, se puede crear uno de red.}

\item{buscfam}{en vez de buscar el nombre de la especie busca a través de la familia, si T no muestra la lista de especies, sólo la familia, si F no seleccona por familia}
}
\description{
Busca el código del grupo, especie y la familia a partir del nombre o fragmento de la especie. Es independiente de mayúsculas y minúsculas y puede sacar más de una especie si una parte de sus nombres coincide.
}
\examples{
buscacod("sph")
}
\seealso{
Other datos_especies: 
\code{\link{AbrvEsp}()},
\code{\link{BuscaAphia}()},
\code{\link{buscaesp}()},
\code{\link{camptoyear}()},
\code{\link{hidrotodec}()},
\code{\link{talpes.camp}()}
}
\concept{datos_especies}
