% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CampsDNS.R, R/CampsDNS.camp.R
\name{CampsDNS.camp}
\alias{CampsDNS.camp}
\title{Listado de campañas presentes en el directorio e información sobre ellas}
\usage{
CampsDNS.camp(dns)

CampsDNS.camp(dns)
}
\arguments{
\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}
}
\value{
Un data.frame con columnas con la lista de archivos campXXX,lanceXXX,faunaXXX,ntallXXX e hidroXXX en el directorio, rellena los que falten con "."

Un data.frame con columnas con la lista de archivos campXXX,lanceXXX,faunaXXX,ntallXXX e hidroXXX en el directorio, rellena los que falten con "."
}
\description{
Función de acceso a datos:
A partir de ficheros de fauna.dbf presentes en directorio comprueba presencia de especie

Función de acceso a datos:
A partir de ficheros de fauna.dbf presentes en directorio comprueba presencia de especie
}
\examples{
CampsDNS.camp("Porc")
CampsDNS.camp("Porc")
}
