% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capwords.r
\name{capwords}
\alias{capwords}
\title{capwords}
\usage{
capwords(s, strict = FALSE)
}
\arguments{
\item{s}{A character vector of length one.}

\item{strict}{= FALSE If \code{TRUE}, force all letters of a word to 
lowercase after the first letter.  Default is \code{FALSE}.}
}
\value{
The character vector originally submitted via argument \code{s} with
  the first letter of each word capitalized.
}
\description{
Capitalize each word in a character vector of length one.
}
\examples{
#   ---- Capitalize each of 'hello' and 'there'.  
capwords("hello there")

#   ---- Use the strict argument.  Returns 'Camp'.
capwords("CAMP",strict=TRUE)
}
\author{
WEST Inc.
}
