% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chuck_zeros.R
\name{chuck.zeros}
\alias{chuck.zeros}
\title{chuck.zeros}
\usage{
chuck.zeros(b, e, beg.buff, end.buff, df)
}
\arguments{
\item{b}{describe argument}

\item{e}{describe argument}

\item{beg.buff}{describe argument}

\item{end.buff}{describe argument}

\item{df}{describe argument}
}
\value{
describe return value
}
\description{

}
\details{
other comments found in file
}
\examples{
#insert examples

}
\author{
WEST Inc.
}
\seealso{
\code{\link{related routine}}, \code{\link{related routine}}
}

