% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.bt.camp.R
\name{CV.bt.camp}
\alias{CV.bt.camp}
\title{Abundancias, SE y coeficientes de variación paramétricos calculados con bootstrap}
\usage{
CV.bt.camp(gr, esp, camp, dns, ind = "p", cor.time = TRUE, kg = TRUE,
  Rbt = 1000, excl.sect = NA, verbose = TRUE)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros. 9 incluye todos los grupos a excepción del 6}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que sacar datos año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX", Mediterráneo "MXX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew" o "Porc", Cantábrico "Cant" o "Cnew", Golfo de Cádiz "Arsa" y Mediterráneo "Medi"}

\item{ind}{El dato a procesar, si (p)eso o (n)umero}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{kg}{Si T saca los resultados en Kg, si F en gramos}

\item{Rbt}{número de iteraciones bootstrap para calcular los intervalos de confianza}

\item{excl.sect}{Excluye los sectores o subsectores dados como caracteres}

\item{verbose}{si T avisa distintas especies pueden ser de distintas características}
}
\value{
Devuelve el valor medio por bootstrap de la media, el error estándar y el CV
}
\description{
Función de resultados a partir de bootstrap con Rbt iteraciones conservando el diseño de muestreo
}
\seealso{
{\link{grafhistbox}}
}
