% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CVlist.camp}
\alias{CVlist.camp}
\title{Abundancias en número y peso totales para todas las especies de un grupo en una campaña}
\usage{
CVlist.camp(gr = "1", camp, dns, kg = T, dec = 3, percents = F,
  cor.time = T, excl.sect = NA)
}
\arguments{
\item{gr}{Grupo del que se quieren los datos: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{camp}{Campaña de la que se solicitan los resultados un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{kg}{Si T saca los resultados en Kg, si F en gramos}

\item{dec}{Número de decimales de los resultados}

\item{percents}{Si T devuelve los resultados en porcentaje del total}

\item{cor.time}{si T corrige abundancias a duración de lance de 30 minutos}

\item{excl.sect}{Excluye los sectores o subsectores dados como caracteres}
}
\value{
Devuelve un data.fram con valores en orden decreciente de abundancia con columnas gr,esp,especie,familia,camp,weight,number,nlan (número de lances en que aparece cada especie)
}
\description{
Función de resultados
}

