% Generated by roxygen2 (4.0.2): do not edit by hand
\name{datagegr.camp}
\alias{datagegr.camp}
\title{Extrae los datos del FAUNA de una especie en concreto en una edad.}
\usage{
datagegr.camp(gr, esp, camp, dns = "Pnew", plus = 8, mediahora = 1,
  cor.time = T)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustaceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Codigo de la especie numerico o caracter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña o campañas a representar en el mapa}

\item{dns}{Elige el origen de las bases de datos Porcupine "Pnew????" Cantábrico "Cant????" Golfo de Cadiz "Arsa" proporciona los datos para Medits pero no saca mapas}

\item{plus}{Edad plus: incluir la edad considerada como plus, solo afecta si se pide la edad solicitada que une todas las edades mayores}

\item{mediahora}{Valor para obtener abundancias por hora si media hora es mayor}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}
}
\value{
Devuelve un data.frame con campos lance, lat, long y abundancia por edad 0,1,2...Plus de edad, lance, peso número subestrato...
}
\description{
Función de acceso a datos: saca las abundancias por edad y lance de la especie
}

