% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datlan.camp.r
\name{datlan.camp}
\alias{datlan.camp}
\title{Características del lance}
\usage{
datlan.camp(
  camp,
  dns,
  incl2 = TRUE,
  incl0 = FALSE,
  hidro = FALSE,
  outhidro = FALSE,
  excl.sect = NA,
  redux = FALSE,
  year = T,
  quarter = T,
  bio = F
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa", combinados con "dnsred" busca los datos en el servidor de Santander si se han creado las RODBCs}

\item{incl2}{Si T se incluyen los lances extra no incluidos para las abundancias o biomasas estratificadas}

\item{incl0}{Si T se incluyen los lances nulos}

\item{hidro}{Si T muestra datos de hidrografía. Necesita que exista fichero base de datos de hidro: HIDROXYY.dbf}

\item{outhidro}{si T saca los datos del fichero hidro al final de todo el proceso como salida}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{redux}{Si T elimina datos de longitud y latitud de virada y muestra la media de las profundidades de largada y virada}

\item{year}{si T incluye una columna con el año al final de los datos}

\item{quarter}{si T incluye una columna con el trimestre de los datos teniendo en cuenta la fecha del lance, puede cambiar a mitad de la campaña, cuidado con campañas IBTS adscritas a un trimestre particular.}

\item{bio}{reduce el data.frame a los datos para los proyectos de biología, con datos en formato decimal y hexadecimal y las zonas ICES}
}
\value{
Devuelve un data.frame con datos de cada lance, las variables dependen de la selección de hidro y redux. En cualquier caso incluye variables weight.time con el factor de calibración para lances con menos tiempo del estándar y arsect: el área del sector al que corresponde el lance dentro del muestreo
}
\description{
Función de acceso a datos:
Extrae las características de los lances para una campaña determinada
}
\details{
Un problema que ocurre al utilizar el CampR con ficheros dbf de las primeras campañas
puede ser que al fichero lanceXXX.dbf le falte algún campo, habitualmente
el campo \strong{ESTN} utilizado en las últimas versiones del \strong{CAMP} para ligar lances con las estaciones de CTD.
El error usual es \strong{$ operator is invalid for atomic vectors}
Si se detecta este error revisar la estructura de lanceXXX.dbf con la de
otros ficheros de lances de los últimos años
}
\examples{
  print(datlan.camp(Nsh[24:28],"Cant",hidro=FALSE,excl.sect=c("A")))
  print(datlan.camp("P16","Porc",bio=T))
}
\seealso{
{\link{MapLansGPS}}
}
