% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datlan.camp.r
\name{datlan.camp}
\alias{datlan.camp}
\title{Características del lance}
\usage{
datlan.camp(camp, dns, incl2 = TRUE, incl0 = FALSE, hidro = FALSE,
  excl.sect = NA, redux = FALSE, year = T, quarter = T, bio = F)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)gr Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{incl2}{Si T se incluyen los lances extra no incluidos para las abundancias o biomasas estratificadas}

\item{incl0}{Si T se incluyen los lances nulos}

\item{hidro}{Si T muestra datos de hidrografía. Necesita que exista fichero base de datos de hidro: HIDROXYY.dbf}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{redux}{Si T elimina datos de longitud y latitud de virada y muestra la media de las profundidades de largada y virada}

\item{year}{si T incluye una columna con el año al final de los datos}

\item{quarter}{si T incluye una columna con el trimestre de los datos teniendo en cuenta la fecha del lance, puede cambiar a mitad de la campaña, cuidado con campañas IBTS adscritas a un trimestre particular.}

\item{bio}{reduce el data.frame a los datos para los proyectos de biología, con datos en formato decimal y hexadecimal y las zonas ICES}
}
\value{
Devuelve un data.frame con datos de cada lance, las variables dependen de la selección de hidro y redux. En cualquier caso incluye variables weight.time con el factor de calibración para lances con menos tiempo del estándar y arsect: el área del sector al que corresponde el lance dentro del muestreo
}
\description{
Función de acceso a datos:
Extrae las características de los lances para una campaña determinada
}
\examples{
  print(datlan.camp(Nsh[24:28],"Cant",hidro=FALSE,excl.sect=c("A")))
  print(datlan.camp("P16","Porc",bio=T))
}
\seealso{
{\link{MapLansGPS}}
}
