% Generated by roxygen2 (4.0.2): do not edit by hand
\name{datmuest.camp}
\alias{datmuest.camp}
\title{Datos del muestreo de la especie esp en la campaña}
\usage{
datmuest.camp(gr, esp, camp, dns = "Cant", excl.sect = NA)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}
}
\value{
Devuelve un list con nombre de la especie, número total medido (sexado si se sexa) del fichero de tallas  peso total capturado (del fichero de capturas en kg) y rango de tallas
}
\description{
Saca un resumen de los ejemplares y peso muestreados de la especie esp en la campaña camp
}

