% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtallan.camp.r
\name{dtallan.camp}
\alias{dtallan.camp}
\title{Valores absolutos en número por talla}
\usage{
dtallan.camp(
  gr,
  esp,
  camp,
  dns,
  lances = NA,
  cor.time = TRUE,
  depth_range = NA,
  sex = FALSE,
  muestr = TRUE,
  MeanLan = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{lances}{Da la opción de escribir un número de lance y saca los valores solo para ese lance o grupo de lances.}

\item{cor.time}{Corrige el tiempo del lance a 30 minutos si está fuera del tiempo (especialmente importante para 1983 y 1984)}

\item{depth_range}{NA por defecto, si no poner dos valores, min y max para establecer los límites batimétricos de los lances.}

\item{sex}{Por defecto (F) suma todos los individuos como indet. T saca los datos por sexo si los hay, no afecta si sólo hay indeterminados (3)}

\item{muestr}{Por defecto (T) no pondera los datos por el peso total en la captura del lance, si T coge los medidos realmente}

\item{MeanLan}{por defecto (T) saca la media por lance, dividiendo por el número de lances utilizados}
}
\value{
Devuelve un data frame con dos campos, $talla con las tallas de la especie en la campaña y número con el número de ejemplares de cada una de las tallas
}
\description{
Da Valores absolutos (totales por lance o grupos de lances) en número por talla y sexo (si lo hay)
}
\examples{
dtallan.camp(gr=1,esp=10,camp="N14",dns="Cant",lances=108,muestr=T)
dtallan.camp(gr=1,esp=10,camp="N14",dns="Cant",lances=108,muestr=F)
dtallan.camp(gr=1,esp=10,camp="N14",dns="Cant",lances=NA,muestr=F)
}
\seealso{
Other Distribuciones de tallas: 
\code{\link{denstall.camp}()},
\code{\link{dtall.campb}()},
\code{\link{dtall.camp}()},
\code{\link{dtall.lan}()},
\code{\link{dtallan.peso}()},
\code{\link{dtallbarplot}()}
}
\concept{Distribuciones de tallas}
