% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecolmatrix.camp.R
\name{ecolmatrix.camp}
\alias{ecolmatrix.camp}
\title{Extrae los datos del FAUNA de una especie en concreto.}
\usage{
ecolmatrix.camp(gr, esp = 999, camp, dns, ind = "p", incl6 = FALSE,
  codes = TRUE)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 para deshechos y otros. 9 incluye todos los grupos a excepción del 6 ver parámetro incl6 mas abajo}

\item{esp}{ha de ser 999 cuando se quiere incluir todas las especies del grupo, o elegir todas las especies deseadas con los codigos de las especies}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{ind}{Elige el valor (n)úmero o (p)eso que se utiliza en el resultado}

\item{incl6}{Si F se excluye el grupo 6 de otros (en general material no orgánico o antropogénico). Si T lo incluye dentro del grupo de "especies"}

\item{codes}{Si T da los códigos de las especies, si F el nombre científico}
}
\value{
Devuelve un data.frame con columnas lan, una con el peso o número de cada especie del grupo solicitado, lat, long, prof
}
\description{
Crea un data frame con campos lance, peso, numero, subestrato (en doble cifra), área subestrato  y posición geográfica. Usando códigos "9" "999" extrae los datos para el conjunto de todas las especies y grupos
}
\examples{
ecolmatrix.camp(1,999,"P01","Pnew",ind="p")
}
\seealso{
{\link{MapEcol.camp}}, {\link{ecolgr.camp}}
}
