% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edadstr.camp}
\alias{edadstr.camp}
\title{Calcula las abundancias estratificadas por edad}
\usage{
edadstr.camp(gr, esp, camp, dns = "Pnew", plus = 8, cor.time = T,
  AltAlk = NA)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camp}{Campaña de la que se extraen los datos un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{AltAlk}{ALK alternativa tomada de un fichero de edad del Camp edadXYY.dbf}
}
\description{
Función de resultados: abundancias estratificadas por edad para cada estrato batimétrico a partir de los datos del camp.
}
\examples{
edadstr.camp("1"," 45","P01","Pnew",8)
}
\seealso{
GetAlk.camp {\link{GetAlk.camp}}
}

