% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_catch_table.r
\name{F.allCatch.table}
\alias{F.allCatch.table}
\title{F.allCatch.table}
\usage{
F.allCatch.table(site, min.date, max.date, output.file)
}
\arguments{
\item{site}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}

\item{output.file}{<describe argument>}
}
\value{
<describe return value>
}
\description{
List all catches, regardless of species, etc. 

   Input:
   site = site ID of the place we want, trap locaton 

   Output:
   The take table is output to a CSV file


   Build table in Access file containing trap visits we want
}
\details{
<other comments found in file>
   ---- Write out the data set
   ---- Send messages back to the interface
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

