% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_catch_table.r
\name{F.allCatch.table}
\alias{F.allCatch.table}
\title{F.allCatch.table}
\usage{
F.allCatch.table(site, min.date, max.date, output.file)
}
\arguments{
\item{site}{The identification number of the site for which estimates are
required.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{output.file}{The name of the file prefix under which output is to be 
saved.}
}
\value{
A \code{csv} containing catch information for all caught species.  A
  data frame containing the same information is also returned, but invisibly.
}
\description{
Lists all catch, regardless of species.
}
\details{
This function utilizes the All-Catch Series query to find and
  identify all pertinent records within a catch.  See the Section Structure
  Query Language (SQL) Queries in \code{F.run.sqlFile} for more information.
}
\examples{
 
\dontrun{
#   ---- Obtain the catch over the specified period on 
#   ---- the American.
site <- 57000
min.date <- "2013-01-01"
max.date <- "2013-06-01"
output.file <- NA
df <- F.allCatch.table(site,min.date,max.date,output.file)
}
}
\seealso{
\code{F.chinookByDate.table}, \code{F.get.all.fish.data}
}

