% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_passage.r
\name{F.annual.passage}
\alias{F.annual.passage}
\title{F.annual.passage}
\usage{
F.annual.passage(site, taxon, run, min.date, max.date, output.file, ci = TRUE)
}
\arguments{
\item{site}{<describe argument>}

\item{taxon}{<describe argument>}

\item{run}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}

\item{output.file}{<describe argument>}

\item{ci=TRUE}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Estimate passage for all years in the data base.

   site = site number
   taxon = taxon number
   run = run number
   output.file = name of .CSV file to put passage estimates into.  Full path is allowed.

   Annual production is estimated for every year in the data base, from minimum date to maximum 
   date.  "years" are (min.date to min.date+365days), (min.date+365days, min.date+2(365)days), etc...


   ---- Pull the min and max dates from the data base
}
\details{
<other comments found in file>
   Open ODBC channel
   Pull all visits where fish were processed in order to determine min and max dates
   USE THIS CODE TO COMPUTE ESTIMATES FOR EVERY YEAR IN THE DATA BASE
sql.visit <- paste( "SELECT visitTime",  
    " FROM (", s.tab, " INNER JOIN ", ss.tab, " ON ", s.tab, ".siteID", " = ", ss.tab, ".siteID", ")", 
    " INNER JOIN ", tv.tab, " ON ", ss.tab, ".subsiteID", "=", tv.tab, ".trapPositionID", 
    " WHERE (((", s.tab, ".siteID)=", site, ") ",
    "AND ((", tv.tab, ".visitTypeID)=2 Or (", tv.tab, ".visitTypeID)=4) ",
    "AND ((", tv.tab, ".includeCatchID)=1 Or (", tv.tab, ".includeCatchID)>250)); ",
    sep="")

visit <- sqlQuery(ch, sql.visit)
F.sql.error.check(visit)

strt.date <- min( visit$visitTime )
end.date <- max( visit$visitTime )
   ---- Call F.passage, once for each year, to estimate passage.  This is done so that catch and efficiency models 
        are estimated within year only.
secs.in.year <- round(60*60*24*365.25)
yr.cuts <- seq( strt.date, end.date+secs.in.year, by=secs.in.year )
class(yr.cuts) <- class(strt.date)
       Loop over years
   ---- Plot results
   ---- Open PNG device
   ---- Write annual passage table to a file, if called for
   ---- Send messages back to the interface
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

