% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_batch_date.r
\name{F.assign.batch.date}
\alias{}
\alias{F.assign.batch.date}
\title{F.assign.batch.date - Assign batch dates}
\usage{
F.assign.batch.date(df)
}
\arguments{
\item{df}{The data frame containing trap visit records for which a day must be 
assigned.}
}
\value{
Function \code{F.assign.batch.date} maps trapping records containing 
  trapping-instance variable \code{EndTime}s to particular days, taking into
  consideration that trapping instances often take place over midnight, i.e.,
  two separate calendar days.  This function also considers multi-day
  trapping instances.
  
  The function works by utilizing a standardized time, via which trapping 
  periods are partitioned into particular days.  The cut time utilized to 
  assign minutes to a particular day is 4:00 AM Pacific Time (regardless of 
  Daylight Savings).  As of release 4.5, this value is set via global 
  variable \code{samplePeriodCutTime}.  The time of 4:00 AM was chosen, 
  rather than midnight, in order to not prematurely cut-off trapping 
  instances that often continue past midnight.
  
  Generally, the variable created by function \code{F.assign.batch.date}, 
  \code{batchDate}, is the same as variable \code{SampleDate} found in 
  query-derived table \code{TempSumUnmarkedByTrap_Run_Final} in any CAMP 
  Database.  Variable \code{batchDate} replaces variable \code{SampleDate} in
  all dataframes for which function \code{F.assign.batch.date} is called.  To
  avoid confusion, new variable \code{batchDate} replaces old variable 
  \code{SampleDate} after the running of this function.
}
\description{
Assigns batch dates to all trap visis in a data frame based on
\code{samplePeriodCutTime}.
}
\details{
In order to process the sequence of dates correctly, 
function \code{F.assign.batch.date} first identifies the earliest and latest 
dates in the provided data frame's trapping \code{EndTime} variable.  Given 
\code{EndTime} dates and times, it then moves back one day in time, and two 
forward; this encapsulates the earliest and latest \code{EndTime} dates. 
Given the unique dates within this resulting range, each of which start at 
12:00 midnight, it then adds the value of global variable 
\code{samplePeriodCutTime};  this effectively moves the traditional 
day-separation time of midnight to the value of \code{samplePeriodCutTime}, 
or 4:00 AM as of release 4.5, for the purposes of assigning fish-counts to
particular days.  See Examples.  

All calculations consider daylight savings time and assume that measurements 
take place within the Pacific Time Zone.
}
\examples{
# Set the time zone to Pacific.
time.zone <- "America/Los_Angeles"

# Create a data set of three trapping instances.  
origDate <- data.frame(trapVisitID=c(1234,1235),
SampleDate=strptime(c("2014-01-24","2014-01-24"),"\%F"),
StartTime=strptime(c("2014-01-23 14:28:00","2014-01-24 03:59:00"),"\%Y-\%m-\%d \%H:\%M:\%S", tz=time.zone),
EndTime=strptime(c("2014-01-24 03:59:00","2014-01-24 04:00:00"),"\%Y-\%m-\%d \%H:\%M:\%S", tz=time.zone))

# BatchDate for \\code{trapVisitID=1234} is now 1/23, due to that trap stopping before 4:00 AM.  
newDate <- F.assign.batch.date(origDate)
}
\seealso{

}

