% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_batch_date.r
\name{F.assign.batch.date}
\alias{F.assign.batch.date}
\title{F.assign.batch.date}
\usage{
F.assign.batch.date(df)
}
\arguments{
\item{df}{The data frame containing trap visit records for which a day must be 
assigned.}
}
\value{
The same data frame \code{df} submitted to the function, with new
variable \code{batchDate} replacing old variable \code{SampleDate}.
}
\description{
Assigns batch dates to all trap visits in a data frame via global
  variable \code{samplePeriodCutTime}.
}
\details{
In order to process the sequence of dates correctly, function 
  \code{F.assign.batch.date} first identifies the earliest and latest dates 
  in the provided data frame's trapping \code{EndTime} variable.  Given 
  \code{EndTime} dates and times, it then moves back one day in time, and two
  forward; this encapsulates the earliest and latest \code{EndTime} dates. 
  Given the unique dates within this resulting range, each of which start at 
  12:00 midnight, it then adds the value of global variable 
  \code{samplePeriodCutTime}, or 4:00 AM.  The time of 4:00 AM was chosen so 
  as to not prematurely cut-off trapping instances that often continue past 
  midnight. This effectively moves the traditional day-separation time of 
  midnight to 4:00 AM.
  
  Function \code{F.assign.batch.date} maps trapping records containing 
  trapping-instance variable \code{EndTime}s to particular days, taking into 
  consideration that trapping instances often take place over midnight, i.e.,
  two separate calendar days.  This function also considers multi-day 
  trapping instances.  All calculations consider daylight savings time and 
  assume that measurements take place within the Pacific Time Zone.
  
  Generally, the variable created by function \code{F.assign.batch.date}, 
  \code{batchDate}, is the same as variable \code{SampleDate} found in 
  query-derived table \code{TempSumUnmarkedByTrap_Run_Final} in any CAMP 
  Database.  See section Structured Query Language (SQL) Queries in
  \code{F.run.sqlFile}. Variable \code{batchDate} replaces variable
  \code{SampleDate} in all dataframes for which function
  \code{F.assign.batch.date} is called.
}
\examples{
\dontrun{
#   ---- Set the time zone to Pacific.
time.zone <- "America/Los_Angeles"

#   ---- Create a data set of three trapping instances.  
origDate <- data.frame(trapVisitID=c(1234,1235),
SampleDate <- strptime(c("2014-01-24","2014-01-24"),"\%F"),
StartTime <- strptime(c("2014-01-23 14:28:00","2014-01-24 03:59:00"),
  "\%Y-\%m-\%d \%H:\%M:\%S", tz=time.zone),
EndTime <- strptime(c("2014-01-24 03:59:00","2014-01-24 04:00:00"),
  "\%Y-\%m-\%d \%H:\%M:\%S", tz=time.zone))

#   ---- BatchDate for trapVisitID=1234 is now 1/23, 
#   ---- since that trap stopped before 4 AM.  
newDate <- F.assign.batch.date(origDate)
}
}
\author{
WEST Inc.
}
