% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Report_Criteria.r
\name{F.buildReportCriteria}
\alias{F.buildReportCriteria}
\title{F.buildReportCriteria}
\usage{
F.buildReportCriteria(site, min.date, max.date)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}
}
\value{
Within R, function \code{build_Report_Criteria} returns a data frame 
  containing one row and row column of the total number of visits at a site 
  between the specified \code{min.date} and \code{max.date}.  Within Access, 
  via the \code{RODBC} package, it creates table \code{TempReportCriteria_Trapvisit}, via query series  
  Build Report Criteria.
}
\description{
Update Access database table \code{TempReportCriteria_Trapvisit}
  to contain all unique trapping instances for the site and inclusive
  calendar dates specified by the user.
}
\details{
Function \code{build_Report_Criteria} is the
workhorse function associated with many query series, and sets up data
within the Access CAMP database for further processing.  See function 
\code{F.run.sqlFile} for more details on query series.
}
\examples{
\dontrun{
#   ---- American River at Watt Avenue, 2013 Season
site <- 57000                 
min.date <- "2013-01-16"
max.date <- "2013-06-08" 

#   ---- Obtain inclusive trap visits.  
nvisits <- F.buildReportCriteria( site, min.date, max.date )
}
}
\author{
WEST Inc.
}
\seealso{
\code{sqlQuery}, \code{F.run.sqlFile}, \code{F.sql.error.check}
}

