% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Report_Criteria.r
\name{F.buildReportCriteria}
\alias{F.buildReportCriteria}
\title{F.buildReportCriteria}
\usage{
F.buildReportCriteria(site, min.date, max.date)
}
\arguments{
\item{site}{The \code{siteID} for which data are requested.}

\item{min.date}{The start date for which data are requested.}

\item{max.date}{The end date for which data are requested.}
}
\value{
A data frame 
  containing one row and one column of the total number of visits at \code{siteID}  
  between \code{min.date} and \code{max.date}.
}
\description{
Update a table in the Access CAMP database to contain a particular 
  site's trapVisitIDs between two dates specified by the user.
}
\section{Side effect}{
 The query run by this function, 
\code{QryBuildReportCriteria.sql}, constructs a temporary table 
in the CAMP Access file which is later used by other tables and queries
(i.e., \code{TempReportCriteria_Trapvisit}). 
Many series of queries start and data extractions start with this one.
}
\examples{

}
\seealso{
\code{sqlQuery}, \code{F.run.sqlFile}, \code{F.sql.error.check}
}

