% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Report_Criteria_Release.r
\name{F.buildReportCriteriaRelease}
\alias{F.buildReportCriteriaRelease}
\title{F.buildReportCriteriaRelease}
\usage{
F.buildReportCriteriaRelease(site, min.date, max.date)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}
}
\value{
Within R, the total integer count of efficiency trials.  Within
  Access, table \code{blah} is created, itemizing the releases.
}
\description{
Update a table in the underlying Access database to contain release
trapVisitIDs and count the number of efficiency trials.
}
\details{
Unique efficiency trials are identified by unique \code{releaseIDs}
  falling within the specified \code{min.date} and \code{max.date}.
  
  Note that the difference between this and the other
  \code{buildReportCriteria} query sequence is this one's focus on releases
  and subsequent recaptures.
}
\examples{

}

