% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_catch_table.r
\name{F.byCatch.table}
\alias{F.byCatch.table}
\title{F.byCatch.table}
\usage{
F.byCatch.table(site, min.date, max.date, output.file)
}
\arguments{
\item{site}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}

\item{output.file}{<describe argument>}
}
\value{
<describe return value>
}
\description{
List all non-chinook catches,

   Input:
   site = site ID of the place we want, trap locaton

   Output:
   The take table is output to a CSV file


   Build table in Access file containing trap visits we want
}
\details{
<other comments found in file>
   ---- Write out the data set
   ---- Produce a pie chart and write out the data for it
   ---- Open PNG device
   ---- Send messages back to the interface
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

