% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_model.r
\name{F.catch.model}
\alias{F.catch.model}
\alias{catch.model}
\alias{catchModel}
\alias{catchmodel}
\title{Estimate a catch regression model.}
\usage{
F.catch.model(catch.df)
}
\arguments{
\item{catch.df}{A data frame containing all requested trapping data for a
particular trapPosition.  See 'Details'.}
}
\value{
Function \code{F.catch.model} serves two main purposes.  The first utilizes
  cubic splines to estimate total catch per trapping position over the date
  range provided.  The second utilizes spline results to impute estimates for
  periods when the trap was not functioning.
}
\details{
\bold{Splines}: Prior to model fitting, \code{trapVisitIDs} are identified
as possible night-time trapping instances via examination of StartTimes and
EndTimes along with calculated sunrises and sunsets.  As of the 1.5 release,
all sunrises and sunsets are set to 7:00 AM and 7:00 PM local time.  Given
  daylight and trapping times, all trapVisitIDs with some portion of trapping
  taking place at night are identified, along with the overall proportion of a
  trapping instance that takes place at night.

  Following computation of night-time variables, a null Poisson model of
  total catch is fit, via a long link, with an indicator variable for
  night-time trapping included if at least 90% of all the trapping instances
  trapped at night, as defined above;  otherwise, only an intercept term is
  included.  These models include log-transformed trap sampling time, in
  hours, as an offset.  All models fit via glm.

  After the initial intercept-only model, increasingly complex models are
  fit.  More complex models require at least ten caught fish, over any
  combination of trapping instances.  Prior to a cubic spline, each of a
  polynomial linear, quadratic, and cubic is fit to the data.  (Note that a
  cubic polynomial model is a cubic spline with no internal knots.) Each
  incorporates a night variable if at least 90% of the trapping instances
  fished at night.  Models only consider the next complex model if two
  conditions are met.  First ,the difference in the Akaike Information
  Criterion (AIC), when comparing the current model to the previous model,
  must be greater than two, after rounding both models to four decimals.
  Second, the number of unique trapping instances, divided by 15, rounded
  down, must be greater than or equal to the model's degrees of freedom,
  excluding an intercept.  This means that for a linear model, at least 15
  data points are required.  Similar logic requires 30 for a quadratic.  The
  table below summarizes the relationship between the number of data points (or trapping instances) and maximal
  model possible.

 \tabular{cll}{
  \bold{Number of}    \tab                           \tab \bold{Number of} \cr
  \bold{Coefficients} \tab \bold{Model Type}         \tab \bold{Efficiency Trials} \cr
        ------------  \tab ---------------------------------------- \tab -------------------- \cr
   0   \tab intercept-only                           \tab 15 \cr
   1   \tab linear                                   \tab 30 \cr
   2   \tab quadratic                                \tab 45 \cr
   3   \tab cubic                                    \tab 60 \cr
   4   \tab cubic spline with one internal knot      \tab 75 \cr
   5   \tab cubis spline with two internal knots     \tab 90 \cr
   ... \tab ...                                      \tab ... \cr
   k                  \tab cubic spline with (k - 3) internal knots \tab 15*(k - 1) \cr
 }

   When required, cubic splines utilize function bs to obtain an appropriate
   basis for use in model fitting. The model degrees of freedom determines
   the number of internal knots utilized, via parameter df.  Individual cubic splines include
   no intercept, although all cubic-spline Poisson catch models
   utilizes an overall intercept, so as to center models along the outcome axis.

\bold{Imputation}: <describe imputation>
}
\section{}{
  Imputation:  The trap-specific imputation procedure utilizes the final catch spline result obtained
via the process described above.  Specifically, it sweeps through all temporally sorted rows of the catch dataframe for the
trap of interest, replacing all instances of "Not fishing," which spline-estimated fish.  All estimates loop
over gaps one at a time, predicting catch for a maximum of up to 24 hours.  All "Not fishing" gaps estimate on
hours, in tandem with the temporal unit utilized in Poisson model offsets.  For example, for a 56-hour gap, predictions
occur for each 24-hour period, with catch estimated proportionally for any "leftover" preceding and antecedent times.  See examples.

Catch is not estimated for individual "Not fishing" episodes of duration less than two hours.  In these cases, the sampleMinutes
associated with
}
\examples{
F.catch.model( catch.df ) 

}
\seealso{
blah blah blah
}

