% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinook_by_date.r
\name{F.chinookByDate.table}
\alias{F.chinookByDate.table}
\title{F.chinookByDate.table - Retrieve data frame of chinnook catches}
\usage{
F.chinookByDate.table(site, min.date, max.date, output.file)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{output.file}{The name of the file prefix under which output is to be 
saved.}
}
\value{
A \code{csv} containing Chinook salmon catch information.  A data
  frame containing the same information is also returned, but invisibly.
}
\description{
Summarize chinook salmon by date between two dates.
}
\details{
As its name implies, function \code{F.chinookByDate.table} retrieves
  only Chinook salmon records, as specified by setting variable 
  \code{taxon=161980}.  The function utilizes helper function 
  \code{F.get.all.catch.data} to handle the sum-chinook-by-trap query series.
  Note that function \code{F.get.catch.data}, utilized in the estimation of 
  passage, differs from \code{F.get.all.catch.data}.
}
\examples{

# requires an mdb.

}
\seealso{
F.get.catch.data.
}

