% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_model.r
\name{F.efficiency.model}
\alias{F.efficiency.model}
\title{F.efficiency.model}
\usage{
F.efficiency.model(obs.eff.df, plot = T, method = 1, max.df.spline = 4,
  plot.file = NA)
}
\arguments{
\item{obs.eff.df}{A data frame with at least variables \code{batchDate} and 
\code{efficiency}, where \code{efficiency} is \code{NA} for all days 
requiring an estimate.}

\item{plot}{A logical indicating if resulting efficiency should be plotted.}

\item{method}{A scalar specifying the type of extrapolation to do. 
\code{method = 1} takes average of entire season. \code{method = 2} uses 
earliest observed efficiency for each unique interval between distinct
efficiency trials; i.e., a so called "step-function" constant model. 
\code{method = 3} is a B-spline model with up to \code{max.df.spline}
degrees of freedom.}

\item{max.df.spline}{The maximum degrees of freedom allowed for splines.}

\item{plot.file}{The name of the file prefix under which output is to be 
saved.  Set to NA to plot to the Plot window.}
}
\value{
A data frame with all observed and imputed \code{efficiency} values, 
  where variable \code{gam.estimated} identifies days with imputed values.
}
\description{
Compute and estimate efficiency for all missing combinations of
  traps and days;  i.e., "impute" a value for efficiency when missing.
}
\section{Efficiency Methodologies}{


Selection of \code{method} allows for efficiency estimation to vary, based on
need.
  
  \itemize{ 
  \item{\code{method=1} : A "Seasonal-mean model".  This means that a
  ratio-of-means bias-corrected global efficiency is calculated, for each 
  trap, where the efficiency is estimated via \deqn{\frac{nCaught
  + 1}{nReleased + 1}}{(nCaught + 1) / (nReleased + 1).}  Values
  for variables \code{nCaught} and \code{nReleased} originate via function
  \code{F.get.releases} and the querying of an underlying Access database. 
  The resulting ratio-of-means efficiency is then applied to all trapping
  days, regardless if data were missing or not.}
  
  \item{\code{method=2} : A "Constant model."  This means that this model
  utilizes a step function to estimate efficiency, where here, steps may
  increase or decrease over time.  Each interval of time for which efficiecy
  needs to be estimated utilizes the earliest observed efficiency between
  efficiency trials.  Note that this method has not been fully programmed to
  be compatible with boostrapping in function \code{F.bootstrap.passage}.  This
  option should not be used until \code{F.bootstrap.passage} can handle its
  results.}
  
  \item{\code{method=3} : A "B-spline model."  This efficiency estimation 
  methodology chops up the date range into small temporal bits.  For each,
  possibly different degree polynomials are fit, with each individual
  polynomial fitting the data locally.  Each individual polynomial is then
  fit together into one final curve, with several conditions ensuring its
  smoothness. One condition includes matching the polynomial estimate at the
  end of one time point with the connecting starting time point of the next
  polynomial.  Others include ensuring that the first derivatives (local
  slopes) and second derivatives (local convexity) at these connecting time
  points are equal.  
  
  Estimation requires at least one trapping instance to occur within the 
  earliest and latest date with an efficiency trial.  All three can occur on 
  the same one day.  Generalized additive models, or GAMs, first fit the
  observed efficiency-trial data as
  \deqn{\frac{nCaught}{nReleased}}{nCaught / nReleased} outcome
  proportions against a null model with a binomial link.  This leads to a
  constant, or intercept-only model, where the intercept is the weighted 
  avearge of caught fish over all efficiency trials.
  
  Assuming at least ten efficiency trials, higher-order polynomial models are
  considered.  Akaike Information Criterion (AIC), rounded to four decimal
  places, determines winning models. Models with lower AICs trump all others,
  with the winning model having to be at least two AIC points less than its
  nearest competitor.  Function \code{bs} identifies the B-spline basis to
  use in higher-order models.  
  
  In generalized additive models, consideration of higher-order polynomials 
  often starts with cubic polynomials partitioned in a temporal piecewise
  fashion, chopped at the time values at which observed efficiencies were
  recorded. Currently, however, no polynomials of degree greater than three
  are considered in efficiency models;  thus, models could either be 
  intercept-only, or cubic.
  
  The \eqn{\beta}s from the final selected GAM are saved for use in
  bootstrapping via function \code{F.boostrap.passage}.  They are also 
  utilized to form predictions over the entire efficiency trial season, for
  each trap.  Model efficiency values are used for all days, i.e., in lieu of
  observed data.  All dates outside the efficiency trial season use an
  estimated mean spline calculated over all dates.}
  
  \item{\code{method=4} : An "Enhanced Efficiency model."  To be developed
  at a later date.}
  
  }
}
\examples{
\dontrun{
#   ---- Fit an efficiency model for each unique trapPositionID 
#   ---- in data frame obs.eff.df.  
F.efficiency.model( obs.eff.df, plot=T, method=1, max.df.spline=4, plot.file=NA)
}
}
\author{
Trent McDonald (tmcdonald@west-inc.com)
}
\seealso{
\code{F.get.releases}, \code{F.bootstrap.passage}
}

