% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_efficiency.enh.r
\name{F.est.efficiency.enh}
\alias{F.est.efficiency.enh}
\title{F.est.efficiency.enh}
\usage{
F.est.efficiency.enh(release.df, batchDate, df.spline = 4, plot = TRUE,
  plot.file = NA)
}
\arguments{
\item{release.df}{A data frame produced by \code{F.get.release.data}. 
Contains information on releases and recaptures.  This data frame has one 
line per release trial per trap, with trap identified via variable 
\code{TrapPositionID}.}

\item{batchDate}{A POSIX-formatted vector of dates.}

\item{df.spline}{The default degrees of freedom to use in the estimation of 
splines. Default is 4 (1 internal knot).}

\item{plot}{A logical indicating if efficiencies are to be plotted over time,
per trap.}

\item{plot.file}{The name to which a graph of efficiency is to be output, if 
\code{plot=TRUE}.}
}
\value{
Results from running the enhanced efficiency model fitting process 
  include \code{csv}s of efficiency trials missing a covariate, for each
  \code{TrapPositionID} at a \code{subSiteID}.  This prevents these
  efficiency trials from inclusion in the model fitting process.
  
  Each \code{TrapPositionID} also results in a series of plots depicting the
  fitted temporal spline, at each point of the backwards-fitting process. 
  These could number many, depending on the number of covariates available
  for possible exlcusion.
  
  Each \code{trapPositionID} also outputs a \code{png} containing model
  fitting information, including plots of efficiency versus each considered
  covariate, along with plotted temporal trends of each covariate against
  time.  Additional plots include the final fitted temporal spline (along
  with an prediction "curve" derived from the available data), as well as a
  final "plot" depicting model summary statistics, obtained via the
  \code{summary} function against the logistic efficiency-trial model fits.
  
  Note that no passage estimates result from the fitting of enhanced
  efficiency models.  This is because bootstrapping does not occur, but also
  because estimation of passage is not the goal of the model fitting process.
  Use the regular function sequence; i.e., functions without the
  \code{".enh"} to estimation passage for one-year intervals of interest.
}
\description{
Estimate trap efficiency for every sample period, per trap.
}
\examples{
\dontrun{
#   ---- Estimate the efficiency.  
theEff <- F.est.efficiency(release.df,batchDate,df.spline=4,plot=TRUE,plots.file=NA)
}
}
\seealso{
\code{run_passage.enh}, \code{get_release_data},
  \code{est_passage.enh}, \code{est_efficiency.enh}, \code{eff_model.enh}
}
\author{
WEST Inc.
}
