% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_efficiency.r
\name{F.est.efficiency}
\alias{F.est.efficiency}
\title{F.est.efficiency}
\usage{
F.est.efficiency(release.df, batchDate, method = 1, df.spline = 4,
  plot = TRUE, plot.file = NA)
}
\arguments{
\item{release.df}{<describe argument>}

\item{batchDate}{<describe argument>}

\item{method=1}{<describe argument>}

\item{df.spline=4}{<describe argument>}

\item{plot=TRUE}{<describe argument>}

\item{plot.file=NA}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Estimate trap efficiency for every sample period. 

   Input:
   release.df = data frame produced by F.get.release.data().  Contains
       information on releases, and recaptures. This data frame has one line per release trial 
       per trap (TrapPositionID).

   Note: Run season is a vector length 2 of dates for beginning and 
   ending of run, stored as an attribut of the data frame.

   Output:
   A data frame containing interval and capture efficiency and
   $gam.estimated.  $gam.estimated is 'Yes' if efficiency for that interval was estimated
   by the GAM model, rather than being empirical.

    batchDate <- bd
      method <- 3
      df.spline <- 3
      plot.file <- file.root


   ---- Check that we actually caught released fish.  If not, 0 efficiency, cannot do estimate.
}
\details{
<other comments found in file>
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

