% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_catch_data.r
\name{F.get.all.catch.data}
\alias{F.get.all.catch.data}
\title{F.get.all.catch.data}
\usage{
F.get.all.catch.data(site, taxon, min.date, max.date)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}
}
\value{
A data frame entitled \code{visit}.  To be included, a record has to
  be from the correct \code{site} and \code{taxon}, and be between the values
  specified by \code{min.date} and \code{max.date}.
}
\description{
Fetch catch data for a single taxon from an Access data base.
}
\details{
Function \code{F.get.all.catch.data} is called from within function
\code{F.all.catch.table}, and helps
}
\examples{
\dontrun{

# requires an mdb.
}
}
\seealso{
\code{F.all.catch.table}
}

