% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_fish_data.r
\name{F.get.all.fish.data}
\alias{F.get.all.fish.data}
\title{F.get.all.fish.data}
\usage{
F.get.all.fish.data(site, min.date, max.date)
}
\arguments{
\item{site}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Fetch ALL the fish data from Access data base between two dated. 
   This is a version of F.get.indiv.fish.data, but that routine returns records for a single taxon.

   input:
   site = site ID of place we want to do estimates for. Note; Site is a location along a river where
       we can compute passage.  Subsites are places in the river (thalweg, right bank, etc) at the project.
       Project is a parent of subsite.
       Site is also a parent of subsite.
       Project is the funding source of the site.  Site and Project are parallel fields.  In theory,
       this makes it easier query sites and projects because sometimes the same site is sampled by
       two different projects.
   min.date and max.date = minimum and maximum dates for a visit to be included. A string in format "YYYY-MM-DD"



   ---NOTE: build_report_Criteria must be run before here. 


   *******
   Open ODBC channel
}
\details{
<other comments found in file>
NA
   Subsampling is handled above in the SQL with [CatchRaw].[n]*[CatchRaw].[subsampleDenominator]/[CatchRaw].[subsampleNumerator]
}
\examples{
\dontrun{
<insert examples>

}
}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

