% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_fish_data.r
\name{F.get.all.fish.data}
\alias{F.get.all.fish.data}
\title{F.get.all.fish.data}
\usage{
F.get.all.fish.data(site, min.date, max.date)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}
}
\value{
A data frame containing all catch, regardless of taxon, between the 
  specified dates.
}
\description{
Fetch all fish data, regardless of taxon, from an Access data 
  base between two dates.
}
\details{
This is a generalization of \code{F.get.indiv.fish.data} which
  returns records for a single taxon.  Function \code{F.get.all.fish.data}
  utilizes the All Catch query series in order to obtain the proper data.
  See section Structured Query Language (SQL) Queries in function 
  \code{F.run.sqlFile}.
}
\examples{
\dontrun{
#   ---- Obtain all fish on the American between the specified dates. 
site <- 57000
min.date <- "2013-01-01"
max.date <- "2013-06-01" 
F.get.all.fish.data(site,min.date,max.date)
}
}
\author{
WEST Inc.
}
\seealso{
\code{F.get.indiv.fish.data}
}

