% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_release_data.r
\name{F.get.release.data}
\alias{F.get.release.data}
\title{F.get.release.data}
\usage{
F.get.release.data(site, taxon, min.date, max.date)
}
\arguments{
\item{site}{The identification number of the site for which estimates are
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{min.date}{The start date for data to include. This is a text string in
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as
\code{min.date}.}
}
\value{
A data frame summarizing efficiency trials for the site of interest
  for all traps between the dates indicated.  Data include \code{Recaps}
  numerators and \code{nReleased} denominators.
}
\description{
Fetch data on efficiency trials from an Access database.
}
\details{
Function \code{F.get.release.data} utilizes query sequences Build 
  Report Criteria and Build Report Criteria Release to obtain all results 
  within the specified \code{min.date} and \code{max.date}.  See section
  Structured Query Language (SQL) Queries in function 
  \code{F.run.sqlFile} for more details.
  
  Query results include one record for every trap visit within a 
  \code{releaseTime}, say 7 days, even if the trap visit did not catch any 
  marked fish.  In this case, zeros are recorded for all combinations of 
  \code{releaseID}, \code{trapPositionID}, and \code{trapVisitID}.
  
  Given a specific release, the resulting data frame tells how many fish from
  the release were captured on subsequent trap visits, for each trap.  These 
  result from collapsing all trap visits and computing the total number of 
  each release's captured fish.  Note that generally, at any one time, more
  than one release can "go," and so a single trap visit may catch fish from
  multiple releases.  Total recaptures are summarized over unique
  combinations of trap visits and positions, via variables \code{releaseID}
  and \code{trapPositionID}.
  
  Release records need to have both variables \code{IncludeTest} and 
  \code{IncludeCatch} flagged as \code{"Yes"} for inclusion in efficiency 
  estimation. Recaptures that took place during half-cone operations are 
  multiplied by the value of the \code{halfConeMulti} global variable, which is set 
  at 2. Half cone operations are identified by variable \code{HalfCone} 
  having a value of \code{"Yes"}.
  
  Variables \code{HrsToFirstVisitAfter} and \code{HrsToLastVisitAfter} are 
  used in function \code{F.est.efficiency} as helper variables to derive a 
  batch date when the \code{meanEndTime} variable is \code{NA}.
}
\section{Mean Recapture Time}{
 The mean recapture time is estimated for each 
  unique grouping of \code{releaseID} and \code{trapPositionID}.  In the case
  of no recaptures, the mean recapture time is recorded as \code{NA}.  In all
  other cases, the mean recapture time is calculated as the weighted mean of 
  recapture times, weighting on the number of caught fish.  For example, 
  suppose fishing takes place at a particular trap over 7 consecutive days. 
  If the bulk of fish were caught on day 7, then the mean recapture time 
  would be near that 7th day of fishing.  This is in contrast to a "straight"
  mean recapture time, which would estimate a mean recapture time 3.5 days
  after the start of fishing, regardless of the temporal distribution of 
  captured fish over the entire 7-day period.
}
\examples{
\dontrun{
#   ---- Fetch all Chinook salmon efficiency data on the American River 
#   ---- between Jan. 16, 2013 and June 8th, 2013.  
site <- 57000
taxon <- 161980
min.date <- "2013-01-01"
max.date <- "2013-06-01"
df <- F.get.release.data(site,taxon,min.date,max.date)
}
}
\author{
WEST Inc.
}
\seealso{
\code{F.run.sqlFile}, \code{F.summarize.releases}, \code{F.release.summary}
}

