% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_by_catch.r
\name{F.getByCatch}
\alias{F.getByCatch}
\title{F.getByCatch}
\usage{
F.getByCatch(site, min.date, max.date)
}
\arguments{
\item{site}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Retreive all non-Chinook data between two dates at site.


   ---NOTE: build_report_Criteria must be run before here.


   *******
   Open ODBC channel
}
\details{
<other comments found in file>
NA
NA
   Execute the final SQL statement
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

