% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_by_catch.r
\name{F.getByCatch}
\alias{F.getByCatch}
\title{F.getByCatch}
\usage{
F.getByCatch(site, min.date, max.date)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}
}
\value{
A \code{csv} with all non-Chinook records between the two dates
  specified at the specified site.
}
\description{
Retreive all non-Chinook catch data between two dates at a site.
}
\details{
Function \code{F.getByCatch} utilizes query sequence By Catch to
  obtain all non-Chinook records within a catch.  See section Structured
  Query Language (SQL) Queries in function \code{F.run.sqlFile} for details.
}
\examples{
\dontrun{
#' #   ---- Fetch all by-catch data from the American.
site <- 57000
min.date <- "2013-01-01"
max.date <- "2013-06-01"
df <- F.getByCatch(site,min.date,max.date)
}
}
\seealso{
\code{F.run.sqlFile}
}
\author{
WEST Inc.
}
