% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_by_catch.r
\name{F.getByCatch}
\alias{F.getByCatch}
\title{F.getByCatch}
\usage{
F.getByCatch(site, min.date, max.date)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}
}
\value{
A \code{csv} with all non-Chinook records between the two dates
  specified at the specified site.
}
\description{
Retreive all non-Chinook data between two dates at site.
}
\examples{
\dontrun{

# requires mdb.

}
}
\seealso{

}

