% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_freq.r
\name{F.length.frequency}
\alias{F.length.frequency}
\title{F.length.frequency}
\usage{
F.length.frequency(site, taxon, run, min.date, max.date, output.file,
  by.lifestage)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{run}{The text seasonal identifier.  This is a one of \code{"Spring"}, 
\code{"Fall"}, \code{"Late Fall"}, or \code{"Winter"}.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{output.file}{The name of the file prefix under which output is to be 
saved.  Set to NA to plot to the Plot window.}

\item{by.lifestage}{When set to \code{TRUE}, distributions are reported for 
individual life stages.  Otherwise, they are collapsed and reported for all
fish.}
}
\value{
Either a single histogram, if variable \code{by.lifestage} is set to
  \code{FALSE}. Otherwise, a histogram for each individual life stage present
  within the data.
}
\description{
Plot frequency distribution of fork lengths.
  
  Input: db = full path and name of the Access data base to retrieve data
  from site = site ID of the place we want, trap locaton taxon = taxon number
  (from luTaxon) to retrieve run = run ID of fish we want to do estimates
  for. by.lifestage = if TRUE, produce histograms by lifestage, otherwise,
  lump all fish.
  
  Output: A graph, in "file".
}
\details{
Function \code{F.length.freq} utilizes function \code{F.get.indiv.fish.data} 
to obtain individual fish fork length data.
}
\examples{
\dontrun{
#   ---- Obtain graphical histograms for the American. 
site <- 57000
taxon <- 161980
run <- "Fall"
min.date <- "2014-01-01"
max.date <- "2014-06-06"
output.file <- "American"
by.lifestage <- TRUE

F.length.frequency(site,taxon,run,min.date,max.date,output.file,by.lifestage)
}
}
\author{
WEST Inc.
}
\seealso{
\code{F.get.indiv.fish.data}
}

