% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.lifestage.passage.assignLS3group.R
\name{F.lifestage.passage.assignLS3group}
\alias{F.lifestage.passage.assignLS3group}
\title{F.lifestage.passage.assignLS3group}
\usage{
F.lifestage.passage.assignLS3group(site, taxon, min.date, max.date, output.file,
  ci = TRUE, autoLS = TRUE, reclassifyFL = FALSE)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{output.file}{A text string indicating a prefix to append to all output.}

\item{ci}{A logical indicating if 95% bootstrapped confidence intervals 
should be estimated along with passage estimates.}

\item{autoLS}{Default is FALSE, no analytical life stage assignment is done.
If TRUE, the analytical life stage assignment is done, see details.}

\item{reclassifyFL}{A logical indicating if passage should be estimated via
forklength-based class groups.}
}
\value{
jared
}
\description{
3 life stage groups are fit and the program will decide if weight will be used or not
}
\details{
3 life stage groups are fit and the program will decide if weight will be used or not
}
\examples{
\dontrun{
#   ---- Calculate specified auto lifestage on the American.
site <- 57000
taxon <- 161980
min.date <- "2013-01-01"
max.date <- "2013-05-30"
output.file <- "Testing"
ci <- TRUE
autoLS <- TRUE
reclassifyFL <- FALSE

F.lifestage.passage.assignLS3group(site,taxon,min.date,max.date,output.file,ci,autoLS,reclassifyFL)

}
}
\author{
WEST Inc.
}
\seealso{
\code{F.run.passage}, \code{passageWithLifeStageAssign},
  \code{F.lifestage.passage.forkLength}, \code{F.lifestage.passage.assignLS},
  \code{F.lifestage.passage.assignLS2group}, 
  \code{F.lifestage.passage.assignLS2groupNoWeight},
  \code{F.lifestage.passage.assignLS3groupNoWeight},
  \code{F.lifestage.passage.assignLSNoWeight}
}

