% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.lifestage.passage.forkLength.R
\name{F.lifestage.passage.forkLength}
\alias{F.lifestage.passage.forkLength}
\title{F.lifestage.passage.forkLength}
\usage{
F.lifestage.passage.forkLength(site, taxon, min.date, max.date, by,
  output.file, ci = TRUE, enhmodel = FALSE, autols = FALSE,
  nls = NULL, weightuse = NULL)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{by}{A text string indicating the temporal unit over which daily 
estimated catch is to be summarized.  Can be one of \code{"day"}, 
\code{"week"}, \code{"month"}, \code{"year"}.}

\item{output.file}{A text string indicating a prefix to append to all output.}

\item{ci}{A logical indicating if 95\% bootstrapped confidence intervals 
should be estimated along with passage estimates.}

\item{enhmodel}{A logical indicating if enhanced efficiency models should 
be used to estimate trap efficiencies.  Default is \code{TRUE}.}

\item{autols}{Default of \code{FALSE} leads to no assigning of no analytical
life stage. If \code{TRUE}, assignment of analytical life stage is performed. 
See Details.}

\item{nls}{Number of life stage groups to estimate. Ignored if 
\code{autols=FALSE}.  See Details.}

\item{weightuse}{A logical indicating if variable \code{weight} should be used for
the analytical life stage assignment;  the default is \code{NULL}. Ignored
if \code{autols=FALSE}.  See Details.}
}
\value{
A \code{csv} table of passage estimates over the specified date 
  range, with fork-length groups down the rows, and Fall run across the columns.  A 
  \code{png} displaying proportion-of-catch bar charts of Fall run and fork-length groups.
  For each run and fork-length-group combination found within the specified data 
  range, an additional series of output.  A \code{csv} of daily passage 
  estimates for all traps operating at least one day, and catching at least 
  one fish, for all days within the specified date range. A \code{png} of 
  catch versus time, for all inclusive traps.    A \code{png} of daily 
  efficiency estimates, and accompanying \code{csv} for all traps operating 
  at least one day, and catching at least one fish, for all days within the 
  specified time period.  Finally, a bar chart of passage summarizing catch
  over the time period specified via \code{by}.
}
\description{
Estimate production by fork-length group and Fall run for all days within a
  date range.
}
\details{
Function \code{F.lifestage.passage.forkLength} is the main workhorse function for
  estimating passage with respect to forklength groups, restricted to Fall run.  As such, it
  calls several separate functions, some of which contain queries designed to
  run against an Access database.
  
  Generally, queries against a database comprise two main efforts.  The first
  involves a query for efficiency trial data, generally called "release" 
  data, and conducted via function \code{F.get.release.data}, while the 
  second queries for catch data via function \code{F.get.catch.data}.
  
  Once catch data are obtained, fish are partitioned as to whether or not 
  they were assigned and caught during a half-cone operation.  Function 
  \code{F.est.passage} wraps the functions that conduct the actual passage 
  estimation, which involves statistical fits of each of catch and efficiency
  over time.
  
  All calls to function \code{F.run.passage} result in daily passage 
  estimates, and courser temporal estimates, based on the value specified via
  \code{by}.  Regardless of the temporal partitioning, estimates are always 
  additionally summarized by year.  Function runs with \code{by} specified 
  as \code{"year"} output only one set of annual estimates.
  
  The difference between the specified \code{max.date} and \code{min.date}
  must be less than or equal to 366 days, as calculated via function
  \code{difftime}.
  
  Selection of \code{"week"} for input variable \code{by} results in weeks 
  displayed as customized Julian weeks, where weeks number 1-53.  The 
  specific mapping of days to weeks can be found within the \code{Dates} 
  table of any associated CAMP Access database.
  
  Forklength groupings are specified via global variable 
  \code{forkLengthCutPoints} in \code{GlobalVars}, and by default, include up
  to four distinct groupings.  However, if no fish exist for a particular 
  grouping, no output associated with that grouping is created.  When
  \code{reclassify=TRUE}, the biologically recorded \code{lifeStage} is 
  redefined via groups specifed in data frame \code{forkLengthCutPoints}, as 
  defined in \code{GlobalVars}.  Default behavior leads to four separate 
  fork-length-based groups.  Similar to reports that break out totals by 
  biologically assigned \code{lifeStage}s, reports utilizing breakout by fork
  length may not report totals for all four groups, if the river and date 
  range specified caught no fish with that particular range of fork lengths. 
  The remapping of \code{lifeStage} to reflect fork-length-based groupings is
  performed by function \code{reclassifyLS}.
}
\examples{
\dontrun{
#   ---- Estimate passage of Fall run based on fork-length groups
#   ---- on the American.
site <- 57000
taxon <- 161980
min.date <- "2013-01-01"
max.date <- "2013-06-01"
by <- "week"
output.file <- NA
ci <- TRUE
nls <- NULL
weightuse <- NULL
autols <- FALSE
reclassifyFL <- TRUE

F.lifestage.passage.forkLength(site,taxon,min.date,max.date,by,
  output.file,ci,nls,weightuse,autols,reclassify)
}
}
\seealso{
\code{F.get.release.data}, \code{F.get.catch.data},
  \code{F.est.passage}
}
\author{
WEST Inc.
}
