% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifestage_passage.r
\name{F.lifestage.passage}
\alias{F.lifestage.passage}
\title{F.lifestage.passage}
\usage{
F.lifestage.passage(site, taxon, min.date, max.date, output.file, ci = TRUE)
}
\arguments{
\item{site}{The identification number of the site for which estimates are
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{output.file}{The name of the file prefix under which output is to be 
saved.  Set to NA to plot to the Plot window.}

\item{ci}{A logical indicating if 95% bootstrapped confidence intervals 
should be estimated along with passage estimates.}
}
\value{
A table of passage estimates, with lifestages down the rows, and runs
  across the columns.
  
  A bar chart displaying relative proportions over both run and life stage.
  
  For each unique combination of run and life stage containing at least one
  fish, a graph \code{png} of catch over time, a \code{csv} of daily catch
  and passage.
  
  A graphical display of efficiency over time, for each trap, along with an
  accompanying \code{csv} tabular datasheet.
}
\description{
Estimate annual passage / production estimates between specified.
}
\details{
The date range difference specified via \code{max.date} and
  \code{min.date} must be less than or equal to 366 days.  Note that this
  cutoff allows for leap-year annual estimates.
  
  Passage requires the division of a non-zero catch by a non-zero efficiency.
  A non-zero catch numerator ensures that passage estimates are greater than
  zero, while a non-zero efficiency denominator ensures that passage 
  estimates are less than infinity.  The program exits if either of these 
  conditions are violated.
  
  Assuming validity, catch data are summarized via unique combinations of 
  \code{trapVisitID}, \code{FinalRun}, and \code{lifeStage}.  After 
  summarizing, all catch data are collapsed so as to have one line per 
  combination of these variables.  The use of a plus-count algorithm divvies
  up non-randomly selected fish into appropriate \code{FinalRun} and 
  \code{lifeStage} categories.  In this way, all fish are ultimately utilized
  in all pasasge estimates by life stage and run.
  
  Observed fish are also corrected for half-cone operations, in which the
  intake of fish is reduced to half of a cone's aperture being covered, so as
  to reduce flow, and thus captured fish.  Generally, all fish captured 
  during half-cone operations are multiplied by the value of the global 
  \code{halfConeMulti} variable, which serves as a multiplier to correct for 
  fish missed.  Variable \code{halfConeMulti} is set to 2.
  
  The unique combinations of run and life stage in the catch data dictate the
  reports generated. Thus, reports spanning different time periods and sites
  may report different run and life stage passage combinations.
}
\examples{
\dontrun{
  
  
}
}

