% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_model.r
\name{F.plot.catch.model}
\alias{F.plot.catch.model}
\title{F.plot.catch.model}
\usage{
F.plot.catch.model(df, file = NA)
}
\arguments{
\item{df}{A data frame that includes variables \code{trapPositionID},
\code{TrapPosition}, \code{batchDate}, \code{imputed.catch},
\code{modAssignedCatch}, and \code{imputedCatch}.}

\item{file}{The name of the file prefix under which output is to be saved. 
Set to \code{NA} to plot to the Plot window.}
}
\value{
A \code{png} graphical display of Daily Raw (un-inflated) catch, per 
  trap, by day.
}
\description{
Plot smoothed, possibly imputed, catch data over days per trap.
}
\details{
The input data frame usually has other catch variables, although 
  these currently serve no graphical purpose.  Variable \code{trapPositionID} contains 
  the numeric code of a trap, while variable \code{TrapPosition} contains in 
  text description.  Variable \code{imputed.catch} holds the proportion of 
  time, for that day, for which imputed occurred;  as such, its values are 
  between zero and one, inclusive.  Variable \code{imputedCatch} contains the
  total number of imputed fish for the day recorded via \code{batchDate}. 
  Variable \code{modAssignedCatch} contains the total number of caught fish 
  for that \code{batchDate} and incoporates any extra fish arising from 
  half-cone operations.  The value of \code{imputedCatch} is necessarily 
  equal to or less than the value of \code{modAssignedCatch}.
  
  The imputation curve displayed in the graph is obtained via function 
  \code{supersmu}, and serves as a reasonable approximation to the actual 
  spline used in the imputation procedure contained in function 
  \code{F.catch.model}.
  
  Jason note 4/26/2016 -- I changed this to graph half-cone adjusted catch. 
  This may not be what Doug wants.
}
\examples{
#   ---- Create a data frame.
batchDates <- as.POSIXct(strftime(seq(from=c(ISOdate(2014,1,1)),by="day",length.out=80),format="\%F"),tz="America/Los_Angeles")
df <- data.frame(trapPositionID=c(rep(12345,80),rep(98765,80)),
                 TrapPosition=c(rep('Left Bank',80),rep('Right Bank',80)),
                 batchDate=batchDates,
                 imputed.catch=rep(0,160),
                 modAssignedCatch=20*rnorm(length(batchDates))+624.26*exp(-(as.numeric(batchDates)/86400 - 16120)^2/260),
                 imputedCatch=rep(0,160))

df$modAssignedCatch <- ifelse(df$trapPositionID == 98765,15*rnorm(nrow(df))+0.50*df$modAssignedCatch,df$modAssignedCatch)
df$modAssignedCatch[df$modAssignedCatch < 0] <- 0
df$imputedCatch[120:124] <- df$modAssignedCatch[120:124]
df$imputed.catch[120:124] <- rep(1,5)

attr(df,"site.name") <- as.factor(c("A River in the Central Valley"))
attr(df,"species.name") <- "Chinook Salmon"
attr(df,"run.name") <- "Fall"
attr(df,"life.stage") <- "Fry"

#   ---- Plot results to plot window.
F.plot.catch.model( df, file=NA )
}
\author{
Trent McDonald (tmcdonald@west-inc.com)
}
\seealso{
\code{F.catch.model}
}

