% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eff_model.r
\name{F.plot.eff.model}
\alias{F.plot.eff.model}
\title{F.plot.eff.model}
\usage{
F.plot.eff.model(df, file)
}
\arguments{
\item{df}{A data frame that includes variables \code{TrapPositionID}, 
\code{batchDate}, \code{nReleased}, \code{nCaught}, \code{efficiency}, 
\code{imputed.eff}, and \code{trapPositionID}.  Currently variables 
\code{TrapPositionID} and \code{trapPositionID} appear to be the same, 
although both are called in subsequent points in the code.}

\item{file}{The name of the file prefix under which output is to be saved. 
Set to \code{NA} to plot to the Plot window.}
}
\value{
Output includes a \code{png} graphical display of estimated 
  efficiencies, per trap, by day, as well as a \code{csv} of the data points 
  utilized in making the graph.  The \code{csv} includes both the estimated 
  values from the specified efficiency model fit by function 
  \code{F.efficiency.model}, as well as the observed counts of fish deriving 
  from efficiency trials.
}
\description{
Plot per-trap efficiency values for all days between specified
  dates.
}
\examples{
\dontrun{
#   ---- Plot efficiency results housed in data frame 
#   ---- df to the plot window.
F.plot.eff.model( df, file=NA )
}
}
\author{
WEST Inc.
}
\seealso{
\code{F.efficiency.model}
}

