% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lifestages.r
\name{F.plot.lifestages}
\alias{F.plot.lifestages}
\title{F.plot.lifestages}
\usage{
F.plot.lifestages(df, file.root, plot.pies = FALSE)
}
\arguments{
\item{df}{A data frame containing runs in the rows and life stages (or 
fork-length groups) across the columns.}

\item{file.root}{A text string indicating a prefix to append to all output.}

\item{plot.pies}{If \code{TRUE}, a pie chart is produced instead of a bar 
chart.}
}
\value{
A graphical \code{png} plotting a bar chart showing relative numbers 
  of estimated passage, with lifestages down the rows, and runs across the 
  columns.  Generally, all passage routines utilize the bar-chart option via 
  paramter \code{plot.pies};  thus, pie charts are never returned.
}
\description{
Produce a plot of the passage associated with all distinct 
  combinations of lifestage (or fork length) and run.
}
\details{
Function \code{F.plot.lifestages} works on both life stages and 
  fork-length groups, depending on if \code{F.run.passage} or 
  \code{passageWithLifeStageAssign} was called in the case of the former, or 
  if \code{F.lifestage.passage.forkLength} in the case of the latter.
  
  Note that function \code{F.lifestage.passage.forkLength} only plots results
  from the Fall run, by design.
}
\examples{
\dontrun{
# Plot passage estimates housed in a data frame
# entitled allCatch.  
F.plot.lifestages(df=allCatch,file.root="PassageResults",plot.pies=FALSE)
}
}
\seealso{
\code{F.run.passage}, \code{passageWithLifeStageAssign}, 
  \code{F.lifestage.passage.forkLength}
}
\author{
WEST Inc.
}
