% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lifestages.r
\name{F.plot.lifestages}
\alias{F.plot.lifestages}
\title{F.plot.lifestages}
\usage{
F.plot.lifestages(df, file.root, plot.pies = FALSE)
}
\arguments{
\item{df}{<describe argument>}

\item{file.root}{<describe argument>}

\item{plot.pies=FALSE}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Produce a plot of the lifestages.

   df = a table with lifestages down the rows and runs across the columns.
   plot.pies = If TRUE, pie charts are produced
}
\details{
<other comments found in file>
   df <- df
   file.root <- output.file
   plot.pies <- F
   Get passage columns and totals row indices
   Get totals
   Pare down the data frame to just passage columns, no totals
   drop the runs with 0 fish. 
   drop the lifestages with no fish
   Reorder the columns to make the bar charts go through time
   Reorder the rows so life stages increase from left to right
   Do the plot
   If file=NA, a pdf graphing device is assumed to be open already.
   Shut down all graphics devices
   ---- Open PNG device
   If file=NA, a pdf graphing device is assumed to be open already.
   Shut down all graphics devices
   ---- Open PNG device
   Y axis label in margin
   the barplots 
   Write out the ls.names along the bottom
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

