% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_passage.r
\name{F.plot.passage}
\alias{F.plot.passage}
\title{F.plot.passage}
\usage{
F.plot.passage(df, out.file = "passage.png")
}
\arguments{
\item{df}{<describe argument>}

\item{out.file="passage.png"}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Plot a time series of passage estimates.  This can be called from
   Sweave files to put a figure in a report.

   This file produces a bar graph of the passage estimates in df. 

   input:
   df = passage estimate data frame from F.est.passage.  This is passage summarized by "day", "week", "month" or "year".
       It is assumed that the percentage of the estimate that was imputed is available in df$pct.imputed.catch. 


   df<-pass
     out.file='passage.png'
}
\details{
<other comments found in file>
   Construct matrix of bar heights
   Compute extent of y axis
   GRaph using barplot
   add a smoother line here. 
lines( supsmu( mp, hgts), lwd=2, col="darkblue" )  # Doug did not like this
   Add x axis labels
   Just label 1st of every month
   Label every day
        mons <-  format(df$date, "%b") 
        yrs <- format(df$date, "%Y")
        dt1 <- paste( "01", mons , sep="" )
        mons <-  as.numeric(format(df$date, "%m")) 
        dt2 <- as.POSIXct( strptime( paste("1",mons+1,yrs), "%d %m %Y" )) 
        dt2 <- dt2 - 24*60*60
        dt2 <- format(dt2, "%d%b")
        
        dt1[1] <- format( min(df$date), "%d%b" )
        dt2[length(df$date)] <- format( max(df$date), "%d%b" )
 "weekly"  (yearly does not get plotted)   
 jason -- add.  need a lookup table that maps days to our defined julian weeks here.  
 jason add. -- copy from passage.r
 can't figure out how to join on posix dates.  so cheating. 
         df <- subset(df, select = -c(date.alone,uniqueDate,julianWeek,julianWeekLabel) )   # jason - don't think we need this here.
       old
         dt1 <- df$date
         dt2 <- c(df$date[-1] - 24*60*60, max(df$date))
         dt1 <- format(dt1, "%d%b")
         dt2 <- format(dt2, "%d%b")
         dt <- paste(dt1, dt2, sep="-")
         my.cex <- .75
   Add y axis labels
   Add total passage
   Add title
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

