% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclassifyLS.R
\name{F.reclassifyLS}
\alias{F.reclassifyLS}
\title{F.reclassifyLS}
\usage{
F.reclassifyLS(catch)
}
\arguments{
\item{catch}{A catch dataframe originating from a catch-query sequence.  See
function \code{F.get.catch.data}.  At the least, \code{catch} must contain
variables entitled \code{forkLength} and \code{lifeStage}.}
}
\value{
A dataframe with original variable \code{lifeStage} values containing
  "\code{Fry}," "\code{Parr}," and "\code{Smolt}" into the label values
  specified via the Global environment dataframe \code{forkLengthCutPoints}.
  See \code{GlobalVars}.
}
\description{
Reclassify lifeStage via forklength into groups specified
via Global Enviornment dataframe \code{forkLengthCutPoints}.
}
\details{
Generally speaking, the process by which a new \code{lifeStage}
  variable is constructed is straight-forward.  Care must be taken, however,
  to ensure that records with no recorded \code{forkLength}, i.e., those
  containing a value of \code{NA}, are mapped to \code{"Unassigned"} in the
  new \code{lifeStage} variable.  This process is programmed in order to
  ensure correct processing of plus counts via function \code{F.assign.1dim},
  for records containing valid \code{FinalRun}, but not \code{lifeStage}, or
  \code{forkLength}, as the case may be.
}
\examples{
 
\dontrun{
catchWithNewLifeStageCats <- reclassifyLS(catch)
}
}
\seealso{
\code{GlobalVars.R}, \code{F.get.catch.data}, \code{F.assign.1dim}
}

