% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_summary.r
\name{F.release.summary}
\alias{F.release.summary}
\title{F.release.summary}
\usage{
F.release.summary(site, taxon, run, min.date, max.date, output.file)
}
\arguments{
\item{site}{<describe argument>}

\item{taxon}{<describe argument>}

\item{run}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}

\item{output.file}{<describe argument>}
}
\value{
<describe return value>
}
\description{
This is the outer wrapper function to summarize releases at a particular site, for a particular run, for a particular taxon for a particular year.

   Inputs:
   site = site number
   taxon = taxon number
   run = run number
   year = year number
   output.file = name of .CSV file to put passage estimates into.  Full path is allowed.

table.names <- data(table.names)  # Names of table in data base is stored in the campass package.  Eventually do this, then delete 'get' statement
table.names <- get( "table.names", env=.GlobalEnv )
db.name <- get( "db.file", env=.GlobalEnv ) 


   ---- Fetch efficiency data
}
\details{
<other comments found in file>
 get a couple of variables from quickie lookups
   ---- Or, could write summary to CSV file
     rs <- attr(release.df,"run.season")
     rs <- paste( format(rs[1], "%d-%b-%Y"), "to", format(rs[2], "%d-%b-%Y"))
   ---- Send messages back to the interface
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

