% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sqlFile.r
\name{F.run.sqlFile}
\alias{}
\alias{F.run.sqlFile}
\title{Opens, formats, and runs all SQL queries from a text file.}
\usage{
F.run.sqlFile(ch, sqlFile, echo = TRUE, check.drops = FALSE, sql.code.dir,
  ...)
}
\arguments{
\item{ch}{Output originating from a call to RODBC function odbcConnectAccess,
which essentially opens a connection between R and Access, or more 
generally, an ODBC database.}

\item{sqlFile}{A text file containing one or more SQL-formatted query strings.}

\item{sql.code.dir}{Directory storing all SQL code for the CAMP RST platform.}

\item{...}{Additional parameters needed for evaluation of the underlying SQL queries contained in sqlFile.  See 'Details.'}
}
\value{
FALSE, assuming successful completion of a SQL query.  Otherwise, an 
  error message to either the Console window, if running directly via R, or 
  to the run_R.out file, if running via the Platform.
}
\details{
Function F.sql_error_check is used immediately after the
use of RODBC function sqlQuery, which queries an Access database via a text
string formatted as a SQL query. Function sqlQuery, if successful, results in
a dataframe, possibly with zero rows. Otherwise, a character vector with
error messsages is returned. Function F.sql.error.check examines if an error
message is returned, and if found, stops execution and prints an error
message to the R_run.out text file or the R Console.
}
\examples{
db.file <- "C:/path/CAMP.mdb"                    # the location of CAMP.mdb
db <- get( "db.file", env=.GlobalEnv )           # assign the mdb string to an object
ch <- odbcConnectAccess(db)                      # open a connection between R and Access

# a good query
ans <- sqlQuery( ch, "SELECT * FROM luRun" )     # select all rows via a SQL query
F.sql.error.check(ans)                           # check for query result validity

#' # a bad query
ans <- sqlQuery( ch, "SELECT * FROM fakeTable" ) # select all rows via a SQL query
F.sql.error.check(ans)                           # check for query result validity

}
\seealso{
sqlQuery
}

