% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_by_date.r
\name{F.size.by.date}
\alias{F.size.by.date}
\title{F.size.by.date}
\usage{
F.size.by.date(site, taxon, run, min.date, max.date, output.file)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{run}{The text seasonal identifier.  This is a one of \code{"Spring"}, 
\code{"Fall"}, \code{"Late Fall"}, or \code{"Winter"}.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{output.file}{The name of the file prefix under which output is to be 
saved.  Set to NA to plot to the Plot window.}
}
\value{
A graphical \code{png} entitled via parameter \code{output.file}, displyaing 
fork length in millimeters as a function of time.  Also, an accompanying \code{csv}
containing the data used in plotting.
}
\description{
Plot fork length (mm) of fish by date of catch for a particular
  site and year.
}
\details{
Function \code{F.size.by.date} first compiles fish data via function
  \code{F.get.indiv.fish.data}.  Trapping instances include both valid and
  invalid catch, as determined via variable \code{includeCatchID}.  Thus, 
  resulting graphs may display both valid and invalid fishing instances, if 
  both trappining instances took place within the time period specified 
  via \code{min.date} and \code{max.date.}
  
  All observations are jittered when plotted, so as to ensure the volume of 
  any one particular date and fork length combination can be visualized.  
  
  In the case when no valid catch records are found, and hence no fish can 
  contribute to the analysis, a blank \code{png} stating as such is returned.
}
\examples{
\dontrun{
# Create a plot of fork length by data for the American.
site <- 57000
taxon <- 161980
run <- "Fall"
min.date <- "2014-01-01"
max.date <- "2014-06-01"
output.file <- "American"

F.size.by.date(site,taxon,run,min.date,max.date,output.file)
}
}
\author{
WEST Inc.
}
\seealso{
\code{F.get.indiv.fish.data}
}

