% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fish_visit.r
\name{F.summarize.fish.visit}
\alias{F.summarize.fish.visit}
\title{F.summarize.fish.visit}
\usage{
F.summarize.fish.visit(catch, variable)
}
\arguments{
\item{catch}{<describe argument>}

\item{variable}{<describe argument>}
}
\value{
<describe return value>
}
\description{
Summarize fish count and fork lengths over visits.

   catch = data frame containing one line per fish or fish group with identical lengths.
       If they did not catch any member of a particular taxon, catch will have 0 rows.  This 
       data frame will usually be the output of F.get.indiv.fish.data()

   variable = summarize over all data (so inflated counts), or assigned counts (measured fish only).

   Output = a data frame with one line per visit, with catch summarized.
}
\details{
<other comments found in file>
   variable <- 'unassigned'
   variable <- 'halfcone'
 jason add - first pass assumed that unassigned fork length is NA.  but...not NA for one day in run i'm investigating.  force this.
   These are variables that are constant within a trapVisit, run, and lifestage
   indexes = all unique combinations of visit, run, and life stage
   indexes = NA for all gaps in the trapping sequence
   These are variables that are constant within a trapVisit, run, and lifestage
   indexes = all unique combinations of visit, run, and life stage
   indexes = NA for all gaps in the trapping sequence
   These are variables that are constant within a trapVisit, run, and lifestage
   indexes = all unique combinations of visit, run, and life stage
   indexes = NA for all gaps in the trapping sequence
   These are variables that are constant within a trapVisit, run, and lifestage
   indexes = all unique combinations of visit, run, and life stage
   indexes = NA for all gaps in the trapping sequence
   Because of the gap lines, there are NA's in indexes (because there are NA's in trapVisitID). 
   Fix these.  To assure indexes is same length as catch, we cannot simply remove NA's here. 
   Solution: set them to -1, then remove -1 from u.ind.  This way the loop below 
   just skips them.
   Initialize place to store summarized catches. 
u.ind <- sort(unique(indexes))  # NA's in indexes are lost here, in the sort
catch.fl <- as.data.frame( matrix( NA, length(u.ind), length(const.vars) + 3 ))
names( catch.fl ) <- c(const.vars, "n.tot", "mean.fl", "sd.fl")
   Count number of fish, compute mean fork length, etc per visitID in catch.  
   Copy over constant variables
   Take weighted averages of fork lengths using 'n' as weights
   Add back in the lines with missing indexes.  These correspond to gaps in trapping
    tmp <- catch[ indexes < 0, const.vars ]
    tmp <- cbind( tmp, matrix( NA, sum(indexes < 0), 3))
    names( tmp ) <- c(const.vars, "n.tot", "mean.fl", "sd.fl")
    catch.fl <- rbind( catch.fl,  tmp )
   Sort the result by trap positing and trap visit
 no measured fish, but both finalrun and lifestage assigned.
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

