% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_index.r
\name{F.summarize.index}
\alias{F.summarize.index}
\title{F.summarize.index}
\usage{
F.summarize.index(dt, summarize.by)
}
\arguments{
\item{dt}{<describe argument>}

\item{summarize.by}{<describe argument>}
}
\value{
<describe return value>
}
\description{
return a list suitable for calling tapply which summarizes a vector by 'summarize.by'.

   dt = a POSIXct date vector
   summarize.by = a string equal to "week", "month", "year", or "day"

 dt <- catch.df$batchDate
}
\details{
<other comments found in file>
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

