% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_effort.r
\name{F.weekly.effort}
\alias{F.weekly.effort}
\title{F.weekly.effort}
\usage{
F.weekly.effort(site, taxon, min.date, max.date, output.file)
}
\arguments{
\item{site}{<describe argument>}

\item{taxon}{<describe argument>}

\item{min.date}{<describe argument>}

\item{max.date}{<describe argument>}

\item{output.file}{<describe argument>}
}
\value{
<describe return value>
}
\description{
compute weekly effort.  Effort is number of days/hours that trap was fishing.

   site = site to do summary for
   min.date = minimum date to include
   max.date = maximum date to include
   output.file = root name of output file(s)

   ---- Check that times are less than 1 year apart
}
\details{
<other comments found in file>
  === update 5/23/2015 === requires new connie-query. ===================================
   Open ODBC channel
   *****
   This SQL file develops the hours fished and TempSamplingSummary table 
   *****
   This SQL generates the daily allocation of minutes, so they sum to 1440
   Now, fetch the result 
 manipulate data
 check to make sure minutes.1 minutes.2 minutes.3 all exist, and if not, insert.
 set up df of dates and time possibilities.  get min and max over all traps so all resulting dfs are the same size.
 get helpful stuff for looping about traps
 clean up the julian week info 
 check for julian week 53
 make nice dfs for csv and for prep for plotting
 combine all the traps into an overall df
 the.sum <- NULL
 for(i in 1:nTraps){
   if(i == 1){
     the.sum <- df.Day[[1]][,c(6:10)]
   } else {
     the.sum <- the.sum + df.Day[[i]][,c(6:10)]
   }
 }  
 df.Day[[nTraps + 1]]$nTrapsDay <- (df.Day[[nTraps + 1]]$Total - df.Day[[nTraps + 1]]$Diff) / 1440
 get site label
 for each trap (and over all traps)
 check for week 53
   Compute locations for, and labels of, months'
   Lower plot
   Compute locations for, and labels of, months'
 output csv of minutes for trap i
   out.week.table2 <- paste(output.file, " Daily Effort Summary Table - ",eff.df3$Position[1],".csv", sep="")  
   write.table( df.Day[[i]], file=out.week.table2, sep=",", row.names=FALSE, col.names=TRUE)
cat(eff.df3Print)
   ---- Send messages back to the interface
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

