% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAlk.camp.R
\name{GetAlk.camp}
\alias{GetAlk.camp}
\title{Extrae ALK para una especie de los ficheros del camp 
 
Función de acceso a datos: recupera la clave talla edad de una especie y campaña determinadas}
\usage{
GetAlk.camp(gr, esp, camp, dns = "Porc", plus = 8, AltAlk = FALSE)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camp}{Campaña de la que se extraen los datos un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{AltAlk}{ALK alternativa tomada de un fichero edadXYY.dbf sin ruta ni extensión}
}
\description{
Extrae ALK para una especie de los ficheros del camp 
 
Función de acceso a datos: recupera la clave talla edad de una especie y campaña determinadas
}
\examples{
GetAlk.camp(1,43,"N93","Cant",8) 
GetAlk.camp(1,45,"P03","Pnew",AltAlk="edadXYY") 
}
\seealso{
Other edades: \code{\link{bubbage.camp}},
  \code{\link{datagegr.camp}}, \code{\link{edadsect.camp}},
  \code{\link{edadstr.camp}},
  \code{\link{grafedtal.camps}},
  \code{\link{grafedtal.camp}},
  \code{\link{logabage.camp}},
  \code{\link{logabage2.camp}}, \code{\link{maphistage}}
}
