% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafedtal.camp.r
\name{grafedtal.camp}
\alias{grafedtal.camp}
\title{Histograma de distribución de tallas con edades 
 
 Función gráfica a partir de los ficheros del camp. Histograma distribución de tallas estratificada por edad}
\usage{
grafedtal.camp(gr, esp, camp, dns = "Porc", plus = 8, cor.time = TRUE,
  excl.sect = NA, AltAlk = NA, ti = FALSE, leg = TRUE, cexleg = 1,
  es = TRUE, plot = TRUE, ymax = NA, out.dat = FALSE)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces sólo hay claves de talla para peces y cigala?}

\item{esp}{Código de la especie numérico o carácter con tres espacios. Sólo admite una especie por gráfica}

\item{camp}{Campaña a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant", Golfo de Cádiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{plus}{Edad plus: Edad considerada como plus, todas las edades mayores se suman como edad +}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{AltAlk}{Clave talla edad alternativa sin ruta ni extensión, NA por defecto usa la clave de la campaña edadXYY.dbf}

\item{ti}{Si T añade título al gráfico, el nombre de la especie en latín}

\item{leg}{Si T añade leyenda}

\item{es}{Si T gráfico en castellano, si F gráfico en inglés}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{ymax}{Valor máximo del eje y}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}
}
\value{
Saca gráfica con distribución de tallas y la distribución de las edades en cada talla. Si out.dat=TRUE saca un data.frame con columnas n(valor del número de la distribución estratifcada para la talla y la edad),talla,edad,camp. Da error si no existe ALK para la especie en la campaña
}
\description{
Histograma de distribución de tallas con edades 
 
 Función gráfica a partir de los ficheros del camp. Histograma distribución de tallas estratificada por edad
}
\examples{
grafedtal.camp(1,43,"P09","Porc",es=FALSE,out.dat=TRUE)
}
\seealso{
{\link{grafedtal.camps}}

Other edades: \code{\link{ALKs.dns.camp}},
  \code{\link{GetAlk.camp}}, \code{\link{bubbage.camp}},
  \code{\link{datagegr.camp}}, \code{\link{edadsect.camp}},
  \code{\link{edadstr.camp}},
  \code{\link{grafedtal.camps}},
  \code{\link{logabage.camp}},
  \code{\link{logabage2.camp}}, \code{\link{maphistage}}
}
