% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListFauna.lan.R
\name{ListFauna.lan}
\alias{ListFauna.lan}
\title{Capturas de un lance específico}
\usage{
ListFauna.lan(camp, dns = "Porc", lan, gr = NA, out = TRUE)
}
\arguments{
\item{camp}{Campaña de la que se extrae los datos del lance (lan) un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc" o "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa"}

\item{lan}{Lance del que se extrae la faunística}

\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados. Si NA, muestra todos los grupos a excepción de 6 desechos y otros no orgánicos}

\item{out}{salida de la función, por defecto TRUE, pero si FALSE no saca el resultado en pantalla (lo asigna a variable para ListFauna.lans)}
}
\value{
Devuelve un data.frame con columnas grupo,esp,lanceui,especie,peso,numero
}
\description{
Muestra un listado de las especies capturadas en peso y número en un lance concreto. Se pueden seleccionar grupos de especies
}
\examples{
ListFauna.lan("12C","Cant",1,gr=c(1,2))
}
\seealso{
Other ListadosFauna: 
\code{\link{ListFauna.camp}()},
\code{\link{ListFauna.groups}()},
\code{\link{ListFauna.lans}()},
\code{\link{ListFaunaTals.camps}()},
\code{\link{ListFaunaTals.camp}()},
\code{\link{ListFaunaTals.groups}()}
}
\concept{ListadosFauna}
