% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListFaunaTals.camp.R
\name{ListFaunaTals.camp}
\alias{ListFaunaTals.camp}
\title{Capturas y tallas mínima y máxima de grupos de especies para todos los lances de una campaña}
\usage{
ListFaunaTals.camp(gr = "1", camp, dns, excl.sect = NA, incl2 = TRUE)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados. 6 desechos y no orgánicos no se puede incluir.}

\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{incl2}{Si T incluye los lances especiales "2"}
}
\description{
Muestras las abundancias totales sin estratificar, en número y peso y las tallas mínima y máxima de un grupo de especies para todos los lances de una campaña
}
\examples{
ListFaunaTals.camp(gr="1",camp="N90",dns="Cant",excl.sect=NA)
}
\seealso{
{\link{ListFauna.camp}}
}
