% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListFaunaTals.camps.R
\name{ListFaunaTals.camps}
\alias{ListFaunaTals.camps}
\title{Capturas y tallas mínima y máxima de grupos de especies para todos los lances de varias campaña}
\usage{
ListFaunaTals.camps(
  gr = "1",
  camps,
  dns,
  excl.sect = NA,
  cor.time = TRUE,
  profrange = NA,
  incl2 = TRUE
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados. 6 desechos no se puede elegir en esta función}

\item{camps}{campañas (años) a representar en el mapa: Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{excl.sect}{Sectores a excluir como carácter, se pueden elegir tanto los sectores como estratos}

\item{profrange}{Si c(profmin,profmax) filtra por ese rango de profundidad, por defecto NA no filtra por profunidades, debe ser o NA o un rango con dos profundidades}

\item{incl2}{Si T incluye los lances especiales "2"}
}
\description{
Ver documentación ListFaunaTals.camp
}
\examples{
ListFaunaTals.camps(gr="1",Nsh[7:27],"Cant",excl.sect=NA,profrange=NA,incl2=TRUE)
ListFaunaTals.camps(gr="1",Nsh,"Cant",excl.sect=NA,incl2=TRUE,profrange=c(400,999))
}
\seealso{
Other ListadosFauna: 
\code{\link{Fauna.camp}()},
\code{\link{ListFauna.camp}()},
\code{\link{ListFauna.groups}()},
\code{\link{ListFauna.lans}()},
\code{\link{ListFauna.lan}()},
\code{\link{ListFaunaTals.camp}()},
\code{\link{ListFaunaTals.groups}()}
}
\concept{ListadosFauna}
