% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maparea.r
\name{maparea}
\alias{maparea}
\title{Mapa general del banco de Porcupine}
\usage{
maparea(
  ICESrect = FALSE,
  ICESlab = FALSE,
  ICESlabcex = 0.7,
  es = TRUE,
  leg = TRUE,
  bw = FALSE,
  dens = 0,
  sectcol = F
)
}
\arguments{
\item{ICESrect}{Si T incluye los los rectángulos ICES}

\item{ICESlab}{Si T incluye las etiquetas de los rectángulos ICES}

\item{ICESlabcex}{tamaño del ICESlab en cex, .5 por defecto subirlo si se quiere más grande}

\item{es}{Textos en español, si F en inglés}

\item{leg}{incluye la leyenda con los colores/patrones de los estratos}

\item{bw}{Si T mapa en blanco y negro respecto a tierra y puntos, en caso contrario en color. Para sacar el diseño de estratos de Porcupine se utiliza sectcol=TRUE y leg=TRUE}

\item{dens}{si mayor de 0 las superficies de los estratos tienen patrones de líneas}

\item{sectcol}{si T pone los sectores con color de fondo, en caso contrario lo deja en blanco, bueno para armap.tot}
}
\description{
Mapa con la estratificación y el área general del banco de Porcupine incluyendo parte de Irlanda
}
\examples{
maparea(sectcol=TRUE,leg=TRUE)
maparea(sectcol=FALSE,leg=FALSE)
}
\seealso{
Other mapas: 
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{NepFUs.camp}()},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{map.check}()},
\code{\link{maphistage}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other Porcupine: 
\code{\link{MapPorc}()},
\code{\link{Psh}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{mapporco}()}
}
\concept{Porcupine}
\concept{mapas}
