% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maparea.r
\name{maparea}
\alias{maparea}
\title{Mapa general del banco de Porcupine}
\usage{
maparea(
  ICESrect = FALSE,
  ICESlab = FALSE,
  ICESlabcex = 0.7,
  es = TRUE,
  leg = TRUE,
  dpth = TRUE,
  bw = FALSE,
  dens = 0,
  sectcol = FALSE,
  SACs = FALSE,
  filSAC = FALSE,
  graf = FALSE,
  xpng = 1000,
  ypng = 900,
  ppng = 15
)
}
\arguments{
\item{ICESrect}{Si T incluye los los rectángulos ICES}

\item{ICESlab}{Si T incluye las etiquetas de los rectángulos ICES}

\item{ICESlabcex}{tamaño del ICESlab en cex, .5 por defecto subirlo si se quiere más grande}

\item{es}{Textos en español, si F en inglés}

\item{leg}{incluye la leyenda con los colores/patrones de los estratos}

\item{dpth}{si T incluye los datos de estratos batimétricos en la leyenda, si F solo los sectores}

\item{bw}{Si T mapa en blanco y negro respecto a tierra y puntos, en caso contrario en color. Para sacar el diseño de estratos de Porcupine se utiliza sectcol=TRUE y leg=TRUE}

\item{dens}{si mayor de 0 las superficies de los estratos tienen patrones de líneas}

\item{sectcol}{si T pone los sectores con color de fondo, en caso contrario lo deja en blanco, bueno para armap.tot}

\item{SACs}{si T incluye la silueta de las areas de especial atención irlandesas SACs}

\item{filSAC}{si T las SACs se llenan con color verde, si F sólo queda el contorno}

\item{graf}{si F el gráfico sale en la pantalla, si nombre fichero va a fichero en el directorio de trabajo del Rstudio ver getwd()}

\item{xpng}{width archivo png si graf es el nombre del fichero}

\item{ypng}{height archivo png si graf es el nombre del fichero}

\item{ppng}{points png archivo si graf es el nombre del fichero}
}
\description{
Mapa con la estratificación y el área general del banco de Porcupine incluyendo parte de Irlanda
}
\examples{
maparea(sectcol=TRUE,leg=TRUE,graf="MapArea.ejemplo")
maparea(sectcol=FALSE,leg=FALSE)
maparea(sectcol=FALSE,leg=FALSE,SACs=T,filSAC=T)
}
\seealso{
Other mapas: 
\code{\link{BelgicaMound}},
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapHidro.camp}()},
\code{\link{MapHidro_b.camp}()},
\code{\link{MapLansGPS}()},
\code{\link{MapLansHH}()},
\code{\link{MatchHidro.ctd}()},
\code{\link{NepFUs.camp}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{VedasBesugo}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{hmmSAC}},
\code{\link{map.check}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maphistage}()},
\code{\link{maphistal.one}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other Porcupine: 
\code{\link{BelgicaMound}},
\code{\link{MapPorc}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{Psh}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{hmmSAC}},
\code{\link{mapporco}()}
}
\concept{Porcupine}
\concept{mapas}
