% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapArsa.r
\name{MapArsa}
\alias{MapArsa}
\title{Mapa del Golfo de Cádiz}
\usage{
MapArsa(xlims = c(-8.14, -5.54), ylims = c(35.95, 37.33), lwdl = 1,
  leg = F, cuadr = FALSE, cuadrMSFD = FALSE, ICESrect = FALSE,
  ICESlab = FALSE, ICESlabcex = 0.5, ax = TRUE, bw = F,
  wmf = FALSE, es = TRUE, places = TRUE)
}
\arguments{
\item{xlims}{Define los limites longitudinales del mapa, los valores por defecto son los del total del área de la campaña}

\item{ylims}{Define los limites latitudinales del mapa, los valores por defecto son los del total del área de la campaña}

\item{lwdl}{Ancho de las líneas del mapa}

\item{leg}{si T incluye una legenda con los estratos de profundidad y los estratos se colorean respectivamente}

\item{cuadr}{Si T saca las cuadrículas de 5x5 millas naúticas}

\item{cuadrMSFD}{Si T dibuja cuadrícula de 10 millas naúticas utilizada para la evaluación de la estrategia marina (MSFD)}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{ICESlab}{Si T incluye las etiquetas de los rectángulos ICES}

\item{ICESlabcex}{tamaño del ICESlab en cex, .5 por defecto subirlo si se quiere más grande}

\item{ax}{Si T saca los ejes x e y}

\item{bw}{si T mapa con tierra en gris, si F tierra en color}

\item{wmf}{Si T saca a fichero metafile Arsaconc.emf}

\item{es}{Si T saca titulos y ejes en español}

\item{places}{Si T saca ciudades y puntos geográficos de referencia}
}
\value{
Saca en pantalla el mapa y es utilizada por otras funciones
}
\description{
Función auxiliar para sacar mapas de la campaña ARSA
}
\examples{
MapArsa()
}
\seealso{
{\link{MapCant}}, {\link{mapporco}}

Other mapas base: \code{\link{MapCant}},
  \code{\link{MapIberia}}, \code{\link{MapMedit}},
  \code{\link{MapNort}}, \code{\link{MapPorc}},
  \code{\link{mapporco}}

Other ARSA: \code{\link{As1}}, \code{\link{As2}}
}
\concept{ARSA}
\concept{mapas base}
