% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapArsa.r
\name{MapArsa}
\alias{MapArsa}
\title{Mapa del Golfo de Cádiz}
\usage{
MapArsa(
  xlims = c(-8.15, -5.52),
  ylims = c(35.95, 37.335),
  lwdl = 1,
  leg = F,
  cuadr = FALSE,
  cuadrMSFD = FALSE,
  ICESrect = FALSE,
  ICESrectcol = 2,
  ICESlab = FALSE,
  escala = FALSE,
  cex.scala = 0.6,
  FU = NA,
  ColFU = "white",
  dens = 20,
  FUsLab = FALSE,
  ICESlabcex = 0.8,
  ax = TRUE,
  bw = F,
  graf = FALSE,
  es = TRUE,
  places = TRUE
)
}
\arguments{
\item{xlims}{Define los limites longitudinales del mapa, los valores por defecto son los del total del área de la campaña}

\item{ylims}{Define los limites latitudinales del mapa, los valores por defecto son los del total del área de la campaña}

\item{lwdl}{Ancho de las líneas del mapa}

\item{leg}{si T incluye una legenda con los estratos de profundidad y los estratos se colorean respectivamente}

\item{cuadr}{Si T saca las cuadrículas de 5x5 millas naúticas}

\item{cuadrMSFD}{Si T dibuja cuadrícula de 10 millas naúticas utilizada para la evaluación de la estrategia marina (MSFD)}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{ICESrectcol}{color de las líneas de los rectangulos}

\item{ICESlab}{Si T incluye las etiquetas de los rectángulos ICES}

\item{cex.scala}{tamaño de la fuente de la escala, .6 por defecto}

\item{FU}{Por defecto NA pero si se incluye un vector con la lista de las unidades funcionales las pinta (las disponibles en MapNort son FU31, FU25, FU26)}

\item{ColFU}{por defecto "chartreuse" o "white" selecciona el color de las unidades funcionales de cigala}

\item{dens}{densidad (transparanecia) del color de la FU, 20 por defecto, menos no se aprecia mucho, más de 20 sale color fijo, 0 quita el color de relleno}

\item{FUsLab}{Por defecto F, pero si T incluye una etiqueta con los nombres de las FUs seleccionadas en FUs}

\item{ICESlabcex}{tamaño del ICESlab en cex, .5 por defecto subirlo si se quiere más grande}

\item{ax}{Si T saca los ejes x e y}

\item{bw}{si T mapa con tierra en gris, si F tierra en color}

\item{graf}{Si F no saca nada, si pones el nombre de un gráfico saca un grafico png y al final del porceso dice dónd está el mapa con ese nombre:}

\item{es}{Si T saca titulos y ejes en español}

\item{places}{Si T saca ciudades y puntos geográficos de referencia}

\item{escaala}{si T incluye una escala del mapa}
}
\value{
Saca en pantalla el mapa y es utilizada por otras funciones
}
\description{
Funci?n auxiliar para sacar mapas de la campaña ARSA
}
\examples{
MapArsa()
MapArsa(ICESrect = T,ICESlab = T,ICESrectcol = T,FU="FU30",ColFU="white",dens=20,xlims = c(-7.7,-6),ylims = c(36,37.3))
}
\seealso{
{\link{MapCant}}, {\link{mapporco}}

Other mapas base: 
\code{\link{MapCant}()},
\code{\link{MapIberia}()},
\code{\link{MapMedit}()},
\code{\link{MapNort}()},
\code{\link{MapPorc}()},
\code{\link{mapporco}()},
\code{\link{mapscale}()}

Other ARSA: 
\code{\link{As1}},
\code{\link{As2}}
}
\concept{ARSA}
\concept{mapas base}
