% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCant.r
\name{MapCant}
\alias{MapCant}
\title{Mapa del Cantábrico y Galicia para una o varias especies y también combina campañas superponiéndolas
para dar una idea de la distribución de la especie. Si es muy escasa se puede pasar a representar
solo a puntos de aparición}
\usage{
MapCant(
  gr,
  esp,
  camps,
  dns = "Cant",
  puntos = FALSE,
  DpthStr = TRUE,
  bw = FALSE,
  add = FALSE,
  esc.mult = 1,
  ti = FALSE,
  ind = "p",
  ceros = F,
  es = T,
  places = T
)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 desechos y otros, 9 escoge todos los orgánicos pero excluye desechos}

\item{esp}{Código de la especie númerico o caracter con tres espacios. 999 para todas las especies del grupo}

\item{camps}{Campañas a representar en el mapa, con MapCant1 se puede sacar más de  un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: solo para el Cantábrico "Cant"}

\item{puntos}{si TRUE en el mapa muestra sólo los puntos en los que aparece la especie opcionalmente con color por estratos de profundidad, si false son puntos proporcionales al tamaño de la abundancia}

\item{DpthStr}{si TRUE los puntos salen por estrato de profundidad con legenda explicándolo}

\item{bw}{si T los colores salen en blanco, si F en lightblue}

\item{add}{Si T añade los puntos al gráfico actual, si F dibuja uno nuevo}

\item{esc.mult}{Varia el tamaño de los puntos valores mayores de 1 disminuyen tamaño, valores de menos de cero lo aumentan}

\item{ti}{Si T añade titulo al mapa, el nombre de la especie en latín}

\item{ind}{Si "p" representa los datos de biomasa, si "n" representa los datos de abundancia}

\item{ceros}{Si T representa los ceros con cruces +, si F no representa lances sin captura}

\item{color}{Color de los puntos que representan las abundancias}
}
\description{
Crea un mapa para la campaña Demersales con información para la especie solicitada en la campañas solicitadas, para varias campañas ver también {\link{maphist}}
}
\seealso{
{\link{maphist}}

Other mapas base: 
\code{\link{MapArsa}()},
\code{\link{MapIberia}()},
\code{\link{MapMedit}()},
\code{\link{MapNort}()},
\code{\link{MapPorc}()},
\code{\link{mapporco}()},
\code{\link{mapscale}()}

Other Galicia Cantabrico: 
\code{\link{MapNort}()}
}
\concept{Galicia Cantabrico}
\concept{mapas base}
