% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCant.r
\name{MapCant}
\alias{MapCant}
\title{Mapa de Demersales Norte para una sola campaña}
\usage{
MapCant(gr, esp, camp, dns = "Cant", color = 1, add = FALSE,
  escala = NA, ti = FALSE, ind = "p", ceros = F)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados 6 desechos y otros, 9 escoge todos los orgánicos pero excluye desechos}

\item{esp}{Código de la especie númerico o caracter con tres espacios. 999 para todas las especies del grupo}

\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: solo para el Cantábrico "Cant"}

\item{color}{Color de los puntos que representan las abundancias}

\item{add}{Si T añade los puntos al gráfico actual, si F dibuja uno nuevo}

\item{escala}{Varia el tamaño de los puntos}

\item{ti}{Si T añade titulo al mapa, el nombre de la especie en latín}

\item{ind}{Si "p" representa los datos de biomasa, si "n" representa los datos de abundancia}

\item{ceros}{Si T representa los ceros con cruces +, si F no representa lances sin captura}
}
\description{
Crea un mapa para el Cantábrico y Galicia con información para la especie solicitada en la campaña solicitada, sólo una campaña para mas campañas ver maphist
}
\seealso{
{\link{maphist}}

Other mapas base: \code{\link{MapArsa}},
  \code{\link{MapIberia}}, \code{\link{MapMedit}},
  \code{\link{MapNort}}, \code{\link{mapporco}}
}
\concept{Norte Demersales}
\concept{mapas base}
