% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MapEcol.camp}
\alias{MapEcol.camp}
\title{Mapa de índices ecológicos para un grupo en una campaña}
\usage{
MapEcol.camp(gr, esp = "999", camp, dns = "Pnew", ind = "n",
  indec = "div", plot = T, bw = F, ti = T, idi = "l", es = T,
  out.dat = F, layout = NA)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{ha de ser 999 cuando se quiere incluir todas las especies del grupo, o elegir todas las especies deseadas con los codigos de las especies}

\item{camp}{Campaña de la que se extraen los datos: un año comcreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Pnew", Cantábrico "Cant, Golfo de Cádiz "Arsa" (únicamente para sacar datos al IBTS, no gráficos)}

\item{ind}{Elige el valor (n)úmero o (p)eso sobre el que se calculan los índices de diversidad, dominancia....}

\item{indec}{Elige el índice ecológico a representar: opciones disponibles: 'div' 'simp', 'domsimp' y 'nesp'}

\item{plot}{Si T saca un gráfico en pantalla}

\item{bw}{Gráfico en blanco en negro si T o en color si F}

\item{ti}{Añade el nombre de la especie en latín sin T, si F no añade titulo}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{es}{Si T saca los titulos y rotulos en español, si F en inglés}

\item{out.dat}{Si T el resultado final de la función es la figura en pantalla, pero los datos en objeto}

\item{layout}{Organización de gráficos en filas ó columnas c(r,c)}
}
\value{
Saca el mapa de diversidad en la campaña seleccionada.
}
\description{
Utiliza los datos del Camp representar la variación geográfica de los índices ecológicos
}
\examples{
ecolgr.camp(1,999,"P08","Pnew",ind="n")
}
\seealso{
MapEcol.camp {\link{MapEcol.camp.r}}
}

