% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapGaltal.R
\name{MapGaltal}
\alias{MapGaltal}
\title{Mapa distribución en Galicia entre tallas tmin y tmax}
\usage{
MapGaltal(gr, esp, camps, dns = "Cant", tmin = 1, tmax = 999,
  cor.time = TRUE, ind = "n", sex = NA, bw = TRUE, ti = TRUE,
  plot = TRUE, idi = "l", cexleg = 1, years = TRUE, layout = NA,
  out.dat = TRUE, leg = TRUE, ceros = TRUE, escmult = 0.25,
  incl2 = FALSE)
}
\arguments{
\item{gr}{Grupo de la especie: Sólo hay tallas para 1 peces, 2 crustáceos decápodos y 3 algunos moluscos}

\item{esp}{Código de la especie numerico o caracter con tres espacios. 999 para todas las especies del grupo}

\item{dns}{Elige el origen de las bases de datos: sólo Galicia "Cant" o "Cnew"}

\item{tmin}{Talla mínima del intervalo de tallas a incluir}

\item{tmax}{Talla máxima del intervalo de tallas a incluir}

\item{cor.time}{Si T corrige las abundancias en función de la duración del lance}

\item{ind}{Permite elegir entre número "n" o peso "p" peso s?lo funciona cuando existen a y b en especies.dbf y se elige s?lo una especie}

\item{sex}{Permite elegir entre machos(1), hembras(2) o indeterminados(3), NA escoge sin tener en cuenta el sexo}

\item{bw}{gráfico en blanco en negro si T o en color si F}

\item{ti}{Si T a?ade titulo al mapa, el nombre de la especie en latín}

\item{plot}{Saca el gráfico (T) o lo guarda como objeto para componer con otros gráficos (F)}

\item{idi}{Nombre científico de la especie ("l") o nombre común ("e")}

\item{cexleg}{Varía el tamaño de letra de los ejes y del n?mero de la leyenda}

\item{years}{Si T saca los años como nombre de campa?a en los paneles lattice de campa?as}

\item{layout}{Organización de gráficos en filas y columnas c(r,c)}

\item{out.dat}{Si T el resultado final de la funcion es la figura en pantalla, pero los datos en objeto}

\item{leg}{Si T añade la leyenda}

\item{ceros}{Añade puntos para los datos igual a 0 si T, si F no incluye x en los valores 0}

\item{escmult}{Varía la relación de tamaño de los puntos con la leyenda y el máximo en los datos}

\item{camp}{Campaña a representar en el mapa de un año comcreto (XX): Demersales "NXX". Solo Demersales para Galicia}
}
\value{
Si out.dat=TRUE devuelve un data.frame con columnas: lan,lat,long,prof,numero (de individuos entre tmin y tmax),camp, si out.dat=F saca el gráfico en pantalla o como objeto para combinar con otros gráficos con print.trellis
}
\description{
Saca mapas de distribución de una especie entre tallas tmin y tmax para varias campañas
}
\examples{
MapGaltal(1,50,Nsh[21:29],"Cant",1,23,layout=c(3,3),out.dat=TRUE,ceros=FALSE,incl2=TRUE)
}
\seealso{
Other mapas: \code{\link{GrafMarksGPS_2}},
  \code{\link{GrafMarksGPS}}, \code{\link{MapEcol.camp}},
  \code{\link{MapLansGPS}}, \code{\link{armap.camp}},
  \code{\link{armap.tot}}, \code{\link{map.check}},
  \code{\link{maparea}}, \code{\link{maphistage}},
  \code{\link{maphistal}}, \code{\link{maphist}}

Other tallas: \code{\link{maphistal}}
}
