% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapLansGPS.R
\name{MapLansGPS}
\alias{MapLansGPS}
\title{Mapa con los lances en segmentos}
\usage{
MapLansGPS(
  camp,
  dns = "Porc",
  leg = F,
  incl0 = FALSE,
  add = F,
  xlims = NA,
  ylims = NA,
  ti = FALSE,
  col = "navy",
  lwd = 2,
  places = TRUE,
  Nlans = FALSE,
  cexlans = 0.8,
  rumbo = FALSE,
  es = TRUE,
  bw = FALSE,
  ax = TRUE,
  ErrLans = NA,
  ICESrect = FALSE,
  ICESlab = FALSE,
  ICESlabcex = 0.8,
  graf = FALSE,
  xpng = 1200,
  ypng = 800,
  ppng = 15
)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa"}

\item{incl0}{si T incluye los lances nulos con color negro}

\item{add}{si T añade las líneas al gráfico anterior si F lo empieza de cero}

\item{xlims}{Define los limites longitudinales del mapa, si se deja en NA toma los límites de long del área definida en la campaña}

\item{ylims}{Define los limites latitudinales del mapa, si se deja en NA toma los límites de lat del área definida en la campaña}

\item{ti}{si TRUE inclye el nombre de la campaña en el gráfico}

\item{col}{Define el color de los segmentos}

\item{lwd}{Define el ancho de los segmentos}

\item{places}{si T por defecto, incluye las etiquetas de países y ciudad en tierra, no funciona en Porcupine}

\item{Nlans}{si T incluye los números de lance por encima de los segmentos}

\item{cexlans}{tamaño de las etiquetas de los lances si se incluyen}

\item{rumbo}{si T incluye}

\item{bw}{Si T gráfico en blanco y negro por default, si F gráfico en color}

\item{ax}{Si T saca los ejes x e y}

\item{ErrLans}{NA por defecto, pero si se da un vector con varios lances con errores que sale de qcdistlan.camp()$lances solo saca los lances con error}
}
\value{
Devuelve un data.frame con datos de cada lance, las variables dependen de la selección de hidro y redux
}
\description{
gráfica todos los lances de una campaña con las distancias recorridas en cada uno de ellos
}
\examples{
MapLansGPS("12C","Cant",ti=TRUE,xlims=c(-10.4,-7.4),ylims=c(41.8,44.6))
MapLansGPS(123,"Arsa",ti=TRUE,ErrLans=c(1,7,40),Nlans=T,graf="Maplans123")
}
\seealso{
{\link{datlan.camp}}, {\link{qcdistlan.camp}}

Other mapas: 
\code{\link{BelgicaMound}},
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapHidro.camp}()},
\code{\link{MapHidro_b.camp}()},
\code{\link{MapLansHH}()},
\code{\link{NepFUs.camp}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{hmmSAC}},
\code{\link{map.check}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other PescaWin: 
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapLansHH}()},
\code{\link{arMapLansGPS}()},
\code{\link{pasa.arte}()},
\code{\link{pasa.lan}()}
}
\concept{PescaWin}
\concept{mapas}
