% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapLansGPS.r
\name{MapLansGPS}
\alias{MapLansGPS}
\title{Gráfico con los lances en segmentos}
\usage{
MapLansGPS(camp, dns = "Pnew", incl0 = FALSE, xlims = NA, ylims = NA,
  places = TRUE, es = T, bw = FALSE, ax = T)
}
\arguments{
\item{camp}{Campaña de la que se extraen los datos: año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa"}

\item{incl0}{Si T se incluyen los lances nulos}

\item{xlims}{Define los limites longitudinales del mapa, si se deja en NA toma los límites de long del área definida en la campaña}

\item{ylims}{Define los limites latitudinales del mapa, si se deja en NA toma los límites de lat del área definida en la campaña}

\item{places}{si T por defecto, incluye las etiquetas de países y ciudad en tierra, no funciona en Porcupine}

\item{bw}{Si T gráfico en blanco y negro por default, si F gráfico en color}

\item{ax}{Si T saca los ejes x e y}
}
\value{
Devuelve un data.frame con datos de cada lance, las variables dependen de la selección de hidro y redux
}
\description{
gráfica todos los lances de una campaña con las distancias recorridas en cada uno de ellos
}
\examples{
MapLansGPS("12C","Cant")
}
\seealso{
{\link{datlan.camp}}, {\link{qcdistlan.camp}}

Other mapas: \code{\link{GrafMarksGPS_2}},
  \code{\link{GrafMarksGPS}}, \code{\link{MapEcol.camp}},
  \code{\link{MapGaltal}}, \code{\link{NepClos}},
  \code{\link{armap.camp}}, \code{\link{armap.tot}},
  \code{\link{map.check}}, \code{\link{maparea}},
  \code{\link{maphistage}}, \code{\link{maphistal}},
  \code{\link{maphist}}

Other PescaWin: \code{\link{GrafMarksGPS_2}},
  \code{\link{GrafMarksGPS}}, \code{\link{pasa.lan}}
}
