% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapLansHH.R
\name{MapLansHH}
\alias{MapLansHH}
\title{Gráfico con los lances en segmentos}
\usage{
MapLansHH(
  fic,
  dns = "Cant",
  nurows = NA,
  incl0 = FALSE,
  incl2 = TRUE,
  xlims = NA,
  ylims = NA,
  col = 2,
  lwd = 2,
  places = TRUE,
  Nlans = FALSE,
  bearing = FALSE,
  es = T,
  bw = FALSE,
  ax = TRUE,
  graf = FALSE,
  xpng = 1400,
  ypng = 800,
  ppng = 15
)
}
\arguments{
\item{fic}{Nombre del fichero HH con los datos a comprobar en formato HH de DATRAS, con su path o sin el si se está en el mismo directorio}

\item{dns}{Elige el supuesto origen de los datos: Porcupine "Porc", Cantábrico "Cant", Golfo de Cádiz "Arsa" para otras campañas IBTS "Other"}

\item{incl0}{Si T se incluyen los lances nulos}

\item{xlims}{Define los limites longitudinales del mapa, si se deja en NA toma los límites de long del área definida en la campaña}

\item{ylims}{Define los limites latitudinales del mapa, si se deja en NA toma los límites de lat del área definida en la campaña}

\item{col}{Define el color de los segmentos}

\item{lwd}{Define el ancho de los segmentos}

\item{places}{si T por defecto, incluye las etiquetas de países y ciudad en tierra, no funciona en Porcupine}

\item{Nlans}{si T incluye los números de lance por encima de los segmentos}

\item{bw}{Si T gráfico en blanco y negro por default, si F gráfico en color}

\item{ax}{Si T saca los ejes x e y}

\item{graf}{si F el gráfico sale en la pantalla, si nombre fichero va a fichero en el directorio de trabajo del Rstudio ver getwd()}

\item{xpng}{width archivo png si graf es el nombre del fichero}

\item{ypng}{height archivo png si graf es el nombre del fichero}

\item{ppng}{points png archivo si graf es el nombre del fichero}

\item{nrows}{Número de filas del fichero que se quieren obtener, por defecto NA y coge todo el fichero}

\item{Bearing}{si T marca el rumbo hacia el punto final con un punto en el final del lance.}
}
\value{
Devuelve un data.frame con datos de cada lance, las variables dependen de la selección de hidro y redux
}
\description{
gráfica todos los lances de una campaña con las distancias recorridas en cada uno de ellos
}
\examples{
MapLansHH(fic=icesDatras::getDATRAS("HH","SP-NORTH",2001,4),dns="Other",graf="MapLans_ejemplo")
MapLansHH(fic=icesDatras::getDATRAS("HH","IE-IGFS",2019,4),dns="Other")
}
\seealso{
{\link{MapLansGPS}}, {\link{armap.camp}}

Other mapas: 
\code{\link{BelgicaMound}},
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapEcol.camp}()},
\code{\link{MapGaltal}()},
\code{\link{MapHidro.camp}()},
\code{\link{MapHidro_b.camp}()},
\code{\link{MapLansGPS}()},
\code{\link{MatchHidro.ctd}()},
\code{\link{NepFUs.camp}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{arMapLansGPS}()},
\code{\link{armap.camp}()},
\code{\link{armap.tot}()},
\code{\link{hmmSAC}},
\code{\link{map.check}()},
\code{\link{mapICESStatRec.plotrix}()},
\code{\link{mapICESStatRec}()},
\code{\link{maparea}()},
\code{\link{maphistage}()},
\code{\link{maphistal}()},
\code{\link{maphist}()}

Other PescaWin: 
\code{\link{GrafMarksGPS_2}()},
\code{\link{GrafMarksGPS}()},
\code{\link{MapLansGPS}()},
\code{\link{arMapLansGPS}()},
\code{\link{pasa.arte}()},
\code{\link{pasa.lan}()}
}
\concept{PescaWin}
\concept{mapas}
