% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MapNort}
\alias{MapNort}
\title{Mapa del Cantábrico y Galicia}
\usage{
MapNort(lwdl = 0.5, cuadr = F, cuadrMSFD = F, ICESrect = F, bw = T,
  es = F, ax = T, strat = F, places = F, xlims = c(-10.25, -1.4),
  ylims = c(41.82, 44.48))
}
\arguments{
\item{lwdl}{Ancho de las líneas del mapa}

\item{cuadr}{Si T saca las cuadrículas de 5x5 millas naúticas}

\item{cuadrMSFD}{Si T dibuja caudrícula de 10 millas naúticas utilizada para la evaluación de la estrategia marina (MSFD)}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{bw}{si T mapa con tierra en gris, si F tierra en color}

\item{es}{si T saca textos en español, si F en inglés}

\item{ax}{Si T saca los ejes x e y}

\item{strat}{Si marca los sectores geográficos (los batimetricos salen con las líneas correspondientes}

\item{places}{Si T saca ciudades y puntos geográficos de referencia}

\item{xlims}{Define los limites longitudinales del mapa, por defecto -10.25 y -1.4 oeste}

\item{ylims}{Define los limites longitudinales del mapa, por defecto 41.82 y 44.48 norte}
}
\value{
Saca en pantalla el mapa y es utilizada por otras funciones
}
\description{
Función auxiliar para sacar el mapa de la campaña Demersales
}
\seealso{
MapCant {\link{MapCant}} mapporco {\link{mapporco}}
}

