% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapporco.R
\name{mapporco}
\alias{mapporco}
\title{Mapa del Banco de Porcupine sin referencias en tierra}
\usage{
mapporco(
  xlims = c(-15.5, -10.5),
  ylims = c(50.5, 54.5),
  lwdl = 1,
  latlonglin = TRUE,
  cuadr = FALSE,
  ICESrect = FALSE,
  ICESlab = FALSE,
  ICESlabcex = 0.7,
  label = FALSE,
  colo = 2,
  dens = 0,
  bw = F,
  places = TRUE,
  es = TRUE,
  ax = TRUE,
  escala = FALSE,
  cex.scala = 0.6,
  graf = FALSE,
  corners = FALSE,
  leg = FALSE,
  sectcol = FALSE,
  FU = NA,
  FUsLab = FALSE,
  colFU = rgb(0, 1, 0.2),
  densfu = 100,
  limfu = "darkgreen",
  ColFU = rgb(0, 1, 0, 0.2),
  SACs = FALSE,
  filSAC = FALSE
)
}
\arguments{
\item{xlims}{Define los limites longitudinales del mapa, los valores por defecto son los del total del área de la campaña}

\item{ylims}{Define los limites latitudinales del mapa, los valores por defecto son los del total del área de la campaña}

\item{lwdl}{Ancho de las líneas del mapa}

\item{latlonglin}{si T saca líneas longi y latitudinales muy claritas}

\item{cuadr}{Si T saca las cuadrículas de 5x5 millas naúticas}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{ICESlab}{Si T incluye las etiquetas de los rectángulos ICES}

\item{ICESlabcex}{tamaño del ICESlab en cex, .5 por defecto subirlo si se quiere más grande}

\item{label}{Si T saca las etiquetas de cada una de las cuadriculas numeradas consecutivamente por estratos 1A,1B,2B,3A,3B}

\item{colo}{Color de las etiquetas, por defecto rojas}

\item{dens}{si mayor de 0 las superficies de los estratos tienen patrones de líneas}

\item{bw}{Si T mapa en blanco y negro respecto a tierra y puntos, en caso contrario en color. Para sacar el diseño de estratos de Porcupine se utiliza sectcol=TRUE y leg=TRUE}

\item{es}{Textos en español, si F en inglés}

\item{ax}{Si T saca los ejes x e y}

\item{escala}{si T incluye una escala}

\item{cex.scala}{tamaño fuente escala, por defecto si hay escala .6}

\item{graf}{Si F no saca nada, si pones el nombre de un gráfico lo saca saca como archivo png y al final del proceso dice dónde está el mapa con ese nombre:}

\item{corners}{Si T coloca dos puntos rojos en los extremos nordeste y suroeste para ajustar mapas al PescaWin con ax=F}

\item{leg}{incluye la leyenda con los colores/patrones de los estratos}

\item{sectcol}{si T pone los sectores con color de fondo, en caso contrario lo deja en blanco, bueno para armap.tot}

\item{FU}{Por defecto NA pero si se incluye un vector con la lista de las unidades funcionales las pinta (las disponibles en Porcupine sólo es FU16)}

\item{FUsLab}{Por defecto F, pero si T incluye una etiqueta con los nombres de las FUs seleccionadas en FUs}

\item{colFU}{Por defecto rgb(0,1,0,.2), color para las zonas de las FU de cigalas}

\item{limfu}{color de los bordes de las FU}

\item{SACs}{si T incluye los polígonos de las areas speiales de conservación Irlandesas}

\item{filSAC}{si T los poligonos están con fondo de color}
}
\value{
Saca en pantalla el mapa y es utilizada por otras funciones, si wmf=TRUE lo saca a metafile para fondo del pescawin
}
\description{
Función auxiliar para sacar el mapa de la campaña Porcupine
}
\examples{
mapporco(xlims = c(-15.3,-8.10),corners = F,leg=T,sectcol=T,dens=10,es=F) para sustituir a maparea
mapporco(corners = F,leg=F,sectcol=F,dens=0,es=F) # para sustituir a mapporco
}
\seealso{
{\link{MapNort}}, {\link{MapCant}}

Other mapas base: 
\code{\link{MapArsa}()},
\code{\link{MapCant}()},
\code{\link{MapIberia}()},
\code{\link{MapMedit}()},
\code{\link{MapNort}()},
\code{\link{MapPorc}()},
\code{\link{mapscale}()}

Other Porcupine: 
\code{\link{BelgicaMound}},
\code{\link{MapPorc}()},
\code{\link{NewPorcCanyon}},
\code{\link{PorShelfSAC}},
\code{\link{PorcNWSAC}},
\code{\link{PorcSWSAC}},
\code{\link{Psh}},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{hmmSAC}},
\code{\link{maparea}()}
}
\concept{Porcupine}
\concept{mapas base}
