% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapporco.r
\name{mapporco}
\alias{mapporco}
\title{Mapa del Banco de Porcupine sin referencias en tierra}
\usage{
mapporco(xlims = c(-15.5, -10.5), ylims = c(50.5, 54.5), lwdl = 1,
  cuadr = FALSE, ICESrect = FALSE, label = FALSE, colo = 2, ax = TRUE,
  wmf = FALSE, corners = FALSE)
}
\arguments{
\item{xlims}{Define los limites longitudinales del mapa, los valores por defecto son los del total del área de la campaña}

\item{ylims}{Define los limites latitudinales del mapa, los valores por defecto son los del total del área de la campaña}

\item{lwdl}{Ancho de las líneas del mapa}

\item{cuadr}{Si T saca las cuadrículas de 5x5 millas naúticas}

\item{ICESrect}{Si T saca los rectangulos ices de 1 grado de latitud por medio de longitud}

\item{label}{Si T saca las etiquetas de cada una de las cuadriculas numeradas consecutivamente por estratos 1A,1B,2B,3A,3B}

\item{colo}{Color de las etiquetas, por defecto rojas}

\item{ax}{Si T saca los ejes x e y}

\item{wmf}{Si T saca a fichero metafile porconc.emf}

\item{corners}{Si T coloca dos puntos rojos en los extremos nordeste y suroeste para ajustar mapas al PescaWin con ax=F}
}
\value{
Saca en pantalla el mapa y es utilizada por otras funciones, si wmf=TRUE lo saca a metafile para fondo del pescawin
}
\description{
Función auxiliar para sacar el mapa de la campaña Porcupine
}
\seealso{
{\link{MapNort}}, {\link{MapCant}}

Other mapas base: \code{\link{MapArsa}},
  \code{\link{MapCant}}, \code{\link{MapIberia}},
  \code{\link{MapMedit}}, \code{\link{MapNort}}

Other Porcupine: \code{\link{NepClos}}, \code{\link{Psh}},
  \code{\link{maparea}}
}
