% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_buff_days.r
\name{max.buff.days}
\alias{max.buff.days}
\title{max.buff.days}
\usage{
\method{max}{buff.days}(df, trap)
}
\arguments{
\item{df}{<describe argument>}

\item{trap}{<describe argument>}
}
\value{
<describe return value>
}
\description{
df <- df2
}
\details{
<other comments found in file>
 calculate beginning buffer
 calculate ending buffer
}
\examples{
<insert examples>

}
\author{
WEST Inc.
}
\seealso{
\code{\link{<related routine>}}, \code{\link{<related routine>}}
}

