% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Msh.R
\docType{data}
\name{Msh}
\alias{Msh}
\title{Serie histórica de la campaña Medits}
\format{
A vector with codes for the surveys from 1994 to 2011
}
\usage{
Msh
}
\description{
Serie histórica campaña Medits
}
\details{
Serie completa con las campañas realizadas en el Mediterraneo en la campaña MEDITS
}
\seealso{
Other Series datos: 
\code{\link{As1}},
\code{\link{As2}},
\code{\link{CampsDNS.camp}()},
\code{\link{Nsh}},
\code{\link{PresenciaEsp.camp}()},
\code{\link{Psh}}
}
\concept{MEDITS}
\concept{Series datos}
\keyword{datasets}
