% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nsh.r
\docType{data}
\name{Nsh}
\alias{Nsh}
\title{Serie histórica de la campaña Demersales plataforma Norte}
\format{Vector con los códigos de las campañas realizadas desde 1983 hasta 2016}
\usage{
Nsh
}
\description{
Serie histórica Demersales Norte
}
\details{
Serie completa con las campañas de cuarto trimestre realizadas en el Cantábrico y Galicia, el año 1987 no hubo campaña
}
\seealso{
Other Series datos: \code{\link{As1}}, \code{\link{As2}},
  \code{\link{Msh}}, \code{\link{Psh}}
}
\keyword{datasets}
