% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passageWithLifeStageAssign.R
\name{passageWithLifeStageAssign}
\alias{passageWithLifeStageAssign}
\title{passageWithLifeStageAssign}
\usage{
passageWithLifeStageAssign(site, taxon, min.date, max.date, output.file,
  ci = TRUE, nLS = NULL, weightUse = NULL, autoLS = FALSE,
  reclassifyFL = FALSE)
}
\arguments{
\item{site}{The identification number of the site for which estimates are 
required.}

\item{taxon}{The species identifier indicating the type of fish of interest. 
This is always \code{161980}; i.e., Chinook Salmon.}

\item{min.date}{The start date for data to include. This is a text string in 
the format \code{\%Y-\%m-\%d}, or \code{YYYY-MM-DD}.}

\item{max.date}{The end date for data to include.  Same format as 
\code{min.date}.}

\item{output.file}{The name of the file prefix under which output is to be 
saved.  Set to NA to plot to the Plot window.}

\item{ci}{A logical indicating if 95% bootstrapped confidence intervals 
should be estimated along with passage estimates.}

\item{nLS}{Number of life stage groups to estimate. Ignored if 
\code{autoLS=FALSE}.  See Details.}

\item{weightUse}{A logical indicating if variable weight should be used for
the analytical life stage assignment;  the default is \code{NULL}. Ignored
if \code{autoLS=FALSE}.  See Details.}

\item{autoLS}{Default of \code{FALSE} leads to no assigning of no analytical
life stage. If \code{TRUE}, assignment of analytical life stage is done. 
See Details.}

\item{reclassifyFL}{A logical indicating if passage should be estimated via 
forklength-based class groups.}
}
\value{
A data frame containing daily passage estimates, corrected for times 
  not fishing, along with associated standard errors.
  
  A bar chart displaying relative proportions over both run and life stage.
  
  For each unique combination of run and life stage containing at least one 
  fish, a graph \code{png} of catch over time, a \code{csv} of daily catch 
  and passage.
  
  A graphical display of efficiency over time, for each trap, along with an 
  accompanying \code{csv} tabular datasheet.
  
  In the case when \code{autoLS=TRUE}, a plot (\code{pdf}) and a confusion 
  matrix (\code{csv}) comparing the analytical and morphometric life stage 
  assignments.
}
\description{
Estimate annual passage / production estimates between days
specified.
}
\details{
The date range difference specified via \code{max.date} and
  \code{min.date} must be less than or equal to 366 days.  Note that this
  cutoff allows for leap-year annual estimates.
  
  Passage requires the division of a non-zero catch by a non-zero efficiency.
  A non-zero catch numerator ensures that passage estimates are greater than
  zero, while a non-zero efficiency denominator ensures that passage 
  estimates are less than infinity.  The program exits if either of these 
  conditions are violated.
  
  Assuming validity, catch data are summarized via unique combinations of 
  \code{trapVisitID}, \code{FinalRun}, and \code{lifeStage}.  After 
  summarizing, all catch data are collapsed so as to have one line per 
  combination of these variables.  The use of a plus-count algorithm divvies
  up non-randomly selected fish into appropriate \code{FinalRun} and 
  \code{lifeStage} categories.  In this way, all fish are ultimately utilized
  in all pasasge estimates by life stage and run.
  
  Observed fish are also corrected for half-cone operations, in which the
  intake of fish is reduced to half of a cone's aperture being covered, so as
  to reduce flow, and thus captured fish.  Generally, all fish captured 
  during half-cone operations are multiplied by the value of the global 
  \code{halfConeMulti} variable, which serves as a multiplier to correct for 
  fish missed.  Variable \code{halfConeMulti} is set to 2.
  
  The unique combinations of run and life stage in the catch data dictate the
  reports generated. Thus, reports spanning different time periods and sites
  may report different run and life stage passage combinations.  
  
  In the case when \code{autoLS=TRUE}, the life stage is assigned 
  analytically.  Note that the only number of groups allowed by argument 
  (\code{nLS}) is two or three. If \code{NULL}, the function 
  \code{assignLifeStage} determines the number of groups utilized.
  
  If \code{weightUse} is \code{FALSE}, any recorded weight measurements are
  not used in the analytical life stage assignment. If \code{weightUse} is
  \code{NULL}, the function \code{\link{assignLifeStage}} will determine if
  weight will be used or not.
  
  When \code{reclassify=TRUE}, the biologically recorded \code{lifeStage} is 
  redefined via groups specifed in data frame \code{forkLengthCutPoints}, as 
  defined in \code{GlobalVars}.  Default behavior leads to four separate 
  fork-length-based groups.  Similar to reports that break out totals by
  biologically assigned \code{lifeStage}s, reports utilizing breakout by
  forklength may not report totals for all four groups, if the river and date
  range specified caught no fish with that particular range of fork lengths.
}
\examples{
\dontrun{
<insert examples>

}
}
\author{
WEST Inc.
}
\seealso{
\code{assignLifeStage}, \code{assignLSCompare}
}

