% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psh.R
\docType{data}
\name{Psh}
\alias{Psh}
\title{Serie hist칩rica de la campa침a Porcupine realizada por el IEO en el Banco Porcupine}
\format{
A vector with codes for the surveys from 2001 to 2018
}
\usage{
Psh
}
\description{
Serie hist칩rica Porcupine
}
\details{
Serie completa con las campa침as realizadas en Porcupine Bank
}
\seealso{
Other Series datos: 
\code{\link{As1}},
\code{\link{As2}},
\code{\link{ListCamps_b}()},
\code{\link{ListCamps}()},
\code{\link{Msh}},
\code{\link{Nsh}},
\code{\link{PresenciaEsp.camp}()}

Other Porcupine: 
\code{\link{MapPorc}()},
\code{\link{VedaCigala.out}},
\code{\link{VedaCigala}},
\code{\link{maparea}()},
\code{\link{mapporco}()}
}
\concept{Porcupine}
\concept{Series datos}
\keyword{datasets}
