% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcdistlan.camp.R
\name{qcdistlan.camp}
\alias{qcdistlan.camp}
\title{Comprueba la distancia recorrida en los lances y la consistencia con recorrido y la velocidad en los datos del CAMP}
\usage{
qcdistlan.camp(camp, dns = "Cant", todos = FALSE, pc.error = 2)
}
\arguments{
\item{camp}{campaña a revisar los datos en formato Camp Xyy}

\item{dns}{Origen de bases de datos: Cant cantábrico, Pnew Porcupine}

\item{todos}{Por defecto F. Si T lista todos los lances con valores, si no sólo los que pc.error>error}

\item{pc.error}{porcentaje de error aceptable para no mostrar los lances como erróneos}
}
\value{
Devuelve un data.frame con campaña, lance, recorrido según: camp, según la fórmula haversine entre largada y virada, velocidad, velocidad para recorrer el recorrido, tiempo y %s de error en las comparaciones
}
\description{
Sirve para control de calidad y asegurarse que los datos de distancias y posiciones son correctos
}
\examples{
qcdistlan.camp("C14","Cant",pc.error=.01)
}
\references{
distHaversine function gives the haversine calculation of distance between two geographic points (see: library gegosphere)
}
\seealso{
{\link{MapLansGPS}}

Other Control de calidad: \code{\link{qcLW.camp}},
  \code{\link{qcLWbucl.camp}}, \code{\link{qcTalPez.lan}},
  \code{\link{qcdtall.camp}}, \code{\link{qcdtallrev.camp}}
}
