% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcdistlan.camp.R
\name{qcdistlan.camp}
\alias{qcdistlan.camp}
\title{Comprueba la distancia recorrida en los lances y la consistencia con recorrido y la velocidad en los datos del CAMP}
\usage{
qcdistlan.camp(
  camp,
  dns = "Cant",
  todos = FALSE,
  pc.error = 2,
  error.rb = TRUE,
  plots = TRUE,
  es = FALSE,
  esc.mult = 1
)
}
\arguments{
\item{camp}{campaña a revisar los datos en formato Camp Xyy}

\item{dns}{Origen de bases de datos: "Cant" cantábrico, "Porc" o "Pnew" Porcupine, "Arsa" para el Golfo de Cádiz y "Medi" para MEDITS}

\item{todos}{Por defecto F. Si T lista todos los lances con valores, si no sólo los que pc.error>error}

\item{pc.error}{porcentaje de error aceptable para no mostrar los lances como erróneos}

\item{error.rb}{si F no tiene en cuenta los errores de rumbo que tampoco tiene mucho peso}

\item{plots}{si T abre una pantalla con tres gráficos de los errores detectados en error.dist, error.vel y error.rumb}

\item{es}{if T all text will be in Spanish, if F in English}
}
\value{
Devuelve un data.frame con campaña, lance, recorrido, recorrido según la fórmula de Haversine, recorrido según la velocidad x el tiempo, velocidad, tiempo, rumbo, rumbo estimado según posiciones,velocidad calculada a partir de la distancia y el tiempo, y los porcentajes de errores de distancia, velocidad y rumbo, horas de amanecer, salida del sol, mediodía, puesta del sol y noche.
}
\description{
Sirve para control de calidad y asegurarse que los datos de distancias, posiciones, y las horas de largada y virada son correctos y no salen del estándar de largar y virar de día
}
\examples{
qcdistlan.camp("C14","Cant",pc.error=.01)
qcdistlan.camp("216","Arsa",pc.error=.01)
}
\references{
distHaversine function gives the haversine calculation of distance between two geographic points \code{\link[geosphere]{distHaversine}}
}
\seealso{
{\link{MapLansGPS}}

Other Control de calidad: 
\code{\link{qcLW.camp}()},
\code{\link{qcLWbucl.camp}()},
\code{\link{qcTalPez.camp}()},
\code{\link{qcTalPez.lan}()},
\code{\link{qcdtall.camp}()},
\code{\link{qcdtallrev.camp}()},
\code{\link{qclandatr.camp}()}
}
\concept{Control de calidad}
