% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcdtallrev.camp.R
\name{qcdtallrev.camp}
\alias{qcdtallrev.camp}
\title{Revisión de los histogramas de tallas de todas las especies de una campaña}
\usage{
qcdtallrev.camp(camp = "P12", dns = "Porc", nlans = 2)
}
\arguments{
\item{camp}{Campaña a representar en el mapa de un año concreto (XX): Demersales "NXX", Porcupine "PXX", Arsa primavera "1XX" y Arsa otoño "2XX"}

\item{dns}{Elige el origen de las bases de datos: Porcupine "POrc" o "Pnew", Cantábrico "Cant", Golfo de Cadiz "Arsa" (proporciona los datos para Medits pero no saca mapas)}

\item{nlans}{Permite no mostrar la relación talla-peso de una especie si el número de lances en que ha aparecido es menor de nlans (por defecto nlans=2)}
}
\description{
Revisa la distribución de tallas estratificada de todas las especies de peces que hayan aparecido en una campaña.Se puede salir de la función cerrando el gráfico y se puede obtener una función para la campaña y usarla en cada lance. Ver documentación qcdtall.camp
}
\examples{
qcdtallrev.camp(camp="N10",dns="Cant",nlans=2)
}
\seealso{
{\link{qcdtall.camp}}

Other Control de calidad: 
\code{\link{qcLW.camp}()},
\code{\link{qcLWbucl.camp}()},
\code{\link{qcTalPez.camp}()},
\code{\link{qcTalPez.lan}()},
\code{\link{qcdistlan.camp}()},
\code{\link{qcdtall.camp}()},
\code{\link{qclandata.camp}()}
}
\concept{Control de calidad}
