% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcTalPez.lan.r
\name{qcTalPez.lan}
\alias{qcTalPez.lan}
\title{Comprueba que están medidas todas las especies en capturas y viceversa. Comprueba concordancia entre números y pesos en los ficheros de tallas NTALLXXX.dbf y fauna FAUNAXXX.dbf}
\usage{
qcTalPez.lan(camp, dns = "Cant", lan, gr = 1)
}
\arguments{
\item{camp}{campaña a revisar los datos en formato Camp Xyy}

\item{dns}{Origen de bases de datos: "Cant" cantábrico, "Porc" o "Pnew" Porcupine}

\item{lan}{lance para el que se quiere comprobar que estan medidas todas las especies (sólo comprueba peces, por defecto gr=1, si se quiere otros cambiar gr) Si lan=NA comprueba todos los lances de la campaña}

\item{gr}{El grupo que se quiere comprobar 1 peces, 2 crustaceos, 3 cefalópodos. 4, 5 y 6 nunca se miden.}
}
\value{
Devuelve la lista de especies capturadas pero no medidas o viceversa
}
\description{
Sirve para control de calidad y asegurándose que no faltan datos o están incompletos (comprueba coherencia entre faunaXXX.dbf y NtallXXX.dbf y avisa cuando no son coherentes) Existe una funciòn escondida para procesar de una vez toda la campaña \code{qcTalPez.camp(camp=XXX,dns,gr=1)}, pero una vez encontrado que el lance en que está el error es mejor utilizar la función lance por lance
}
\examples{
qcTalPez.lan("C14",dns="Cant",lan=14,gr=1)
}
\seealso{
Other Control de calidad: 
\code{\link{qcLW.camp}()},
\code{\link{qcLWbucl.camp}()},
\code{\link{qcTalPez.camp}()},
\code{\link{qcdistlan.camp}()},
\code{\link{qcdtall.camp}()},
\code{\link{qcdtallrev.camp}()},
\code{\link{qclandatr.camp}()}
}
\concept{Control de calidad}
