% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talpes.camp.r
\name{talpes.camp}
\alias{talpes.camp}
\title{Parámetros a y b relación talla-peso}
\usage{
talpes.camp(gr, esp, dns = "Camp")
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}
}
\description{
Extrae los parámetros a y b de la relación talla-peso de una especie determinada
}
\seealso{
Other datos_especies: 
\code{\link{AbrvEsp}()},
\code{\link{BuscaAphia}()},
\code{\link{buscacod}()},
\code{\link{buscaesp.datras}()},
\code{\link{buscaesp}()},
\code{\link{camptoyear}()},
\code{\link{hidrotodec}()}
}
\concept{datos_especies}
