% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unid.camp}
\alias{unid.camp}
\title{Unidades en que se mide una especie}
\usage{
unid.camp(gr, esp)
}
\arguments{
\item{gr}{Grupo de la especie: 1 peces, 2 crustáceos 3 moluscos 4 equinodermos 5 invertebrados}

\item{esp}{Código de la especie numérico o carácter con tres espacios. 999 para todas las especies del grupo}
}
\description{
Da información de las unidades en que se miden las especies y permite añadir datos a los gráficos
}

